% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_difference_row.R
\name{add_difference_row}
\alias{add_difference_row}
\title{Add difference rows}
\usage{
add_difference_row(x, ...)
}
\arguments{
\item{x}{(\code{gtsummary})\cr
Object with class 'gtsummary'}

\item{...}{Passed to other methods.}
}
\description{
\itemize{
\item \code{\link[=add_difference_row.tbl_summary]{add_difference_row.tbl_summary()}}
}
}
\author{
Daniel D. Sjoberg
}
\keyword{internal}
