% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_glance_statistics.R
\name{add_glance_statistics}
\alias{add_glance_statistics}
\title{Add Model Statistics}
\usage{
add_glance_statistics(
  x,
  include = everything(),
  label = NULL,
  fmt_fun = NULL,
  glance_fun = broom::glance,
  ...
)
}
\arguments{
\item{x}{'tbl_regression' object}

\item{include}{list of statistics to include in output. Must be column
names of the tibble returned by \code{broom::glance()}}

\item{label}{List of formulas specifying statistic labels,
e.g. \code{list(r.squared ~ "R2", p.value ~ "P")}}

\item{fmt_fun}{List of formulas where the LHS is a statistic and the RHS
is a function to format/round the statistics. The default is
\code{style_sigfig(x, digits = 3)}}

\item{glance_fun}{function that returns model statistics. Default is
\code{broom::glance()}. Custom functions must return a single row tibble.}

\item{...}{Additional arguments passed to \verb{glance_fun=}}
}
\value{
gtsummary table
}
\description{
\lifecycle{experimental}
Add model statistics returned from \code{broom::glance()}
}
\examples{
# add example
}
