% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_global.R
\name{add_global}
\alias{add_global}
\title{Adds the global p-value for a categorical variables}
\usage{
add_global(x, ...)
}
\arguments{
\item{x}{\code{tbl_regression} or \code{tbl_uvregression} object}

\item{...}{further arguments passed to or from other methods.}
}
\description{
This function uses \link[car:Anova]{car::Anova} with argument
\code{type = "III"} to calculate global p-values for categorical variables.
If a needed class of model is not supported by
\link[car:Anova]{car::Anova}, please create an
\href{https://github.com/ddsjoberg/gtsummary/issues}{issue} to request support.
Output from \code{tbl_regression} and \code{tbl_uvregression} objects supported.
}
\seealso{
\code{\link{add_global.tbl_regression}},
\code{\link{add_global.tbl_uvregression}}
}
\author{
Daniel D. Sjoberg
}
\keyword{internal}
