% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_global.R
\name{add_global.tbl_uregression}
\alias{add_global.tbl_uregression}
\title{Adds the global p-value for a categorical variables in \code{tbl_uregression} objects}
\usage{
\method{add_global}{tbl_uregression}(x, ...)
}
\arguments{
\item{x}{object with class \code{tbl_uregression} from the \code{\link{tbl_uregression}} function}

\item{...}{arguments to be passed to \code{\link[car]{Anova}}.  Adding \code{test.statistic = }
can change the type of test (e.g. Likelihood-ratio, Wald, etc.).}
}
\description{
This function uses \code{\link[car]{Anova}} from the \code{car} package with \code{type = "III"} to calculate global p-values.
If a needed class of model is not supported by \code{\link[car]{Anova}}, please put in an
issue at https://github.com/ddsjoberg/gtsummary/issues to request support.
}
\examples{
tbl_uregression(
  trial,
  method = "glm",
  y = "response",
  method.args = list(family = binomial),
  exponentiate = TRUE
) \%>\%
  add_global()
}
