% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_n.R
\name{add_n}
\alias{add_n}
\title{Add column with N}
\usage{
add_n(x, ...)
}
\arguments{
\item{x}{(\code{gtsummary})\cr
Object with class 'gtsummary'}

\item{...}{Passed to other methods.}
}
\description{
\itemize{
\item \code{\link[=add_n.tbl_summary]{add_n.tbl_summary()}}
\item \code{\link[=add_n.tbl_svysummary]{add_n.tbl_svysummary()}}
\item \code{\link[=add_n.tbl_regression]{add_n.tbl_regression()}}
\item \code{\link[=add_n.tbl_uvregression]{add_n.tbl_uvregression()}}
\item \code{\link[=add_n.tbl_survfit]{add_n.tbl_survfit()}}
}
}
\seealso{
\code{\link[=add_n.tbl_summary]{add_n.tbl_summary()}}
}
\author{
Daniel D. Sjoberg
}
\keyword{internal}
