% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_nevent.R
\name{add_nevent}
\alias{add_nevent}
\title{Add number of events to a {gtsummary} regression table}
\usage{
add_nevent(x, ...)
}
\arguments{
\item{x}{\code{tbl_regerssion} or \code{tbl_uvregression} object}

\item{...}{further arguments passed to or from other methods.}
}
\description{
This function adds a column of the number of events to tables created with
\link{tbl_regression} or \link{tbl_uvregression}.  Supported
model types include GLMs with binomial distribution family (e.g.
\code{\link[stats]{glm}}, \code{\link[lme4]{glmer}}, and
\code{\link[geepack]{geeglm}}) and Cox
Proportion Hazards regression models (\code{\link[survival]{coxph}}).
}
\seealso{
\link{add_nevent.tbl_regression}, \link{add_nevent.tbl_uvregression},
\link{tbl_regression}, \link{tbl_uvregression}
}
\author{
Daniel D. Sjoberg
}
