% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_nevent.R
\name{add_nevent.tbl_regression}
\alias{add_nevent.tbl_regression}
\title{Add number of events to a {gtsummary} regression table}
\usage{
\method{add_nevent}{tbl_regression}(x, ..)
}
\arguments{
\item{x}{\code{tbl_regerssion} object}

\item{...}{not used}
}
\description{
This function adds a columns of the number of events to tables creates with
\code{\link{tbl_regression}}.  Supported
model types include GLMs with binomial distribution family (e.g.
\code{\link[stats]{glm}}, \code{\link[lme4]{glmer}}, and
\code{\link[geepack]{geeglm}}) and Cox
Proportion Hazards regression models (\code{\link[survival]{coxph}}).
}
\examples{
glm(response ~ trt, trial, family = binomial) \%>\%
  tbl_regression() \%>\%
  add_nevent() \%>\%
  purrr::pluck("table_body")
}
\seealso{
Other tbl_regression: \code{\link{add_global.tbl_regression}},
  \code{\link{bold_italicize_labels_levels}},
  \code{\link{inline_text.tbl_regression}},
  \code{\link{tab_style_bold_p.tbl_regression}},
  \code{\link{tbl_regression}}
}
\author{
Daniel D. Sjoberg
}
\concept{tbl_regression}
