% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_nevent.R
\name{add_nevent.tbl_uvregerssion}
\alias{add_nevent.tbl_uvregerssion}
\title{Add number of events to a {gtsummary} regression table}
\usage{
\method{add_nevent}{tbl_uvregerssion}(x, ..)
}
\arguments{
\item{x}{\code{tbl_uvregerssion} object}

\item{...}{not used}
}
\description{
This function adds a columns of the number of events to tables creates with
\code{\link{tbl_uvregression}}.  Supported
model types include GLMs with binomial distribution family (e.g.
\code{\link[stats]{glm}}, \code{\link[lme4]{glmer}}, and
\code{\link[geepack]{geeglm}}) and Cox
Proportion Hazards regression models (\code{\link[survival]{coxph}}).
}
\seealso{
Other tbl_uvregression: \code{\link{add_global.tbl_uvregression}},
  \code{\link{add_q.tbl_uvregression}},
  \code{\link{bold_italicize_labels_levels}},
  \code{\link{inline_text.tbl_uvregression}},
  \code{\link{tab_style_bold_p.tbl_uvregression}},
  \code{\link{tbl_uvregression}}
}
\author{
Daniel D. Sjoberg
}
\concept{tbl_uvregression}
