% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_nevent.R
\name{add_nevent.tbl_uvregression}
\alias{add_nevent.tbl_uvregression}
\title{Add number of events to a {gtsummary} regression table}
\usage{
\method{add_nevent}{tbl_uvregression}(x, ...)
}
\arguments{
\item{x}{\code{tbl_uvregerssion} object}

\item{...}{not used}
}
\description{
This function adds a column of the number of events to tables created with
\code{\link{tbl_regression}} or \code{\link{tbl_uvregression}}.  Supported
model types include GLMs with binomial distribution family (e.g.
\code{\link[stats]{glm}}, \code{\link[lme4]{glmer}}, and
\code{\link[geepack]{geeglm}}) and Cox
Proportion Hazards regression models (\code{\link[survival]{coxph}}).
}
\section{Reporting Event N}{

The number of events is added to the internal \code{.$table_body} tibble,
and printed to the right of the N column. The number of events is also
accessible via the \code{inline_text()} function for printing in a report.
}

\examples{
trial \%>\%
  tbl_uvregression(
    method = glm,
    y = response,
    method.args = list(family = binomial)
  ) \%>\%
  add_nevent()
}
\seealso{
Other tbl_uvregression: \code{\link{add_global.tbl_uvregression}},
  \code{\link{add_q.tbl_uvregression}},
  \code{\link{bold_italicize_labels_levels}},
  \code{\link{inline_text.tbl_uvregression}},
  \code{\link{tab_style_bold_p.tbl_uvregression}},
  \code{\link{tbl_uvregression}}
}
\author{
Daniel D. Sjoberg
}
\concept{tbl_uvregression}
