% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_q.R
\name{add_q.tbl_uregression}
\alias{add_q.tbl_uregression}
\title{Add a column of q values to \code{tbl_uregression} object to account for
multiple comparisons}
\usage{
\method{add_q}{tbl_uregression}(x, method = "fdr",
  pvalue_fun = fmt_pvalue, ...)
}
\arguments{
\item{x}{\code{tbl_uregression} object}

\item{method}{character argument.  Methods from
\code{stats::}\code{\link[stats]{p.adjust}} are accepted.  Default is \code{method = 'fdr'}.}

\item{pvalue_fun}{function for rounding/formatting p-values.  Default is \code{\link{fmt_pvalue}}.}

\item{...}{further arguments passed to or from other methods}
}
\description{
The adjustments to the p-values is performed with
\code{stats::}\code{\link[stats]{p.adjust}}.  The default method for correction
is false discovery rate (\code{"fdr"})
}
\examples{
trial \%>\%
  tbl_uregression(
    method = "lm",
    y = "age"
  ) \%>\%
  add_global() \%>\%
  add_q()
}
\author{
Esther Drill
}
