% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_strata.R
\name{add_strata}
\alias{add_strata}
\title{Add stratified columns}
\usage{
add_strata(x, strata, include_unstratafied = FALSE)
}
\arguments{
\item{x}{\code{tbl_summary} object}

\item{strata}{variable that tables will be stratified by}

\item{include_unstratafied}{logical indicated whether to include the
original table in the result. Default is \code{FALSE}}
}
\description{
Stratifies existing \code{tbl_summary()} object by a third variable
}
\examples{
add_strata_ex1 <-
  trial \%>\%
  select(age, grade, trt, stage) \%>\%
  mutate(grade = paste("Grade", grade)) \%>\%
  tbl_summary(by = trt, include = -grade, missing = "no") \%>\%
  add_strata(grade)
}
