% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_stack_hierarchical.R
\name{ard_stack_hierarchical}
\alias{ard_stack_hierarchical}
\title{Create and Stack Hierarchical ARDs}
\usage{
ard_stack_hierarchical(
  data,
  hierarchies,
  by = NULL,
  id = NULL,
  include,
  statistic,
  denominator,
  .overall = FALSE,
  .missing = FALSE,
  .attributes = FALSE,
  .total_n = FALSE,
  .shuffle = FALSE
)
}
\description{
Create and Stack Hierarchical ARDs
}
\examples{
ard_stack_hierarchical(
  data = cards::ADAE,
  hierarchies = c(SEX, AESEV),
  by = TRTA,
  id = USUBJID,
  include = everything(),
  statistic = all_categorical() ~ "{n} ({p})",
  denominator = cards::ADSL
)

}
