% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tibble.R
\name{as_data_frame}
\alias{as_data_frame}
\title{The \code{tibble} package's \code{as_data_frame} function is an
alias for \code{tibble::as_tibble}}
\description{
See \code{\link[tibble]{as_tibble}} for details.
}
