% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_gtsummary.R
\name{as_gtsummary}
\alias{as_gtsummary}
\title{Create gtsummary table}
\usage{
as_gtsummary(table_body, ...)
}
\arguments{
\item{table_body}{(\code{data.frame})\cr
a data frame that will be added as the gtsummary object's \code{table_body}}

\item{...}{other objects that will be added to the gtsummary object list}
}
\value{
gtsummary object
}
\description{
This function ingests a data frame and adds the infrastructure around it
to make it a gtsummary object.
}
\details{
Function uses \code{table_body} to create a gtsummary object
}
\examples{
mtcars[1:2, 1:2] |>
  as_gtsummary()
}
