% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tibble.R
\name{as_tibble.fmt_regression}
\alias{as_tibble.fmt_regression}
\alias{as_data_frame.fmt_regression}
\title{Convert \code{fmt_regression} objects to data frame}
\usage{
\method{as_tibble}{fmt_regression}(x, ...)

as_data_frame.fmt_regression(x, ...)
}
\arguments{
\item{x}{object of class \code{fmt_regression} object from
\code{fmt_regression()} function}

\item{...}{further arguments passed to individual methods.}
}
\description{
Convert \code{fmt_regression} objects to data frame
}
\details{
as_data_frame.fmt_regression is alias.
}
\seealso{
\code{\link{fmt_regression}}, \code{\link{as_tibble.fmt_table1}},
\code{\link{as_tibble.fmt_uni_regression}}, \code{\link[tibble]{as_tibble}}
}
