% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-assign_class.R
\name{assign_class}
\alias{assign_class}
\title{Extract class of variable}
\usage{
assign_class(data, variable)
}
\arguments{
\item{data}{data frame}

\item{variable}{string vector of column names from data}
}
\description{
Some packages append non-base classes to data frame columns, e.g.
if data is labeled with the \code{Hmisc} package the class of a string will
be \code{c("labelled", "character")} rather than \code{c("character")} only.  This
simple function extracts the base R class.
}
\keyword{internal}
