% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-assign_dichotomous_value.R
\name{assign_dichotomous_value}
\alias{assign_dichotomous_value}
\title{For dichotomous data, returns that value that will be printed in table.}
\usage{
assign_dichotomous_value(data, variable, summary_type, class)
}
\arguments{
\item{data}{data frame}

\item{variable}{character variable name in \code{data} that will be tabulated}

\item{summary_type}{the type of summary statistics that will be calculated}

\item{class}{class of \code{variable}}
}
\value{
value that will be printed in table for dichotomous data
}
\description{
For dichotomous data, returns that value that will be printed in table.
}
\keyword{internal}
