% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-tbl_summary.R
\name{assign_sort}
\alias{assign_sort}
\title{Assigns categorical variables sort type ("alphanumeric" or "frequency")}
\usage{
assign_sort(variable, summary_type, sort)
}
\arguments{
\item{variable}{variable name}

\item{summary_type}{the type of variable ("continuous", "categorical", "dichotomous")}

\item{sort}{named list indicating the type of sorting to perform. Default is NULL.
The names of the list elements are variable names or '..categorical..' for assigning
all variables of that type.  If both a variable name and '..categorical..' are
specified, the variable name takes precedent}
}
\description{
Assigns categorical variables sort type ("alphanumeric" or "frequency")
}
\author{
Daniel D. Sjoberg
}
\keyword{internal}
