% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-tbl_summary.R
\name{assign_stat_display}
\alias{assign_stat_display}
\title{Assign type of summary statistic}
\usage{
assign_stat_display(summary_type, stat_display)
}
\arguments{
\item{summary_type}{A list that includes specified summary types}

\item{stat_display}{List with up to two named elements.  Names must be
continuous or categorical. Can be \code{NULL}.}
}
\value{
vector of stat_display selections for each variable
}
\description{
Function that assigns default statistics to display, or if specified,
assigns the user-defined statistics for display.
}
\author{
Daniel D. Sjoberg
}
\keyword{internal}
