% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_summary_digits.R
\name{assign_summary_digits}
\alias{assign_summary_digits}
\title{Assign Default Digits}
\usage{
assign_summary_digits(data, statistic, type, digits = NULL)
}
\arguments{
\item{data}{(\code{data.frame})\cr
a data frame}

\item{statistic}{(\verb{named list})\cr
a named list; notably, \emph{not} a \code{\link[=syntax]{formula-list-selector}}}

\item{type}{(\verb{named list})\cr
a named list; notably, \emph{not} a \code{\link[=syntax]{formula-list-selector}}}

\item{digits}{(\verb{named list})\cr
a named list; notably, \emph{not} a \code{\link[=syntax]{formula-list-selector}}.
Default is \code{NULL}}
}
\value{
a named list
}
\description{
Used to assign the default formatting for variables summarized with
\code{tbl_summary()}.
}
\examples{
assign_summary_digits(
  mtcars,
  statistic = list(mpg = "{mean}"),
  type = list(mpg = "continuous")
)
}
