% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-assign_test.R
\name{assign_test}
\alias{assign_test}
\title{determine the appropriate test type given two variables}
\usage{
assign_test(data, var, var_summary_type, by_var, test, group)
}
\arguments{
\item{data}{input data set}

\item{var}{categorical or continuous variable for which a test with \code{by_var} is desired}

\item{var_summary_type}{summary_type from meta data}

\item{by_var}{categorical variable}

\item{test}{list of user defined statistical tests and corresponding variables}
}
\value{
most appropriate test as text of the test function
}
\description{
determine the appropriate test type given two variables
}
\author{
Daniel Sjoberg
}
\keyword{internal}
