% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-assign_var_label.R
\name{assign_var_label}
\alias{assign_var_label}
\title{Assigns variable label to display.}
\usage{
assign_var_label(data, variable, var_label)
}
\arguments{
\item{data}{Data frame.}

\item{variable}{Vector of column name.}

\item{var_label}{list that includes specified variable labels,
e.g. \code{var_label = list(age = "Age, yrs")}}
}
\value{
Vector variable labels.
}
\description{
Preference is given to labels specified in \code{fmt_table1(..., var_label = list())}
argument, then to a label attribute attached to the data frame
(i.e. attr(data$var, "label)), then to the variable name.
}
\examples{
# assign_var_label(mtcars, names(mtcars), list(hp = "Horsepower"))
}
\keyword{internal}
