% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bold_italicize_labels_levels.R
\name{bold_italicize_labels_levels}
\alias{bold_italicize_labels_levels}
\alias{bold_labels}
\alias{italicize_labels}
\alias{bold_levels}
\alias{italicize_levels}
\alias{bold_labels.gtsummary}
\alias{bold_levels.gtsummary}
\alias{italicize_labels.gtsummary}
\alias{italicize_levels.gtsummary}
\alias{bold_labels.tbl_cross}
\alias{bold_levels.tbl_cross}
\alias{italicize_labels.tbl_cross}
\alias{italicize_levels.tbl_cross}
\title{Bold or Italicize}
\usage{
bold_labels(x)

italicize_labels(x)

bold_levels(x)

italicize_levels(x)

\method{bold_labels}{gtsummary}(x)

\method{bold_levels}{gtsummary}(x)

\method{italicize_labels}{gtsummary}(x)

\method{italicize_levels}{gtsummary}(x)

\method{bold_labels}{tbl_cross}(x)

\method{bold_levels}{tbl_cross}(x)

\method{italicize_labels}{tbl_cross}(x)

\method{italicize_levels}{tbl_cross}(x)
}
\arguments{
\item{x}{(\code{gtsummary})
An object of class 'gtsummary'}
}
\value{
Functions return the same class of gtsummary object supplied
}
\description{
Bold or italicize labels or levels in gtsummary tables
}
\examples{
# Example 1 ----------------------------------
tbl_summary(trial, include = c("trt", "age", "response")) |>
  bold_labels() |>
  bold_levels() |>
  italicize_labels() |>
  italicize_levels()
}
\author{
Daniel D. Sjoberg
}
