% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bold_labels.R
\name{bold_labels.fmt_table1}
\alias{bold_labels.fmt_table1}
\title{Bold or unbold variable labels for \code{fmt_table1} objects in Rmarkdown}
\usage{
\method{bold_labels}{fmt_table1}(x, ...)
}
\arguments{
\item{x}{\code{fmt_table1} object}

\item{...}{further arguments passed to or from other methods}
}
\description{
Bold or unbold variable labels for \code{fmt_table1} objects in Rmarkdown
}
\examples{
trial \%>\% fmt_table1() \%>\% bold_labels()
}
