% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bold_labels.R
\name{bold_labels}
\alias{bold_labels}
\title{Bold variable labels in Rmarkdown}
\usage{
bold_labels(x, ...)
}
\arguments{
\item{x}{\code{biostatR} object}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Bold variable labels in \code{biostatR} objects by adding '__'
to each side of the string
}
\seealso{
\code{\link{bold_labels.fmt_table1}},
\code{\link{bold_labels.fmt_regression}},
\code{\link{bold_labels.fmt_uni_regression}}
}
