% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bold_levels.R
\name{bold_levels.fmt_regression}
\alias{bold_levels.fmt_regression}
\title{Bold or unbold variable levels for \code{fmt_regression} objects in Rmarkdown}
\usage{
\method{bold_levels}{fmt_regression}(x, ...)
}
\arguments{
\item{x}{\code{fmt_regression} object}

\item{...}{further arguments passed to or from other methods}
}
\description{
Bold or unbold variable levels for \code{fmt_regression} objects in Rmarkdown
}
\examples{
lm(hp ~ factor(cyl), mtcars) \%>\% fmt_regression() \%>\% bold_levels()
}
