% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bold_levels.R
\name{bold_levels.fmt_uni_regression}
\alias{bold_levels.fmt_uni_regression}
\title{Bold or unbold variable levels for \code{fmt_uni_regression} objects in Rmarkdown}
\usage{
\method{bold_levels}{fmt_uni_regression}(x, ...)
}
\arguments{
\item{x}{\code{fmt_uni_regression} object}

\item{...}{further arguments passed to or from other methods}
}
\description{
Bold or unbold variable levels for \code{fmt_uni_regression} objects in Rmarkdown
}
\examples{
trial \%>\%
  fmt_uni_regression(
    method = "glm",
    y = "response",
    method.args = list(family = binomial),
    exponentiate = TRUE
  ) \%>\%
  bold_levels()
}
