% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bold_levels.R
\name{bold_levels}
\alias{bold_levels}
\title{Bold variable levels in Rmarkdown}
\usage{
bold_levels(x, ...)
}
\arguments{
\item{x}{\code{biostatR} object}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Bold variable levels in \code{biostatR} objects by adding '__'
to each side of the string
}
\seealso{
\code{\link{bold_levels.fmt_table1}},
\code{\link{bold_levels.fmt_regression}},
\code{\link{bold_levels.fmt_uni_regression}}
}
