% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bold_p.R
\name{bold_p.fmt_regression}
\alias{bold_p.fmt_regression}
\title{Bold or unbold p-values for \code{fmt_regression}  objects in Rmarkdown}
\usage{
\method{bold_p}{fmt_regression}(x, t = 0.05, ...)
}
\arguments{
\item{x}{\code{fmt_regression} object}

\item{t}{Determines the threshold below which p-values get bolded. Default is 0.05.}

\item{...}{further arguments passed to or from other methods}
}
\description{
Bold or unbold p-values for \code{fmt_regression}  objects in Rmarkdown
}
\examples{
lm(mpg ~ hp + am, mtcars) \%>\%
  fmt_regression() \%>\%
  bold_p()
}
