% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bold_p.R
\name{bold_p.fmt_uni_regression}
\alias{bold_p.fmt_uni_regression}
\title{Bold or unbold p-values for \code{fmt_uni_regression} objects in Rmarkdown}
\usage{
\method{bold_p}{fmt_uni_regression}(x, t = 0.05, q = FALSE, ...)
}
\arguments{
\item{x}{\code{fmt_uni_regression} object}

\item{t}{Determines the threshold below which p-values get bolded. Default is 0.05.}

\item{q}{logical argument. When TRUE will bold or unbold the q-value column rather than the p-values}

\item{...}{further arguments passed to or from other methods}
}
\description{
Bold or unbold p-values for \code{fmt_uni_regression} objects in Rmarkdown
}
\examples{

trial \%>\%
  fmt_uni_regression(
    method = "lm",
    y = "age"
  ) \%>\%
  bold_p(t = 0.20)
}
