% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge_summary.R
\name{bridge_summary}
\alias{bridge_summary}
\alias{brdg_summary}
\alias{pier_summary_dichotomous}
\alias{pier_summary_categorical}
\alias{pier_summary_continuous2}
\alias{pier_summary_continuous}
\alias{pier_summary_missing_row}
\title{Summary Table Bridges}
\usage{
brdg_summary(
  x,
  calling_function = "tbl_summary",
  by = x$inputs$by,
  include = x$inputs$include,
  type = x$inputs$type
)

pier_summary_dichotomous(
  x,
  variables,
  value = x$inputs$value,
  calling_function = "tbl_summary"
)

pier_summary_categorical(
  x,
  variables,
  missing,
  missing_text,
  missing_stat,
  calling_function = "tbl_summary"
)

pier_summary_continuous2(
  x,
  variables,
  missing,
  missing_text,
  missing_stat,
  calling_function = "tbl_summary"
)

pier_summary_continuous(
  x,
  variables,
  statistic = x$inputs$statistic,
  missing,
  missing_text,
  missing_stat,
  calling_function = "tbl_summary"
)

pier_summary_missing_row(
  x,
  variables = x$inputs$include,
  missing_stat = x$inputs$missing_stat,
  missing_text = x$inputs$missing_text,
  calling_function = "tbl_summary"
)
}
\arguments{
\item{x}{gtsummary object of class \code{"tbl_summary"}}

\item{calling_function}{string indicating the name of the function that is
calling thie bridge function, e.g. \code{call_function = "tbl_summary"}. This is used
internally to organize results.}

\item{by}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
A single column from \code{data}. Summary statistics will be stratified by this variable.
Default is \code{NULL}}

\item{include}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
Variables to include in the summary table. Default is \code{everything()}}

\item{type}{(\code{\link[=syntax]{formula-list-selector}})\cr
Specifies the summary type. Accepted value are
\code{c("continuous", "continuous2", "categorical", "dichotomous")}.
If not specified, default type is assigned via
\code{assign_summary_type()}. See below for details.}

\item{variables}{character list of variables}

\item{value}{named list of values to be summarized. the names are the
variable names.}

\item{missing, missing_text, missing_stat}{Arguments dictating how and if missing values are presented:
\itemize{
\item \code{missing}: must be one of \code{c("ifany", "no", "always")}
\item \code{missing_text}: string indicating text shown on missing row. Default is \code{"Unknown"}
\item \code{missing_stat}: statistic to show on missing row. Default is \code{"{N_miss}"}.
Possible values are \code{N_miss}, \code{N_obs}, \code{N_nonmiss}, \code{p_miss}, \code{p_nonmiss}
}}

\item{statistic}{(\code{\link[=syntax]{formula-list-selector}})\cr
Specifies summary statistics to display for each variable.  The default is
\code{list(all_continuous() ~ "{median} ({p25}, {p75})", all_categorical() ~ "{n} ({p}\%)")}.
See below for details.}
}
\value{
data frame
}
\description{
Bridge function for converting \code{tbl_summary()} (and similar) cards to table bodies.
All bridge functions begin with prefix \verb{brdg_*()}.

This file also contains helper functions for constructing the bridge,
referred to as the piers (supports for a bridge) and begin with \verb{pier_*()}.
}
\examples{
tbl <-
  tbl_summary(
    data = mtcars,
    include = c("cyl", "am", "mpg", "hp"),
    type =
      list(
        cyl = "categorical",
        am = "dichotomous",
        mpg = "continuous",
        hp = "continuous2"
      ),
    value = list(am = 1),
    statistic =
      list(
        c(cyl, am) ~ "{n} ({p}\%)",
        mpg = "{mean} ({sd})",
        hp = c("{mean}", "{median}")
      )
  )

pier_summary_dichotomous(
  x = tbl,
  variables = "am",
  value = list(am = 1)
)

pier_summary_categorical(
  x = tbl,
  variables = "cyl"
)

pier_summary_continuous2(
  x = tbl,
  variables = "hp"
)

pier_summary_continuous(
  x = tbl,
  variables = "mpg"
)
}
