% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge_tbl_summary.R
\name{bridge_tbl_summary}
\alias{bridge_tbl_summary}
\alias{bridge_summary_dichotomous}
\alias{bridge_summary_categorical}
\alias{bridge_summary_continuous2}
\alias{bridge_summary_continuous}
\title{Summary Table Bridges}
\usage{
bridge_summary_dichotomous(
  x,
  variables,
  value,
  missing,
  missing_text,
  missing_stat
)

bridge_summary_categorical(x, variables, missing, missing_text, missing_stat)

bridge_summary_continuous2(x, variables, missing, missing_text, missing_stat)

bridge_summary_continuous(x, variables, missing, missing_text, missing_stat)
}
\arguments{
\item{x}{gtsummary object of class \code{"tbl_summary"}}

\item{variables}{character list of variables}

\item{value}{named list of values to be summarized. the names are the
variable names.}

\item{missing}{TODO:}

\item{missing_text}{TODO:}

\item{missing_stat}{TODO:}
}
\value{
data frame
}
\description{
Bridge functions for converting \code{tbl_summary()} cards to table bodies.
}
\examples{
tbl <-
  tbl_summary(
    data = mtcars,
    type =
      list(
        cyl = "categorical",
        am = "dichotomous",
        mpg = "continuous",
        hp = "continuous2"
      ),
    value = list(am = 1),
    statistic =
      list(
        c(cyl, am) ~ "{n} ({p}\%)",
        mpg = "{mean} ({sd})",
        hp = c("{mean}", "{median}")
      )
  )

bridge_summary_dichotomous(
 x = tbl,
 variables = "am",
 value = list(am = 1),
 missing = "ifany",
 missing_text = "Unknown",
 missing_stat = "{N_miss}"
)

bridge_summary_categorical(
 x = tbl,
 variables = "cyl",
 missing = "ifany",
 missing_text = "Unknown",
 missing_stat = "{N_miss}"
)

bridge_summary_continuous2(
 x = tbl,
 variables = "hp",
 missing = "ifany",
 missing_text = "Unknown",
 missing_stat = "{N_miss}"
)

bridge_summary_continuous(
 x = tbl,
 variables = "mpg",
 missing = "ifany",
 missing_text = "Unknown",
 missing_stat = "{N_miss}"
)
}
