% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_text.R
\name{by_level}
\alias{by_level}
\title{Convert level of by variable to column name}
\usage{
by_level(by_chr, x = x)
}
\arguments{
\item{by_chr}{a level of the by variable inline_text.tbl_regression and
inline_text.tbl_uvregression only}

\item{x}{tbl_summary object}
}
\description{
To be used in the column argument of
}
\author{
Daniel Sjoberg
}
