% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card_continuous.R
\name{card_continuous}
\alias{card_continuous}
\title{Summarize continuous variable}
\usage{
card_continuous(
  cards,
  variable,
  include,
  by = NULL,
  statistic = everything() ~ "{median} ({p25}, {p75})"
)
}
\arguments{
\item{cards}{(\code{card})\cr
An ARD object of class \code{"card"} typically created with \verb{cards::ard_*()} functions.}

\item{variable}{(\code{string})\cr
A single variable name of the continuous variable being summarized.}

\item{include}{(\code{character})\cr
Character vector of the categorical variables to}

\item{by}{(\code{string})\cr
A single variable name of the stratifying variable.}

\item{statistic}{(\code{\link[=syntax]{formula-list-selector}})\cr
Specifies summary statistics to display for each variable.  The default is
\code{everything() ~ "{median} ({p25}, {p75})"}.}
}
\value{
a gtsummary table of class \code{"card_summary"}
}
\description{
Summarize a continuous variable by one or more categorical variables
}
\examples{
library(cards)

bind_ard(
  # the primary ARD with the results
  ard_continuous(
    trial,
    # the order variables are passed here is important.
    # 'trt' is the column stratifying variable and needs to be listed first.
    by = c(trt, grade),
    variables = age
  ) ,
  # add univariate trt tabulation
  ard_categorical(
    trial,
    variables = trt
  ),
  # add missing and attributes ARD
  ard_missing(
    trial,
    by = trt,
    variables = c(grade, age)
  ),
  ard_attributes(
    trial,
    variables = c(trt, grade, age)
  )
) |>
  card_continuous(by = "trt", variable = "age", include = "grade")

bind_ard(
  # the primary ARD with the results
  ard_continuous(trial, by = grade, variables = age),
  # add missing and attributes ARD
  ard_missing(trial, variables = c(grade, age)),
  ard_attributes(trial, variables = c(grade, age))
) |>
  card_continuous(variable = "age", include = "grade")
}
