% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card_summary.R
\name{card_summary}
\alias{card_summary}
\title{ARD summary table}
\usage{
card_summary(
  cards,
  statistic = list(all_continuous() ~ "{median} ({p25}, {p75})", all_categorical() ~
    "{n} ({p}\%)"),
  type = NULL,
  missing = c("ifany", "no", "always"),
  missing_text = "Unknown",
  missing_stat = "{N_miss}",
  include = everything()
)
}
\arguments{
\item{cards}{(\code{card})\cr
An ARD object of class \code{"card"} typically created with \verb{cards::ard_*()} functions.}

\item{statistic}{(\code{\link[=syntax]{formula-list-selector}})\cr
Used to specify the summary statistics for each variable.
Each of the statistics must be present in \code{card} as no new statistics are calculated
in this function.
The default is
\code{list(all_continuous() ~ "{median} ({p25}, {p75})", all_categorical() ~ "{n} ({p}\%)")}.}

\item{type}{(\code{\link[=syntax]{formula-list-selector}})\cr
Specifies the summary type. Accepted value are
\code{c("continuous", "continuous2", "categorical", "dichotomous")}.
Continuous summaries may be assigned \code{c("continuous", "continuous2")}, while
categorical and dichotomous cannot be modified.}

\item{missing, missing_text, missing_stat}{Arguments dictating how and if missing values are presented:
\itemize{
\item \code{missing}: must be one of \code{c("ifany", "no", "always")}
\item \code{missing_text}: string indicating text shown on missing row. Default is \code{"Unknown"}
\item \code{missing_stat}: statistic to show on missing row. Default is \code{"{N_miss}"}.
Possible values are \code{N_miss}, \code{N_obs}, \code{N_nonmiss}, \code{p_miss}, \code{p_nonmiss}
}}

\item{include}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
Variables to include in the summary table. Default is \code{everything()}}
}
\value{
a gtsummary table of class \code{"card_summary"}
}
\description{
The \code{card_summary()} function tables descriptive statistics for
continuous, categorical, and dichotomous variables.
The functions accepts an ARD object.
}
\examples{
library(cards)

ard_stack(
  data = ADSL,
  by = NULL,
  ard_categorical(variables = "AGEGR1"),
  ard_continuous(variables = "AGE"),
  .attributes = TRUE,
  .missing = TRUE
) |>
  card_summary()

ard_stack(
  data = ADSL,
  by = ARM,
  ard_categorical(variables = "AGEGR1"),
  ard_continuous(variables = "AGE"),
  .attributes = TRUE,
  .missing = TRUE
) |>
  card_summary()
}
