% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-checks.R
\name{check_class}
\alias{check_class}
\alias{check_class_data_frame}
\title{Check Argument Class}
\usage{
check_class(
  x,
  class,
  allow_empty = FALSE,
  arg_name = rlang::caller_arg(x),
  call = rlang::caller_env()
)

check_class_data_frame(
  x,
  arg_name = rlang::caller_arg(x),
  call = rlang::caller_env()
)
}
\arguments{
\item{x}{object whose class will be checked}

\item{class}{character vector or string indicating accepted classes.
Passed to \code{inherits(what=class)}}

\item{allow_empty}{logical indicating whether empty arguments are allowed.
Empty is defined by \code{rlang::is_empty()}, where empty elements include
\code{NULL}, \code{list()}, \code{character()}, etc. Default is \code{FALSE}}

\item{arg_name}{string indicating the argument name. Default is \code{rlang::caller_arg(x)}}

\item{call}{The execution environment of a currently running
function, e.g. \code{call = caller_env()}. The corresponding function
call is retrieved and mentioned in error messages as the source
of the error.

You only need to supply \code{call} when throwing a condition from a
helper function which wouldn't be relevant to mention in the
message.

Can also be \code{NULL} or a \link[rlang:topic-defuse]{defused function call} to
respectively not display any call or hard-code a code to display.

For more information about error calls, see \ifelse{html}{\link[rlang:topic-error-call]{Including function calls in error messages}}{\link[rlang:topic-error-call]{Including function calls in error messages}}.}

\item{allow_null}{Logical indicating whether a NULL value will pass the test.
Default is \code{FALSE}}
}
\description{
Check Argument Class
}
\keyword{internal}
