% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-checks.R
\name{check_not_missing}
\alias{check_not_missing}
\title{Check Argument not Missing}
\usage{
check_not_missing(
  x,
  arg_name = rlang::caller_arg(x),
  call = rlang::caller_env()
)
}
\arguments{
\item{x}{argument to check}

\item{arg_name}{string indicating the name of the argument. Used in the error messaging.}

\item{call}{The execution environment of a currently running
function, e.g. \code{call = caller_env()}. The corresponding function
call is retrieved and mentioned in error messages as the source
of the error.

You only need to supply \code{call} when throwing a condition from a
helper function which wouldn't be relevant to mention in the
message.

Can also be \code{NULL} or a \link[rlang:topic-defuse]{defused function call} to
respectively not display any call or hard-code a code to display.

For more information about error calls, see \ifelse{html}{\link[rlang:topic-error-call]{Including function calls in error messages}}{\link[rlang:topic-error-call]{Including function calls in error messages}}.}
}
\description{
Check Argument not Missing
}
\keyword{internal}
