% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cols_label_summary.R
\name{cols_label_summary}
\alias{cols_label_summary}
\title{Relabel columns in summary table}
\usage{
cols_label_summary(x, stat_overall = NULL, stat_by = NULL)
}
\arguments{
\item{x}{\code{tbl_summary} object}

\item{stat_overall}{string vector including text to appear above the overall summary
statistics. \code{N}, the total number of observations, is available for use in the
description. e.g. \code{stat_overall = md("**All Patients**, N = {N}")}}

\item{stat_by}{string vector of text to include above the summary statistics
stratified by a variable.  The following fields are available for use in the
headers: \code{n}, \code{N}, \code{p}, and \code{level}.  \code{n} is the number of observations in
each by group. \code{N} is the total number of observations. \code{p} is the percent
in a by group. \code{level} is the by variable level.
Syntax follows the \code{glue::\link{glue}} function,
e.g. \code{stat_by = md("**{level}**, N = {n} ({style_percent(p, symbol = TRUE)})")}.
Must specify \code{by} along with \code{stat_by}.}
}
\description{
Column labels can be modified using data from the data frame
}
\examples{
tbl_col1 <-
  trial \%>\%
  tbl_summary() \%>\%
  cols_label_summary(stat_overall = md("**All Patients**, N = {N}"))
tbl_col2 <-
  trial \%>\%
  tbl_summary(by = "trt") \%>\%
  cols_label_summary(stat_by = md("**{level}**, N = {n} ({style_percent(p, symbol = TRUE)})"))
}
\seealso{
\link{tbl_summary}, \link[gt]{md}, \link[gt]{html}
}
\author{
Daniel Sjoberg
}
