% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_table_body.R
\name{construct_gtsummary}
\alias{construct_gtsummary}
\alias{construct_gtsummary.default}
\alias{construct_gtsummary.tbl_summary}
\title{Construct Table Body}
\usage{
construct_gtsummary(x)

\method{construct_gtsummary}{default}(x)

\method{construct_gtsummary}{tbl_summary}(x)
}
\arguments{
\item{x}{gtsummary object}
}
\value{
gtsummary object
}
\description{
Construct Table Body
}
\examples{
# TODO: Add example
}
