% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-tbl_custom_summary.R
\name{continuous_summary}
\alias{continuous_summary}
\title{Summarize a continuous variable}
\usage{
continuous_summary(variable)
}
\arguments{
\item{variable}{(\code{string})\cr
String indicating the name of the variable to be summarized. This
variable should be continuous.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}\cr
This helper, to be used with \code{\link[=tbl_custom_summary]{tbl_custom_summary()}}, creates a function
summarizing a continuous variable.
}
\details{
When using \code{continuous_summary()}, you can specify in the \verb{statistic=} argument
of \code{\link[=tbl_custom_summary]{tbl_custom_summary()}} the same continuous statistics than in
\code{\link[=tbl_summary]{tbl_summary()}}. See the \emph{statistic argument} section of the help file of
\code{\link[=tbl_summary]{tbl_summary()}}.
}
\author{
Joseph Larmarange
}
\keyword{internal}
