% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{modify_column_indent}
\alias{tbl_split}
\alias{tbl_split.gtsummary}
\title{Deprecated functions}
\usage{
modify_column_indent(...)

tbl_split(x, ...)

\method{tbl_split}{gtsummary}(...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}\cr
Some functions have been deprecated and are no longer being actively
supported.
}
\keyword{internal}
