% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-tbl_summary.R
\name{df_by}
\alias{df_by}
\title{Simple utility function to get extract and calculate additional information
about the 'by' variable in \code{\link{tbl_summary}}}
\usage{
df_by(data, by)
}
\arguments{
\item{data}{data frame}

\item{by}{character name of the \code{by} variable found in data}
}
\description{
Given a dataset and the name of the 'by' variable, this function returns a
data frame with unique levels of the by variable, the by variable ID, a character
version of the levels, and the column name for each level in the \code{\link{tbl_summary}}
output data frame.
}
\author{
Daniel D. Sjoberg
}
\keyword{internal}
