% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim.R
\name{dim.gtsummary}
\alias{dim.gtsummary}
\title{gtsummary table dimension}
\usage{
\method{dim}{gtsummary}(x)
}
\arguments{
\item{x}{(\code{gtsummary})\cr
a 'gtsummary' table}
}
\value{
integer vector
}
\description{
Returns the dimension of a gtsummary table, that is, the number of rows
and the number of un-hidden columns.

\code{nrow()} calls \code{dim()}; therefore, \code{nrow()} will also work on gtsummary tables.
}
\examples{
tbl <- tbl_summary(trial, include = age, by = trt)

dim(tbl)
nrow(tbl)
}
\keyword{internal}
