% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-add_p.R
\name{.assign_test_tbl_svysummary}
\alias{.assign_test_tbl_svysummary}
\title{Assign test in \code{add_p.tbl_svysummary()}}
\usage{
.assign_test_tbl_svysummary(data, variable, summary_type, by, test)
}
\arguments{
\item{data}{a survey object}

\item{variable}{character variable name}

\item{summary_type}{summary type}

\item{by}{character by variable}

\item{test}{named list of user-supplied tests}
}
\description{
Returns either the user-supplied test, or the default test given the variable
type
}
