% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-add_p.R
\name{.run_add_p_test_fun}
\alias{.run_add_p_test_fun}
\title{Calculate p-values}
\usage{
.run_add_p_test_fun(
  x,
  data,
  variable,
  by = NULL,
  group = NULL,
  type = NULL,
  test.args = NULL
)
}
\arguments{
\item{x}{list returned from \code{.get_add_p_test_fun()}}

\item{data}{data frame/object to calculate p-values from}

\item{variable}{character variable name}

\item{by}{character by variable name}

\item{group}{optional group variable}

\item{type}{optional type variable}
}
\description{
Calculate p-values
}
