% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_tbl_hierarchical.R
\name{filter_tbl_hierarchical}
\alias{filter_tbl_hierarchical}
\alias{tbl_filter}
\alias{tbl_filter.tbl_hierarchical}
\title{Filter Hierarchical Tables}
\usage{
tbl_filter(x, ...)

\method{tbl_filter}{tbl_hierarchical}(x, t, gt = TRUE, eq = FALSE, .stat = "n")
}
\arguments{
\item{x}{(\code{tbl_hierarchical})\cr
A hierarchical gtsummary table of class \code{'tbl_hierarchical'}.}

\item{t}{(scalar \code{numeric})\cr
Threshold used to determine which values will be retained.}

\item{gt}{(scalar \code{logical})\cr
Whether to filter for row sums greater than \code{t} or less than \code{t}. Default is greater than (\code{gt = TRUE}).}

\item{eq}{(scalar \code{logical})\cr
Whether to include the value of \code{t} in the filtered range, i.e. whether to use exclusive comparators (\code{>}, \code{<}) or
inclusive comparators (\code{>=}, \code{<=}) when filtering. Default is \code{FALSE}.}

\item{.stat}{(\code{string})\cr
Statistic to use for sorting when \code{sort = "frequency"}. This statistic must be present in the table for all
hierarchy levels.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}\cr

This function is used to filter hierarchical tables by total row frequencies.
}
\examples{
ADAE_subset <- cards::ADAE |>
  dplyr::filter(
    AESOC \%in\% unique(cards::ADAE$AESOC)[1:5],
    AETERM \%in\% unique(cards::ADAE$AETERM)[1:5]
  )

tbl <- tbl_hierarchical(
  data = ADAE_subset,
  variables = c(SEX, RACE, AETERM),
  by = TRTA,
  denominator = cards::ADSL |> mutate(TRTA = ARM),
  id = USUBJID,
  overall_row = TRUE
)

# Example 1 - Row Sums > 10 ------------------
tbl_filter(tbl, t = 10)

# Example 2 - Row Sums <= 5 ------------------
tbl_filter(tbl, t = 10, gt = FALSE, eq = TRUE)
}
