% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt_beta.R
\name{fmt_beta}
\alias{fmt_beta}
\title{Round and format regression model coefficients}
\usage{
fmt_beta(x)
}
\arguments{
\item{x}{numeric vector}
}
\description{
Takes in a vector of regression model coefficients, and returns a
rounded and formatted coefficient vector.  This is the default
rounding for models formatted with \code{biostatR::fmt_regression()}.  Coefficients
larger than 100 are rounded to the nearest integer, between  10 and 100 to one
decimal place, and all others to 2 decimal places.  These thresholds were
selected with odds ratios and hazard ratios in mind.
}
\examples{
or <- c(100.2342, 11.234, 5.32423, 1.23423469, 0.234)
fmt_beta(or)
}
