% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt_bold_p.R
\name{fmt_bold_p}
\alias{fmt_bold_p}
\title{Bold significant p-values}
\usage{
fmt_bold_p(x, ...)
}
\arguments{
\item{x}{an object created using gtsummary functions}

\item{...}{further arguments passed to other methods.}
}
\description{
Bold p-values in tables create by gtsummary
}
\author{
Daniel Sjoberg
}
