% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt_bold_p.R
\name{fmt_bold_p.tbl_regression}
\alias{fmt_bold_p.tbl_regression}
\title{Bold significant p-values}
\usage{
\method{fmt_bold_p}{tbl_regression}(x, t = 0.05, ...)
}
\arguments{
\item{x}{an object created using \code{tbl_regression} function}

\item{t}{Determines the threshold below which p-values get bolded. Default is 0.05.}

\item{...}{not used}
}
\description{
Bold p-values in tables created by \code{\link{tbl_regression}}
}
\author{
Daniel Sjoberg
}
