% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt_bold_p.R
\name{fmt_bold_p.tbl_summary}
\alias{fmt_bold_p.tbl_summary}
\title{Bold significant p-values}
\usage{
\method{fmt_bold_p}{tbl_summary}(x, t = 0.05, q = FALSE, ...)
}
\arguments{
\item{x}{an object created using \code{tbl_summary} function}

\item{t}{Determines the threshold below which p-values get bolded. Default is 0.05.}

\item{q}{logical argument. When TRUE will bold or unbold the q-value column rather than the p-values}

\item{...}{not used}
}
\description{
Bold p-values in tables created by \code{\link{tbl_summary}}
}
\author{
Daniel Sjoberg
}
