% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt_percent.R
\name{fmt_percent}
\alias{fmt_percent}
\title{Formats percentages to be displayed in tables or text of report.}
\usage{
fmt_percent(x, symbol = FALSE)
}
\arguments{
\item{x}{numeric vector of percentages}

\item{symbol}{logical indicator to include percent symbol in output.  Default is FALSE.}
}
\value{
Formatted percentages
}
\description{
Formats percentages to be displayed in tables or text of report.
}
\examples{
percent_vals <- c(-1, 0, 0.0001, 0.005, 0.01, 0.10, 0.45356, 0.99, 1.45)
fmt_percent(percent_vals)
fmt_percent(percent_vals, symbol = TRUE)
}
\author{
Daniel Sjoberg
}
