% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-table1_header.R
\name{fmt_table1_header}
\alias{fmt_table1_header}
\title{This function creates header rows for \code{fmt_table1} objects}
\usage{
fmt_table1_header(data, by, pvalue)
}
\arguments{
\item{data}{Data frame}

\item{by}{Character variable name in \code{data} that Summary statistics for
\code{variable} are stratified}

\item{pvalue}{logical indicator whether p-values will be reported.}
}
\description{
This function creates header rows for \code{fmt_table1} objects
}
\keyword{internal}
