% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_list.R
\name{formula_list}
\alias{formula_list}
\title{Formula-list Notation}
\description{
Many arguments throughout the gtsummary package accept list and
formula notation, e.g. \code{tbl_summary(statistic=)}. Below enumerates a few
tips and shortcuts for using the list and formulas.
\enumerate{
\item \strong{List of Formulas}

Typical usage includes a list of formulas, where the LHS is a variable
name or a selector.\if{html}{\out{<div class="sourceCode r">}}\preformatted{tbl_summary(statistic = list(age ~ "\{mean\}", all_categorical() ~ "\{n\}"))
}\if{html}{\out{</div>}}
\item \strong{Named List}

You may also pass a named list; however, the tidyselect and gtsummary selectors
are not supported with this syntax.\if{html}{\out{<div class="sourceCode r">}}\preformatted{tbl_summary(statistic = list(age = "\{mean\}", response = "\{n\}"))
}\if{html}{\out{</div>}}
\item \strong{Hybrid Named List/List of Formulas}

Pass a combination of formulas and named elements\if{html}{\out{<div class="sourceCode r">}}\preformatted{tbl_summary(statistic = list(age = "\{mean\}", all_categorical() ~ "\{n\}"))
}\if{html}{\out{</div>}}
\item \strong{Shortcuts}

You can pass a single formula, which is equivalent to passing the formula
in a list.\if{html}{\out{<div class="sourceCode r">}}\preformatted{tbl_summary(statistic = all_categorical() ~ "\{n\}")
}\if{html}{\out{</div>}}

As a shortcut to select all variables, you can omit the LHS of the formula.
The two calls below are equivalent.\if{html}{\out{<div class="sourceCode r">}}\preformatted{tbl_summary(statistic = ~"\{n\}")
tbl_summary(statistic = everything() ~ "\{n\}")
}\if{html}{\out{</div>}}
}
}
