% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gastric}
\alias{gastric}
\title{Results from a simulated study of patients with gastric cancer}
\format{A data frame with 200 rows--one row per patient
\describe{
\item{trt}{Treatment Received}
\item{age}{Age}
\item{albumin}{Albumin}
\item{stage}{T Stage}
\item{grade}{Grade}
\item{response}{Tumor Response}
\item{death}{Patient Died}
\item{ttdeath}{Months to Death/Censor}
}}
\usage{
gastric
}
\description{
A dataset containing the baseline characteristics of 200 patients
with gastric cancer treated with Chemotherapy or
Chemotherapy and Radiation
}
\keyword{datasets}
