% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-get_by_info.R
\name{get_by_info}
\alias{get_by_info}
\title{Simple utility function to get extract and calculate additional information
about the 'by' variable in \code{\link{fmt_table1}}}
\usage{
get_by_info(data, by)
}
\arguments{
\item{data}{data frame}

\item{by}{character name of the \code{by} variable found in data}
}
\description{
Given a dataset and the name of the 'by' variable, this function returns a
data frame with unique levels of the by variable, the by variable ID, a character
version of the levels, and the column name for each level in the \code{\link{fmt_table1}}
output data frame.
}
\keyword{internal}
