% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cards.R
\name{get_cards}
\alias{get_cards}
\title{Extract ARDs}
\usage{
get_cards(x)
}
\arguments{
\item{x}{(\code{gtsummary})\cr a gtsummary table}
}
\value{
list
}
\description{
Extract the ARDs from a gtsummary table. If needed, results may be combined
with \code{cards::bind_ard()}.
}
\examples{
\dontshow{if (gtsummary:::is_pkg_installed('cardx', reference_pkg = 'gtsummary')) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
tbl_summary(trial, by = trt, include = age) |>
  add_overall() |>
  add_p() |>
  get_cards()

glm(response ~ trt, data = trial, family = binomial()) |>
  tbl_regression() |>
  get_cards()
\dontshow{\}) # examplesIf}
}
