% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtsummary-package.R
\docType{package}
\name{gtsummary-package}
\alias{gtsummary}
\alias{gtsummary-package}
\title{gtsummary: Presentation-Ready Data Summary and Analytic Result Tables}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Creates presentation-ready tables summarizing data sets, regression models, and more. The code to create the tables is concise and highly customizable. Data frames can be summarized with any function, e.g. mean(), median(), even user-written functions. Regression models are summarized and include the reference rows for categorical variables. Common regression models, such as logistic regression and Cox proportional hazards regression, are automatically identified and the tables are pre-filled with appropriate column headers.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ddsjoberg/gtsummary}
  \item \url{https://www.danieldsjoberg.com/gtsummary/}
  \item Report bugs at \url{https://github.com/ddsjoberg/gtsummary/issues}
}

}
\author{
\strong{Maintainer}: Daniel D. Sjoberg \email{danield.sjoberg@gmail.com} (\href{https://orcid.org/0000-0003-0862-2018}{ORCID})

Authors:
\itemize{
  \item Daniel D. Sjoberg \email{danield.sjoberg@gmail.com} (\href{https://orcid.org/0000-0003-0862-2018}{ORCID})
  \item Joseph Larmarange (\href{https://orcid.org/0000-0001-7097-700X}{ORCID})
  \item Michael Curry (\href{https://orcid.org/0000-0002-0261-4044}{ORCID})
  \item Emily de la Rua (\href{https://orcid.org/0009-0000-8738-5561}{ORCID})
  \item Jessica Lavery (\href{https://orcid.org/0000-0002-2746-5647}{ORCID})
  \item Karissa Whiting (\href{https://orcid.org/0000-0002-4683-1868}{ORCID})
  \item Emily C. Zabor (\href{https://orcid.org/0000-0002-1402-4498}{ORCID})
}

Other contributors:
\itemize{
  \item Xing Bai [contributor]
  \item Malcolm Barrett (\href{https://orcid.org/0000-0003-0299-5825}{ORCID}) [contributor]
  \item Esther Drill (\href{https://orcid.org/0000-0002-3315-4538}{ORCID}) [contributor]
  \item Jessica Flynn (\href{https://orcid.org/0000-0001-8310-6684}{ORCID}) [contributor]
  \item Margie Hannum (\href{https://orcid.org/0000-0002-2953-0449}{ORCID}) [contributor]
  \item Stephanie Lobaugh [contributor]
  \item Shannon Pileggi (\href{https://orcid.org/0000-0002-7732-4164}{ORCID}) [contributor]
  \item Amy Tin (\href{https://orcid.org/0000-0002-8005-0694}{ORCID}) [contributor]
  \item Gustavo Zapata Wainberg (\href{https://orcid.org/0000-0002-2524-3637}{ORCID}) [contributor]
}

}
\keyword{internal}
