% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtsummary_logo.R
\name{gtsummary_logo}
\alias{gtsummary_logo}
\title{The gtsummary logo, using ASCII or Unicode characters}
\usage{
gtsummary_logo(unicode = l10n_info()$`UTF-8`)
}
\arguments{
\item{unicode}{Whether to use Unicode symbols. Default is \code{TRUE}
on UTF-8 platforms.}
}
\description{
Use \code{\link[crayon:strip_style]{crayon::strip_style()}} to get rid of the colors.
}
\examples{
gtsummary_logo()
}
