% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head.R
\name{head}
\alias{head}
\alias{head.gtsummary}
\alias{tail.gtsummary}
\alias{[.gtsummary}
\title{Display the first or last rows}
\usage{
\method{head}{gtsummary}(x, n = 6L, ...)

\method{tail}{gtsummary}(x, n = 6L, ...)

\method{[}{gtsummary}(x, ...)
}
\arguments{
\item{x}{A \code{gtsummary} object.}

\item{n}{Number of rows to return. Default is \code{6L} as default in \code{\link[utils:head]{utils::head()}} and
\code{\link[utils:head]{utils::tail()}}.}

\item{...}{Additional arguments passed to \code{\link[utils:head]{utils::head()}}, \code{\link[utils:head]{utils::tail()}}, or \code{[}.}
}
\value{
A \code{gtsummary} object with only the first or last \code{n} rows in \code{table_body}.
}
\description{
These functions allow you to view the first (\code{head()}) or last (\code{tail()}) \code{n}
rows of a gtsummary table.
}
\examples{
tbl <- tbl_summary(trial, by = trt, include = grade)

# Example 1 head() ---------------------------
head(tbl, n = 2L)

# Example 2 tail() ---------------------------
tail(tbl, n = 2L)

# Example 3 `[`() ---------------------------
tbl[1:2,]
}
\keyword{internal}
