% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indent_key.R
\name{indent_key}
\alias{indent_key}
\title{Make indices compatible with \code{\link[kableExtra]{add_indent}}}
\usage{
indent_key(x, ...)
}
\arguments{
\item{x}{\code{fmt_table1}, \code{fmt_regression}, or \code{fmt_uni_regression} object}

\item{...}{further arguments passed to or from other methods.}
}
\value{
vector for use in kableExtra::add_indent() or otherwise
}
\description{
Creates vector indicating rows to be indented from \code{fmt_table1} or \code{fmt_regression}
output that can be passed to \code{kableExtra::add_indent}.
}
