% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/italicize_labels.R
\name{italicize_labels.fmt_table1}
\alias{italicize_labels.fmt_table1}
\title{Italicize or unitalicize labels for \code{fmt_table1} objects in Rmarkdown}
\usage{
\method{italicize_labels}{fmt_table1}(x, ...)
}
\arguments{
\item{x}{\code{fmt_table1} object}

\item{...}{further arguments passed to or from other methods}
}
\description{
Italicize or unitalicize labels for \code{fmt_table1} objects in Rmarkdown
}
\examples{
trial \%>\% fmt_table1() \%>\% italicize_labels()
}
