% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/italicize_labels.R
\name{italicize_labels}
\alias{italicize_labels}
\title{Italicize variable labels in Rmarkdown}
\usage{
italicize_labels(x, ...)
}
\arguments{
\item{x}{\code{biostatR} object}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Italicize variable labels in \code{biostatR} objects by adding '_'
to each side of the string
}
\seealso{
\code{\link{italicize_labels.fmt_table1}},
\code{\link{italicize_labels.fmt_regression}},
\code{\link{italicize_labels.fmt_uni_regression}}
}
