% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/italicize_levels.R
\name{italicize_levels.fmt_regression}
\alias{italicize_levels.fmt_regression}
\title{Italicize or unitalicize levels for \code{fmt_regression} objects in Rmarkdown}
\usage{
\method{italicize_levels}{fmt_regression}(x, ...)
}
\arguments{
\item{x}{\code{fmt_regression} object}

\item{...}{further arguments passed to or from other methods}
}
\description{
Italicize or unitalicize levels for \code{fmt_regression} objects in Rmarkdown
}
\examples{
lm(hp ~ factor(cyl), mtcars) \%>\% fmt_regression() \%>\% italicize_levels()
}
