% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/italicize_levels.R
\name{italicize_levels.fmt_uni_regression}
\alias{italicize_levels.fmt_uni_regression}
\title{Italicize or unitalicize levels for \code{fmt_uni_regression} objects in Rmarkdown}
\usage{
\method{italicize_levels}{fmt_uni_regression}(x, ...)
}
\arguments{
\item{x}{\code{fmt_uni_regression} object}

\item{...}{further arguments passed to or from other methods}
}
\description{
Italicize or unitalicize levels for \code{fmt_uni_regression} objects in Rmarkdown
}
\examples{
trial \%>\%
  fmt_uni_regression(
    method = "glm",
    y = "response",
    method.args = list(family = binomial),
    exponentiate = TRUE
  ) \%>\%
  italicize_levels()
}
