% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/italicize_levels.R
\name{italicize_levels}
\alias{italicize_levels}
\title{Italicize variable levels in Rmarkdown}
\usage{
italicize_levels(x, ...)
}
\arguments{
\item{x}{\code{biostatR} object}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Italicize variable levels in \code{biostatR} objects by adding '_'
to each side of the string
}
\seealso{
\code{\link{italicize_levels.fmt_table1}},
\code{\link{italicize_levels.fmt_regression}},
\code{\link{italicize_levels.fmt_uni_regression}}
}
