% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_kable_extra.R
\name{kableExtra_utils}
\alias{kableExtra_utils}
\alias{.escape_html}
\alias{.escape_latex}
\alias{.escape_latex2}
\title{Special Character Escape}
\usage{
.escape_html(x)

.escape_latex(x, newlines = TRUE, align = "c")

.escape_latex2(x, newlines = TRUE, align = "c")
}
\arguments{
\item{x}{character vector}
}
\value{
character vector with escaped special characters
}
\description{
These utility functions were copied from the internals of kableExtra,
and assist in escaping special characters in LaTeX and HTML tables.
These function assist in the creations of tables via \code{as_kable_extra()}.
}
\examples{
\dontshow{if (gtsummary:::is_pkg_installed("kableExtra")) withAutoprint(\{ # examplesIf}
.escape_latex(c("\%", "{test}"))
.escape_html(c(">0.9", "line\nbreak"))
\dontshow{\}) # examplesIf}
}
\seealso{
\code{as_kable_extra()}
}
\keyword{internal}
