% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_print.R
\name{knit_print.fmt_regression}
\alias{knit_print.fmt_regression}
\title{Print \code{fmt_regression} objects in Rmarkdown}
\usage{
\method{knit_print}{fmt_regression}(x, options, ...)
}
\arguments{
\item{x}{object of class \code{fmt_regression} object from \code{\link{fmt_regression}} function}

\item{options}{Copied from the printr package....should we delete?}

\item{...}{further arguments passed to \code{\link[knitr]{kable}}.}
}
\description{
Print \code{fmt_regression} objects in Rmarkdown
}
