% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_print.R
\name{knit_print}
\alias{knit_print}
\title{A custom printing function for Rmarkdown}
\description{
The S3 generic function knit_print is the default printing function in knitr.
The chunk option render uses this function by default. The main purpose of this
S3 generic function is to customize printing of R objects in code chunks. We
can fall back to the normal printing behavior by setting the chunk
option \code{render = normal_print}. See \code{\link[knitr]{knit_print}} for details.
}
