% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{knit_print.tbl_uvregression}
\alias{knit_print.tbl_uvregression}
\title{Print gtsummary objects}
\usage{
\method{knit_print}{tbl_uvregression}(x, ...)
}
\arguments{
\item{x}{object of class \code{tbl_uvregression} object from \code{\link{tbl_uvregression}} function}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print gtsummary objects
}
