% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_header.R
\name{modify_header}
\alias{modify_header}
\title{Modifies header rows for existing \code{fmt_table1}, \code{fmt_regression}, and
\code{fmt_uni_regression} objects}
\usage{
modify_header(x, ...)
}
\arguments{
\item{x}{\code{fmt_table1}, \code{fmt_regression}, or \code{fmt_uni_regression} object}

\item{...}{further arguments passed to or from other methods.}
}
\description{
An S3 generic function. If the top row of a header column is blank,
you may experience printing issues when using \code{knitr}::\code{\link[knitr]{kable}}
}
\seealso{
\code{\link{modify_header.fmt_table1}},
\code{\link{modify_header.fmt_regression}},
\code{\link{modify_header.fmt_uni_regression}}
}
