% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-parse_terms.R
\name{parse_terms}
\alias{parse_terms}
\title{Convert a regression model object to a parsed list matching model terms to variable names}
\usage{
parse_terms(x, tidy_model, show_yesno)
}
\arguments{
\item{x}{regression model object}

\item{tidy_model}{tidied model output from tidy_wrap function}

\item{show_yesno}{categorical and factor variables that are \code{c("No", "Yes")} or
\code{c("no", "yes")} default to dichotomous printing. To force both levels to be shown}
}
\value{
list with one entry for each variable, each element are the terms from
the regression model.For categorical variables, the reference level has been added.
}
\description{
With the \code{parse_terms()} function, each term in the tidied model is parsed and
matched with a variable name.
}
\keyword{internal}
