% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.fmt_regression}
\alias{print.fmt_regression}
\title{Print \code{fmt_regression} objects}
\usage{
\method{print}{fmt_regression}(x, ...)
}
\arguments{
\item{x}{object of class \code{fmt_regression} object from
\code{\link{fmt_regression}} function}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print \code{fmt_regression} objects
}
