% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.fmt_table1}
\alias{print.fmt_table1}
\title{Print \code{fmt_table1} objects}
\usage{
\method{print}{fmt_table1}(x, ...)
}
\arguments{
\item{x}{object of class \code{fmt_table1} object from \code{\link{fmt_table1}} function}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print \code{fmt_table1} objects
}
