% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.tbl_regression}
\alias{print.tbl_regression}
\title{Print gtsummary objects}
\usage{
\method{print}{tbl_regression}(x, ...)
}
\arguments{
\item{x}{object of class \code{tbl_regression} object from \code{\link{tbl_regression}} function}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print gtsummary objects
}
