% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.tbl_summary}
\alias{print.tbl_summary}
\title{Print gtsummary objects}
\usage{
\method{print}{tbl_summary}(x, ...)
}
\arguments{
\item{x}{object of class \code{tbl_summary} object from \code{\link{tbl_summary}} function}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print gtsummary objects
}
