% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.tbl_uregression}
\alias{print.tbl_uregression}
\title{Print gtsummary objects}
\usage{
\method{print}{tbl_uregression}(x, ...)
}
\arguments{
\item{x}{object of class \code{tbl_uregression} object from \code{\link{tbl_uregression}} function}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print gtsummary objects
}
