% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print_gtsummary}
\alias{print_gtsummary}
\alias{print.gtsummary}
\alias{knit_print.gtsummary}
\alias{pkgdown_print.gtsummary}
\title{print and knit_print methods for gtsummary objects}
\usage{
\method{print}{gtsummary}(
  x,
  print_engine = c("gt", "flextable", "huxtable", "kable", "kable_extra", "tibble"),
  ...
)

\method{knit_print}{gtsummary}(
  x,
  print_engine = c("gt", "flextable", "huxtable", "kable", "kable_extra", "tibble"),
  ...
)

\method{pkgdown_print}{gtsummary}(x, visible = TRUE)
}
\arguments{
\item{x}{An object created using gtsummary functions}

\item{print_engine}{String indicating the print method. Must be one of
\code{"gt"}, \code{"kable"}, \code{"kable_extra"}, \code{"flextable"}, \code{"tibble"}}

\item{...}{Not used}
}
\description{
print and knit_print methods for gtsummary objects
}
\author{
Daniel D. Sjoberg
}
\keyword{internal}
