% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_to_name.R
\name{row_to_name}
\alias{row_to_name}
\title{Renames columns to values in a row of data frame}
\usage{
row_to_name(data, row_n = 1, delete_row = TRUE)
}
\arguments{
\item{data}{input data set}

\item{row_n}{row number to take names from. Default is row 1}

\item{delete_row}{logical argument whether to delete row after renaming.
Default is \code{TRUE}}
}
\description{
Renames columns to values in a row of data frame
}
