% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_helpers.R
\name{select_prep}
\alias{select_prep}
\title{Table Body Select Prep}
\usage{
select_prep(table_body, data = NULL)
}
\arguments{
\item{table_body}{a data frame from \code{.$table_body}}

\item{data}{an optional data frame the attributes will be added to}
}
\value{
a data frame
}
\description{
This function uses the information in \code{.$table_body} and adds them
as attributes to \code{data} (if passed). Once they've been assigned as
proper gtsummary attributes, gtsummary selectors like \code{all_continuous()}
will work properly.
}
\keyword{internal}
