% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_p.R
\name{sort_p}
\alias{sort_p}
\title{Sort variables in table by ascending p-values}
\usage{
sort_p(x, ...)
}
\arguments{
\item{x}{an object created using gtsummary functions}

\item{...}{further arguments passed to other methods.}
}
\description{
Sort tables created by gtsummary by p-values
}
\seealso{
\code{\link{sort_p.tbl_summary}},
\code{\link{sort_p.tbl_regression}},
\code{\link{sort_p.tbl_uvregression}}
}
\author{
Karissa Whiting
}
\keyword{internal}
