% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_tbl_hierarchical.R
\name{sort_tbl_hierarchical}
\alias{sort_tbl_hierarchical}
\alias{tbl_sort}
\alias{tbl_sort.tbl_hierarchical}
\title{Sort Hierarchical Tables}
\usage{
tbl_sort(x, ...)

\method{tbl_sort}{tbl_hierarchical}(x, sort = "descending", ...)
}
\arguments{
\item{x}{(\code{tbl_hierarchical}, \code{tbl_hierarchical_count})\cr
a hierarchical gtsummary table of class \code{'tbl_hierarchical'} or \code{'tbl_hierarchical_count'}.}

\item{...}{These dots are for future extensions and must be empty.}

\item{sort}{(\code{string})\cr
type of sorting to perform. Value must be one of:
\itemize{
\item \code{"alphanumeric"} - within each hierarchical section of the ARD, rows are ordered alphanumerically (i.e. A to Z)
by \code{variable_label} text.
\item \code{"descending"} - within each hierarchical section of the ARD, count sums are calculated for each row and rows are
sorted in descending order by sum. If \code{sort = "descending"}, the \code{n} statistic is used to calculate row sums if
included in \code{statistic} for all variables, otherwise \code{p} is used. If neither \code{n} nor \code{p} are present in \code{x} for
all variables, an error will occur.
}

Defaults to \code{"descending"}.}
}
\value{
A \code{gtsummary} of the same class as \code{x}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}\cr

This function is used to sort hierarchical tables. Options for sorting criteria are:
\enumerate{
\item Descending - within each section of the hierarchy table, count sums are calculated for each row and rows are
sorted in descending order by sum (default).
\item Alphanumeric - rows are ordered alphanumerically by label text. By default, \code{\link[=tbl_hierarchical]{tbl_hierarchical()}} sorts tables
in ascending alphanumeric order (i.e. A to Z).
}
}
\examples{
\dontshow{if ((identical(Sys.getenv("NOT_CRAN"), "true") || identical(Sys.getenv("IN_PKGDOWN"), "true"))) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
ADAE_subset <- cards::ADAE |>
  dplyr::filter(AETERM \%in\% unique(cards::ADAE$AETERM)[1:5])

tbl <- tbl_hierarchical(
  data = ADAE_subset,
  variables = c(SEX, RACE, AETERM),
  by = TRTA,
  denominator = cards::ADSL |> mutate(TRTA = ARM),
  id = USUBJID,
  overall_row = TRUE
)

# Example 1 - Descending Frequency Sort ------------------
tbl <- tbl_sort(tbl, sort = "descending")
tbl

# Example 2 - Alphanumeric Sort --------------------------
tbl <- tbl_sort(tbl, sort = "alphanumeric")
tbl
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=tbl_filter]{tbl_filter()}}
}
