% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styfn.R
\name{styfn}
\alias{styfn}
\alias{styfn_number}
\alias{styfn_sigfig}
\alias{styfn_pvalue}
\alias{styfn_ratio}
\alias{styfn_percent}
\title{Style Functions}
\usage{
styfn_number(digits = 0, big.mark = NULL, decimal.mark = NULL, scale = 1, ...)

styfn_sigfig(digits = 2, scale = 1, big.mark = NULL, decimal.mark = NULL, ...)

styfn_pvalue(
  x,
  digits = 1,
  prepend_p = FALSE,
  big.mark = NULL,
  decimal.mark = NULL,
  ...
)

styfn_ratio(digits = 2, big.mark = NULL, decimal.mark = NULL, ...)

styfn_percent(
  symbol = FALSE,
  digits = 0,
  big.mark = NULL,
  decimal.mark = NULL,
  ...
)
}
\arguments{
\item{digits, big.mark, decimal.mark, scale, prepend_p, symbol, ...}{arguments
passed to the \verb{style_*()} functions}
}
\value{
a function
}
\description{
Similar to the \verb{style_*()} family of functions, but these functions return
a \verb{style_*()} \strong{function} rather than performing the styling.
}
\examples{
my_style <- styfn_number(digits = 1)
my_style(3.14)
}
\seealso{
Other style tools: 
\code{\link{style_number}()},
\code{\link{style_percent}()},
\code{\link{style_pvalue}()},
\code{\link{style_ratio}()},
\code{\link{style_sigfig}()}
}
\concept{style tools}
