% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_coef.R
\name{style_coef}
\alias{style_coef}
\title{Implement significant figure like formatting for regression coefficients}
\usage{
style_coef(x)
}
\arguments{
\item{x}{numeric vector}
}
\description{
Converts a numeric argument into a string that has been rounded to a
significant figure-like number suitable for regression model coefficients.
Values with magnitude less than 1, are rounded to two significant figures, and magnitudes
larger than 1 are rounded to three significant figures
}
\examples{
c(0.123, 0.9, 1.1234, 12.345, 101.234, -0.123, -0.9, -1.1234, -12.345, -101.234) \%>\%
  style_coef()
}
\seealso{
style_sigfig
}
\author{
Daniel Sjoberg
}
