% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_style_bold_labels.R
\name{tab_style_bold_labels}
\alias{tab_style_bold_labels}
\title{Bold labels in gtsummary tables}
\usage{
tab_style_bold_labels(x)
}
\arguments{
\item{x}{an object created using gtsummary functions}
}
\description{
Bold labels in gtsummary tables
}
\author{
Daniel Sjoberg
}
