% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_style_bold_levels.R
\name{tab_style_bold_levels}
\alias{tab_style_bold_levels}
\title{Bold levels in gtsummary tables}
\usage{
tab_style_bold_levels(x)
}
\arguments{
\item{x}{an object created using gtsummary functions}
}
\description{
Bold levels in gtsummary tables
}
\author{
Daniel Sjoberg
}
