% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_style_bold_p.R
\name{tab_style_bold_p}
\alias{tab_style_bold_p}
\title{Bold significant p-values}
\usage{
tab_style_bold_p(x, ...)
}
\arguments{
\item{x}{an object created using gtsummary functions}

\item{...}{further arguments passed to other methods.}
}
\description{
Bold p-values in tables created by gtsummary
}
\seealso{
\code{\link{tab_style_bold_p.tbl_summary}},
\code{\link{tab_style_bold_p.tbl_regression}},
\code{\link{tab_style_bold_p.tbl_uvregression}}
}
\author{
Daniel D. Sjoberg
}
\keyword{internal}
