% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_style_bold_p.R
\name{tab_style_bold_p.tbl_regression}
\alias{tab_style_bold_p.tbl_regression}
\title{Bold significant p-values}
\usage{
\method{tab_style_bold_p}{tbl_regression}(x, t = 0.05, ...)
}
\arguments{
\item{x}{an object created using \code{tbl_regression} function}

\item{t}{Determines the threshold below which p-values will be bold. Default is 0.05.}

\item{...}{not used}
}
\description{
Bold p-values in tables created by \code{\link{tbl_regression}}
}
\examples{
tbl_lm_bold_p <-
  lm(age ~ marker + trt + stage, trial) \%>\%
  tbl_regression() \%>\%
  tab_style_bold_p(t = 0.3)
}
\author{
Daniel Sjoberg
}
