% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_style_bold_p.R
\name{tab_style_bold_p.tbl_uregression}
\alias{tab_style_bold_p.tbl_uregression}
\title{Bold significant p-values}
\usage{
\method{tab_style_bold_p}{tbl_uregression}(x, t = 0.05, q = FALSE, ...)
}
\arguments{
\item{x}{an object created using \code{tbl_uregression} function}

\item{t}{Determines the threshold below which p-values get bolded. Default is 0.05.}

\item{q}{logical argument. When TRUE will bold the q-value column rather than the p-values}

\item{...}{not used}
}
\description{
Bold p-values in tables created by \code{\link{tbl_uregression}}
}
\author{
Daniel Sjoberg
}
