% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_style_bold_p.R
\name{tab_style_bold_p.tbl_uvregression}
\alias{tab_style_bold_p.tbl_uvregression}
\title{Bold significant p-values}
\usage{
\method{tab_style_bold_p}{tbl_uvregression}(x, t = 0.05, q = FALSE,
  ...)
}
\arguments{
\item{x}{an object created using \code{tbl_uvregression} function}

\item{t}{Determines the threshold below which p-values will be bold. Default is 0.05.}

\item{q}{logical argument. When TRUE will bold the q-value column rather than the p-values}

\item{...}{not used}
}
\description{
Bold p-values in tables created by \code{\link{tbl_uvregression}}
}
\seealso{
Other tbl_uvregression: \code{\link{add_global.tbl_uvregression}},
  \code{\link{add_nevent.tbl_uvregerssion}},
  \code{\link{add_q.tbl_uvregression}},
  \code{\link{bold_italicize_labels_levels}},
  \code{\link{inline_text.tbl_uvregression}},
  \code{\link{tbl_uvregression}}
}
\author{
Daniel D. Sjoberg
}
\concept{tbl_uvregression}
