% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_style_italicize_labels.R
\name{tab_style_italicize_labels}
\alias{tab_style_italicize_labels}
\title{Italicize labels in gtsummary tables}
\usage{
tab_style_italicize_labels(x)
}
\arguments{
\item{x}{an object created using gtsummary functions}
}
\description{
Italicize labels in gtsummary tables
}
\author{
Daniel Sjoberg
}
