% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head.R
\name{tail.gtsummary}
\alias{tail.gtsummary}
\title{Display the last rows of a gtsummary table}
\usage{
\method{tail}{gtsummary}(x, n = 6L, ...)
}
\arguments{
\item{x}{A \code{gtsummary} object.}

\item{n}{Number of rows to return. Default is \code{6L} as default in \code{tail()}.}

\item{...}{Additional arguments passed to \code{tail()}.}
}
\value{
A \code{gtsummary} object with only the last \code{n} rows in \code{table_body}.
}
\description{
Display the last rows of a gtsummary table
}
\examples{
trial |>
  tbl_summary(by = trt) |>
  tail(n = 2L)

}
