% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_filter.R
\name{tbl_filter}
\alias{tbl_filter}
\alias{tbl_filter.tbl_hierarchical}
\title{Filter Hierarchical Tables}
\usage{
tbl_filter(x, ...)

\method{tbl_filter}{tbl_hierarchical}(x, filter, ...)
}
\arguments{
\item{x}{(\code{tbl_hierarchical}, \code{tbl_hierarchical_count})\cr
A hierarchical gtsummary table of class \code{'tbl_hierarchical'} or \code{'tbl_hierarchical_count'}.}

\item{...}{These dots are for future extensions and must be empty.}

\item{t}{(scalar \code{numeric})\cr
Threshold used to determine which rows will be retained.}

\item{gt}{(scalar \code{logical})\cr
Whether to filter for row sums greater than \code{t} or less than \code{t}. Default is greater than (\code{gt = TRUE}).}

\item{eq}{(scalar \code{logical})\cr
Whether to include the value of \code{t} in the filtered range, i.e. whether to use exclusive comparators (\code{>}, \code{<}) or
inclusive comparators (\code{>=}, \code{<=}) when filtering. Default is \code{FALSE}.}

\item{.stat}{(\code{string})\cr
Statistic to use to calculate row sums. This statistic must be present in the table for all hierarchy levels.
Default is \code{"n"}.}
}
\value{
A \code{gtsummary} of the same class as \code{x}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}\cr

This function is used to filter hierarchical table rows by frequency row sum.
}
\examples{
\dontshow{if ((identical(Sys.getenv("NOT_CRAN"), "true") || identical(Sys.getenv("IN_PKGDOWN"), "true"))) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
ADAE_subset <- cards::ADAE |>
  dplyr::filter(AETERM \%in\% unique(cards::ADAE$AETERM)[1:5])

tbl <- tbl_hierarchical(
  data = ADAE_subset,
  variables = c(SEX, RACE, AETERM),
  by = TRTA,
  denominator = cards::ADSL |> mutate(TRTA = ARM),
  id = USUBJID,
  overall_row = TRUE
)

# Example 1 - Row Sums > 10 ------------------
tbl_filter(tbl, sum(n) > 10)

# Example 2 - Row Sums <= 5 ------------------
tbl_filter(tbl, sum(n) <= 5)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=tbl_sort]{tbl_sort()}}
}
