% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_hierarchical.R
\name{tbl_hierarchical_count}
\alias{tbl_hierarchical_count}
\title{Create hierarchical table}
\usage{
tbl_hierarchical_count(
  data,
  hierarchies,
  by = dplyr::group_vars(data),
  label = NULL,
  denominator = NULL,
  include = everything(),
  digits = NULL,
  summary_row = NULL
)
}
\arguments{
\item{include}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
variables from \code{hierarchy} on whose levels (prior to nesting by the variable) summary statistics should be
calculated for. Must include the last element of \code{hierarchy}.}

\item{summary_row}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
variables from \code{hierarchy} on whose levels (prior to nesting by the variable) a summary row should be displayed.}
}
\description{
Create hierarchical table
}
\examples{
data <- cards::ADAE |>
  dplyr::filter(
    AESOC \%in\% unique(cards::ADAE$AESOC)[1:5],
    AETERM \%in\% unique(cards::ADAE$AETERM)[1:5]
  )

tbl_hierarchical_count(
  data = data,
  hierarchies = c(AESOC, AETERM, AESEV),
  by = TRTA,
  denominator = cards::ADSL \%>\% mutate(TRTA = ARM)
)

}
