% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_merge.R
\name{tbl_merge}
\alias{tbl_merge}
\title{Merge tables}
\usage{
tbl_merge(
  tbls,
  tab_spanner = NULL,
  merge_vars = NULL,
  tbl_ids = NULL,
  quiet = FALSE
)
}
\arguments{
\item{tbls}{(\code{list})\cr
List of gtsummary objects to merge}

\item{tab_spanner}{(\code{character})\cr
Character vector specifying the spanning headers.
Must be the same length as \code{tbls}. The
strings are interpreted with \code{gt::md}.
Must be same length as \code{tbls} argument. Default is \code{NULL}, and places
a default spanning header. If \code{FALSE}, no header will be placed.}

\item{merge_vars}{(\code{character})\cr
Column names that are used as the merge IDs.
The default is \code{NULL}, which merges on
\verb{c(any_of(c("variable", "row_type", "var_label", "label"), cards::all_ard_groups())}.
Any column name included here that does not appear in all tables, will
be removed.}

\item{tbl_ids}{(\code{character})\cr
Optional character vector of IDs that will be assigned to the input tables.
The ID is assigned by assigning a name to the \code{tbls} list, which is
returned in \code{x$tbls}.}

\item{quiet}{(scalar \code{logical})\cr
When \code{FALSE}, a message is printed when unlike tables are merged warning
users of potential row ordering issues.}
}
\value{
A \code{'tbl_merge'} object
}
\description{
Merge gtsummary tables, e.g. \code{tbl_regression}, \code{tbl_uvregression}, \code{tbl_stack},
\code{tbl_summary}, \code{tbl_svysummary}, etc.

This function merges \strong{like tables}.
Generally, this means each of the tables being merged
should have the same structure.
When merging tables with different structures, rows may appear
out of order.
The ordering of rows can be updated with \code{modify_table_body(~dplyr::arrange(.x, ...))}.
}
\examples{
\dontshow{if ((identical(Sys.getenv("NOT_CRAN"), "true") || identical(Sys.getenv("IN_PKGDOWN"), "true")) && gtsummary:::is_pkg_installed('survival')) withAutoprint(\{ # examplesIf}
# Example 1 ----------------------------------
# Side-by-side Regression Models
library(survival)

t1 <-
  glm(response ~ trt + grade + age, trial, family = binomial) \%>\%
  tbl_regression(exponentiate = TRUE)
t2 <-
  coxph(Surv(ttdeath, death) ~ trt + grade + age, trial) \%>\%
  tbl_regression(exponentiate = TRUE)

tbl_merge(
  tbls = list(t1, t2),
  tab_spanner = c("**Tumor Response**", "**Time to Death**")
)

# Example 2 ----------------------------------
# Descriptive statistics alongside univariate regression, with no spanning header
t3 <-
  trial[c("age", "grade", "response")] \%>\%
  tbl_summary(missing = "no") \%>\%
  add_n() \%>\%
  modify_header(stat_0 ~ "**Summary Statistics**")
t4 <-
  tbl_uvregression(
    trial[c("ttdeath", "death", "age", "grade", "response")],
    method = coxph,
    y = Surv(ttdeath, death),
    exponentiate = TRUE,
    hide_n = TRUE
  )

tbl_merge(tbls = list(t3, t4)) \%>\%
  modify_spanning_header(everything() ~ NA_character_)
\dontshow{\}) # examplesIf}
}
\author{
Daniel D. Sjoberg
}
