% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_split.R
\name{tbl_split}
\alias{tbl_split}
\alias{print.tbl_split}
\title{Split gtsummary table}
\usage{
tbl_split(x, variables)

\method{print}{tbl_split}(x, ...)
}
\arguments{
\item{x}{gtsummary table}

\item{variables}{variables at which to split the gtsummary table rows (tables
will be separated after each of these variables)}

\item{...}{not used}
}
\value{
\code{tbl_split} object
}
\description{
\lifecycle{experimental}
The \code{tbl_split} function splits a single gtsummary table into multiple tables
}
\examples{
tbl <-
  tbl_summary(trial) \%>\%
  tbl_split(variables = c(marker, grade))

}
