% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_survival.R
\name{tbl_survival.survfit}
\alias{tbl_survival.survfit}
\title{Tabulate Kaplan-Meier survival probabilites for time-to-event endpoints}
\usage{
\method{tbl_survival}{survfit}(x, times, time_label = "{time}",
  level_label = "{level}, N = {n}", header = "Survival Probability",
  ...)
}
\arguments{
\item{x}{\code{survfit} object. While \code{survfit()} allows stratifcation by multiple
variables, including more than one variable will result in display problems.}

\item{times}{numeric vector of survival times}

\item{time_label}{string defining the label shown for the time column.
Default is "{time}".  The input uses \code{\link[glue]{glue}} notation to
convert the string into a label.  A common label may be "{time} Months", which
would resolve to "6 Months" or "12 Months" depending on the times requested.}

\item{level_label}{used when the input 'x' results are stratified by a variable.
It is a string defining the label shown.  The input uses
\code{\link[glue]{glue}} notation to convert the string into a label.
The default is "{level}, N = {n}", which will print the level of the stratifying
variable.  Other information available to call are {n}', '{variable}',
and '{strata}'.}

\item{header}{string to be displayed as column header of the Kaplan-Meier
estimate.  Default is 'Survival Probability'}

\item{...}{not used}
}
\description{
Function accepts a \code{survfit} object, and provides a tablulated and formmated
summary of the results
}
\examples{
library(survival)
survfit(Surv(ttdeath, death) ~ trt, trial) \%>\%
  tbl_survival(
    times = 1:2,
    time_label = "{time} Years"
  )
}
\author{
Daniel D. Sjoberg
}
