% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-tbl_regression.R
\name{tidy_vetted}
\alias{tidy_vetted}
\title{Vetted tidy models}
\description{
Below is a list of models vetted for use
in \link{tbl_regression} and \link{tbl_uvregression}.  If a model is passed to these
functions and the model is not listed below and  a \code{tidy()} function is
not specified in the \verb{tidy_fun=} argument, the model object will be passed
to \link[broom:tidy]{broom::tidy}.

\itemize{
\item{\link[stats:lm]{stats::lm}}
\item{\link[stats:glm]{stats::glm}}
\item{\link[survival:coxph]{survival::coxph}}
\item{\link[survival:survreg]{survival::survreg}}
\item{\link[lme4:glmer]{lme4::glmer}}
\item{\link[lme4:lmer]{lme4::lmer}}
\item{\link[geepack:geeglm]{geepack::geeglm}}
}
}
\section{model support}{

If \link[broom:tidy]{broom::tidy} or \link[broom.mixed:tidy]{broom.mixed::tidy} support a class of model not listed
above, please submit a \href{https://github.com/ddsjoberg/gtsummary/issues}{GitHub Issue}.
The model can be added to the list of vetted models and unit tests will be
put in place to ensure continued support for the model.
}

