% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyselect_to_list.R
\name{tidyselect_to_list}
\alias{tidyselect_to_list}
\title{Convert tidyselect to variable list}
\usage{
tidyselect_to_list(.data, x, .meta_data = NULL, input_type = NULL)
}
\arguments{
\item{.data}{data with variables to select from}

\item{x}{list of tidyselect formulas}

\item{.meta_data}{meta data from tbl_summary. Default is NULL}

\item{input_type}{indicates type of example to print in deprecation note}
}
\description{
Functions takes a list of tidyselect formulas, e.g. \code{list(starts_with("age") ~ "continuous")},
and returns a named list, e.g. \code{list(age = "continuous")}.
}
\keyword{internal}
