% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_header_stats.R
\name{update_header_stats}
\alias{update_header_stats}
\title{Update Header Statistics}
\usage{
update_header_stats(x, card)
}
\arguments{
\item{x}{(\code{gtsummary})\cr
Object with class 'gtsummary'}

\item{card}{(\code{card})\cr
Object of class 'card', typically created with \code{cards::ard_tabulate()}.
The ARD must be a univariate tabulation.}
}
\value{
gtsummary object
}
\description{
Use this function to update the statistics that can be injected into
gtsummary table headers.
}
\examples{
# Example 1 ----------------------------------
tbl_summary(trial, by = trt, include = age) |>
  update_header_stats(
    # doubling the trial data set and the Ns
    card =
      rbind(trial, trial) |>
      cards::ard_tabulate(variables = trt)
  ) |>
  modify_header(all_stat_cols() ~ "**{level}**  \n N = {n}")
}
