% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_examples.R
\name{utils_examples}
\alias{utils_examples}
\alias{write_html_output}
\alias{read_html_output}
\alias{write_and_read_html_output}
\title{Utilities to write and read tables in html for function documentation}
\usage{
write_html_output(
  tbl,
  example_name = "example",
  out_var_name = deparse(substitute(tbl)),
  method = c("txt", "rds")[2]
)

read_html_output(
  out_var_name,
  example_name = "example",
  method = c("txt", "rds")[2]
)

write_and_read_html_output(
  tbl,
  example_name = "example",
  out_var_name = deparse(substitute(tbl)),
  method = c("txt", "rds")[2]
)
}
\arguments{
\item{tbl}{}
}
\description{
This is still experimental. We are planning to potentially expand this to other
html output. Principally based on \code{\link[=as_gt]{as_gt()}} and \code{\link[gt:as_raw_html]{gt::as_raw_html()}}.
}
\keyword{internal}
