% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_helpers.R
\name{vec_to_df}
\alias{vec_to_df}
\title{Convert character vector to data frame}
\usage{
vec_to_df(x)
}
\arguments{
\item{x}{character vector}
}
\value{
data frame
}
\description{
This is used in some of the selecting we allow for, for example in
\code{as_gt(include=)} you can use tidyselect to select among the call
names to include.
}
\keyword{internal}
