% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ci.R
\name{add_ci}
\alias{add_ci}
\alias{add_ci.tbl_summary}
\title{Add CI Column}
\usage{
add_ci(x, ...)

\method{add_ci}{tbl_summary}(
  x,
  method = list(all_continuous() ~ "t.test", all_categorical() ~ "wilson"),
  include = everything(),
  statistic = list(all_continuous() ~ "{conf.low}, {conf.high}", all_categorical() ~
    "{conf.low}\%, {conf.high}\%"),
  conf.level = 0.95,
  style_fun = list(all_continuous() ~ styfn_sigfig(), all_categorical() ~
    styfn_sigfig(scale = 100)),
  pattern = NULL,
  ...
)
}
\arguments{
\item{x}{(\code{tbl_summary})\cr
a summary table of class \code{'tblsummary'}}

\item{...}{These dots are for future extensions and must be empty.}

\item{method}{(\code{\link[=syntax]{formula-list-selector}})\cr
Confidence interval method. Default is
\code{list(all_continuous() ~ "t.test", all_categorical() ~ "wilson")}.
See details below.}

\item{include}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
Variables to include in the summary table. Default is \code{everything()}}

\item{statistic}{(\code{\link[=syntax]{formula-list-selector}})\cr
Indicates how the confidence interval will be displayed.
Default is \code{list(all_continuous() ~ "{conf.low}, {conf.high}", all_categorical() ~ "{conf.low}\%, {conf.high}\%")}}

\item{style_fun}{(\code{function})\cr
Function to style upper and lower bound of confidence interval. Default is
\code{list(all_continuous() ~ styfn_sigfig(), all_categorical() ~ styfn_sigfig(scale =  100))}.}

\item{pattern}{(\code{string})\cr
Indicates the pattern to use to merge the CI with
the statistics cell. The default is NULL, where no columns are merged.
The two columns that will be merged are the statistics column,
represented by \code{"{stat}"} and the CI column represented by \code{"{ci}"},
e.g. \code{pattern = "{stat} ({ci})"} will merge the two columns with the CI
in parentheses. Default is \code{NULL}, and no merging is performed.}

\item{conf.level(scalar}{\code{real})\cr
Confidence level. Default is \code{0.95}}
}
\value{
gtsummary table
}
\description{
Add a new column with the confidence intervals for proportions, means, etc.
}
\section{method argument}{


Must be one of
\code{c("wilson", "wilson.no.correct", "exact", "asymptotic")} for categorical
variables, and \code{c("t.test", "wilcox.test")} for continuous variables.
\subsection{Continuous Variables}{

Confidence intervals for means are calculated using \code{t.test()} and
\code{wilcox.test()} for pseudo-medians.
}

\subsection{Categorical/Dichotomous Variables}{

Methods \code{c("wilson", "wilson.no.correct")} are calculated with
\code{prop.test(correct = c(TRUE, FALSE))}.
The default method, \code{"wilson"}, includes the Yates continuity correction.
Methods \code{c("exact", "asymptotic")} are calculated with \code{Hmisc::binconf(method=)}.
}
}

\examples{
# Example 1 ----------------------------------
add_ci_ex1 <-
  trial \%>\%
  select(marker, response, trt) \%>\%
  tbl_summary(
    missing = "no",
    statistic = all_continuous() ~ "{mean} ({sd})"
  ) \%>\%
  add_ci()

# Example 2 ----------------------------------
add_ci_ex2 <-
  trial \%>\%
  select(response, grade) \%>\%
  tbl_summary(
    statistic = all_categorical() ~ "{p}\%",
    missing = "no"
  ) \%>\%
  add_ci(pattern = "{stat} ({ci})") \%>\%
  modify_footnote(everything() ~ NA)
}
