% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_difference.R
\name{add_difference}
\alias{add_difference}
\title{Add differences}
\usage{
add_difference(x, ...)
}
\arguments{
\item{x}{(\code{gtsummary})\cr
Object with class 'gtsummary'}

\item{...}{Passed to other methods.}
}
\description{
\itemize{
\item \code{\link[=add_difference.tbl_summary]{add_difference.tbl_summary()}}
\item \code{\link[=add_difference.tbl_svysummary]{add_difference.tbl_svysummary()}}
\item \code{\link[=add_difference.tbl_hierarchical]{add_difference.tbl_hierarchical()}}
}
}
\author{
Daniel D. Sjoberg
}
\keyword{internal}
