% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_difference.tbl_svysummary.R
\name{add_difference.tbl_svysummary}
\alias{add_difference.tbl_svysummary}
\title{Add differences between groups}
\usage{
\method{add_difference}{tbl_svysummary}(
  x,
  test = NULL,
  group = NULL,
  adj.vars = NULL,
  test.args = NULL,
  conf.level = 0.95,
  levels = NULL,
  include = everything(),
  pvalue_fun = label_style_pvalue(digits = 1),
  estimate_fun = list(c(all_continuous(), all_categorical(FALSE)) ~ label_style_sigfig(),
    all_dichotomous() ~ label_style_sigfig(scale = 100, suffix = "\%"), all_tests("smd")
    ~ label_style_sigfig()),
  ...
)
}
\arguments{
\item{x}{(\code{tbl_summary})\cr
table created with \code{tbl_summary()}}

\item{test}{(\code{\link[=syntax]{formula-list-selector}})\cr
Specifies the tests/methods to perform for each variable, e.g.
\code{list(all_continuous() ~ "svy.t.test", all_dichotomous() ~ "emmeans", all_categorical(FALSE) ~ "svy.chisq.test")}.

See below for details on default tests and \link[=tests]{?tests} for details on available
tests and creating custom tests.}

\item{group}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
Variable name of an ID or grouping variable. The column can be used to
calculate p-values with correlated data.
Default is \code{NULL}. See \link{tests} for methods that utilize the \code{group} argument.}

\item{adj.vars}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
Variables to include in adjusted calculations (e.g. in ANCOVA models).
Default is \code{NULL}.}

\item{test.args}{(\code{\link[=syntax]{formula-list-selector}})\cr
Containing additional arguments to pass to tests that accept arguments.
For example, add an argument for all t-tests, use
\code{test.args = all_tests("t.test") ~ list(var.equal = TRUE)}.}

\item{conf.level}{(\code{numeric})\cr
a scalar in the interval \verb{(0, 1)} indicating the confidence level. Default is 0.95}

\item{levels}{(\code{vector})\cr
a length-two vector of the \code{tbl_svysummary(by=)} levels to compare. The
difference is calculated as \code{levels[1]} minus \code{levels[2]}.
This argument is required when the \code{by} variable has more than two levels, and
allows the user to select which two groups to compare. When \code{by} has exactly
two levels, this argument is optional and can be used to flip the direction of
the difference (e.g. \code{levels[2]} minus \code{levels[1]}). Default is \code{NULL}.}

\item{include}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
Variables to include in output. Default is \code{everything()}.}

\item{pvalue_fun}{(\code{function})\cr
Function to round and format p-values. Default is \code{label_style_pvalue()}.
The function must have a numeric vector input, and return a string that is
the rounded/formatted p-value (e.g. \code{pvalue_fun = label_style_pvalue(digits = 2)}).}

\item{estimate_fun}{(\code{\link[=syntax]{formula-list-selector}})\cr
List of formulas specifying the functions
to round and format differences and confidence limits.
Default is
\verb{list(c(all_continuous(), all_categorical(FALSE)) ~ label_style_sigfig(), all_categorical() ~ \\(x) paste0(style_sigfig(x, scale = 100), "\%"))}}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
a gtsummary table of class \code{"tbl_summary"}
}
\description{
Adds difference to tables created by \code{\link[=tbl_summary]{tbl_summary()}}.
The difference between two groups (typically mean or rate difference) is added
to the table along with the difference's confidence interval and a p-value (when applicable).
}
\examples{
\dontshow{if ((identical(Sys.getenv("NOT_CRAN"), "true") || identical(Sys.getenv("IN_PKGDOWN"), "true")) && gtsummary:::is_pkg_installed("broom", ref = "cardx")) withAutoprint(\{ # examplesIf}
# Example 1 ----------------------------------
survey::svydesign(~1, data = as.data.frame(Titanic), weights = ~Freq) |>
  tbl_svysummary(
    by = Survived,
    value = list(Class ~ "1st", Age = "Child"),
    include = c(Class, Age)
  ) |>
  add_difference()

# Example 2 ----------------------------------
# Select two groups to compare when `by=` has 3+ levels
survey::svydesign(~1, data = trial, weights = ~1) |>
  tbl_svysummary(by = grade, include = c(age, marker), missing = "no") |>
  add_difference(levels = c("I", "III"))
\dontshow{\}) # examplesIf}
}
