% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_global_p.R
\name{add_global_p}
\alias{add_global_p}
\title{Adds the global p-value for a categorical variables}
\usage{
add_global_p(x, ...)
}
\arguments{
\item{x}{\code{tbl_regression} or \code{tbl_uvregression} object}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
This function uses \link[car:Anova]{car::Anova} with argument
\code{type = "III"} to calculate global p-values for categorical variables.
Output from \code{tbl_regression} and \code{tbl_uvregression} objects supported.
}
\section{Note}{

If a needed class of model is not supported by
\link[car:Anova]{car::Anova}, please create a
\href{https://github.com/ddsjoberg/gtsummary/issues}{GitHub Issue} to request support.
}

\seealso{
\code{\link{add_global_p.tbl_regression}},
\code{\link{add_global_p.tbl_uvregression}}
}
\author{
Daniel D. Sjoberg
}
