% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_n.R
\name{add_n}
\alias{add_n}
\title{Adds a column with N (or N missing) for each variable}
\usage{
add_n(x, missing = FALSE, last = FALSE)
}
\arguments{
\item{x}{object with class \code{tbl_summary} from the \code{\link{tbl_summary}} function}

\item{missing}{logical argument indicating whether to print N (\code{missing = FALSE}),
or N missing (\code{missing = TRUE}).  Default is \code{FALSE}}

\item{last}{logical indicator to include overall  column last. Default is \code{FALSE}}
}
\description{
The function assumes the DEFAULT header are in use. Only modify header rows
after all columns has been added.
}
\examples{
trial \%>\% tbl_summary(by = "trt") \%>\% add_n()
}
