% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_n.R
\name{add_n}
\alias{add_n}
\title{Adds column with N to gtsummary table}
\usage{
add_n(x, ...)
}
\arguments{
\item{x}{Object created from a gtsummary function}

\item{...}{Additional arguments passed to other methods.}
}
\description{
Adds column with N to gtsummary table
}
\seealso{
\code{\link[=add_n.tbl_summary]{add_n.tbl_summary()}}, \code{\link[=add_n.tbl_svysummary]{add_n.tbl_svysummary()}}, \code{\link[=add_n.tbl_survfit]{add_n.tbl_survfit()}},
\link{add_n.tbl_regression}, \link{add_n.tbl_uvregression}
}
\author{
Daniel D. Sjoberg
}
\keyword{internal}
