% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_nevent.R
\name{add_nevent}
\alias{add_nevent}
\title{Add p-values}
\usage{
add_nevent(x, ...)
}
\arguments{
\item{x}{(\code{gtsummary})\cr
Object with class 'gtsummary'}

\item{...}{Passed to other methods.}
}
\description{
\itemize{
\item \code{\link[=add_nevent.tbl_regression]{add_nevent.tbl_regression()}}
}
}
\seealso{
\code{\link[=add_p.tbl_summary]{add_p.tbl_summary()}}
}
\author{
Daniel D. Sjoberg
}
\keyword{internal}
