% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_nevent.R
\name{add_nevent_regression}
\alias{add_nevent_regression}
\alias{add_nevent.tbl_regression}
\title{Add event N}
\usage{
\method{add_nevent}{tbl_regression}(x, location = "label", ...)
}
\arguments{
\item{x}{(\code{tbl_regression}/\code{tbl_uvregression})\cr
a \code{tbl_regression} or \code{tbl_uvregression} table}

\item{location}{(\code{character})\cr
location to place Ns. Select one or more of \code{c('label', 'level')}.
Default is \code{'label'}.

When \code{"label"} total Ns are placed
on each variable's label row. When \code{"level"} level counts are placed on the
variable level for categorical variables, and total N on the variable's label
row for continuous.}

\item{...}{These dots are for future extensions and must be empty.}
}
\description{
Add event N
}
\examples{
# TODO: Re-add after tbl_uvregression() is migrated
# Example 1 ----------------------------------
# add_nevent.tbl_regression_ex1 <-
#   trial \%>\%
#   select(response, trt, grade) \%>\%
#   tbl_uvregression(
#     y = response,
#     method = glm,
#     method.args = list(family = binomial),
#   ) \%>\%
#   add_nevent()

# Example 2 ----------------------------------
glm(response ~ age + grade, trial, family = binomial) |>
  tbl_regression(exponentiate = TRUE) |>
  add_nevent(location = "level")
}
