% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_overall.R
\name{add_overall}
\alias{add_overall}
\title{Adds a column with overall summary statistics to an existing \code{tbl_summary} object where descriptive
statistics are split by a variable}
\usage{
add_overall(x, last = FALSE)
}
\arguments{
\item{x}{object with class \code{tbl_summary} from the \code{\link{tbl_summary}} function}

\item{last}{logical indicator to include overall  column last. Default is \code{FALSE}}
}
\description{
The function assumes the DEFAULT headers are in use. Only modify header rows
after Overall column has been added.
}
\examples{
tbl_overall <- trial \%>\% tbl_summary(by = "trt") \%>\% add_overall()
}
\seealso{
Other tbl_summary: \code{\link{add_comparison}},
  \code{\link{add_n}}, \code{\link{add_q.tbl_summary}},
  \code{\link{add_stat_label}},
  \code{\link{bold_italicize_labels_levels}},
  \code{\link{cols_label_summary}},
  \code{\link{inline_text.tbl_summary}},
  \code{\link{tab_style_bold_p.tbl_summary}},
  \code{\link{tbl_summary}}
}
\author{
Daniel D. Sjoberg
}
\concept{tbl_summary}
