% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_overall.R
\name{add_overall}
\alias{add_overall}
\alias{add_overall.tbl_custom_summary}
\title{Add overall column}
\usage{
add_overall(x, ...)

\method{add_overall}{tbl_custom_summary}(
  x,
  last = FALSE,
  col_label = "**Overall**  \\nN = {N}",
  statistic = NULL,
  digits = NULL,
  ...
)
}
\arguments{
\item{x}{(\code{gtsummary})\cr
Object with class 'gtsummary'}

\item{...}{Passed to other methods.}
}
\description{
\itemize{
\item \code{\link[=add_overall.tbl_summary]{add_overall.tbl_summary()}}
}
}
\seealso{
\code{\link[=add_overall.tbl_summary]{add_overall.tbl_summary()}}
}
\author{
Daniel D. Sjoberg
}
\keyword{internal}
