% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_p.R
\name{add_p}
\alias{add_p}
\title{Add p-values}
\usage{
add_p(x, ...)
}
\arguments{
\item{x}{(\code{gtsummary})\cr
Object with class 'gtsummary'}

\item{...}{Passed to other methods.}
}
\description{
\itemize{
\item \code{\link[=add_p.tbl_summary]{add_p.tbl_summary()}}
\item \code{\link[=add_p.tbl_svysummary]{add_p.tbl_svysummary()}}
\item \code{\link[=add_p.tbl_continuous]{add_p.tbl_continuous()}}
\item \code{\link[=add_p.tbl_cross]{add_p.tbl_cross()}}
\item \code{\link[=add_p.tbl_survfit]{add_p.tbl_survfit()}}
}
}
\author{
Daniel D. Sjoberg
}
\keyword{internal}
