% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_p.R
\name{add_p}
\alias{add_p}
\title{Adds p-values to gtsummary table}
\usage{
add_p(x, ...)
}
\arguments{
\item{x}{Object created from a gtsummary function}

\item{...}{Additional arguments passed to other methods.}
}
\description{
Adds p-values to gtsummary table
}
\seealso{
\link{add_p.tbl_summary}, \link{add_p.tbl_cross}
}
\author{
Daniel D. Sjoberg
}
