% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_p.R
\name{add_p.tbl_continuous}
\alias{add_p.tbl_continuous}
\title{P-values for \code{tbl_continuous}}
\usage{
\method{add_p}{tbl_continuous}(
  x,
  test = NULL,
  pvalue_fun = NULL,
  include = everything(),
  test.args = NULL,
  group = NULL,
  ...
)
}
\description{
P-values for \code{tbl_continuous}
}
