% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_p.R
\name{add_p.tbl_summary}
\alias{add_p.tbl_summary}
\title{Add p-values to summary table}
\usage{
\method{add_p}{tbl_summary}(
  x,
  test = NULL,
  pvalue_fun = styfn_pvalue(),
  group = NULL,
  include = everything(),
  test.args = NULL,
  ...
)
}
\arguments{
\item{x}{(\code{tbl_summary})\cr
table created with \code{tbl_summary()}}

\item{test}{(\code{\link[=syntax]{formula-list-selector}})\cr
Specifies the statistical tests to perform for each variable, e.g. \code{ list(all_continuous() ~ "t.test", all_categorical() ~ "fisher.test")}.
Common tests include \code{"t.test"}, \code{"aov"}, \code{"wilcox.test"}, \code{"kruskal.test"},
\code{"chisq.test"}, \code{"fisher.test"}, and \code{"lme4"} (for clustered data).
See \link{tests} for details, more tests, and instruction for implementing a custom test.

Default tests when \code{group} argument not specified are \code{"kruskal.test"}
for continuous variables (\code{"wilcox.test"} when "by" variable has two levels),
\code{"chisq.test.no.correct"} for categorical variables with all expected cell
counts >=5, and \code{"fisher.test"} for categorical variables with any expected cell count <5.}

\item{pvalue_fun}{(\code{function})\cr
Function to round and format p-values. Default is \code{styfn_pvalue()}.
The function must have a numeric vector input, and return a string that is
the rounded/formatted p-value (e.g. \code{pvalue_fun = styfn_pvalue(digits = 2)}).}

\item{group}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
Variable name of an ID or grouping variable. The column can be used to
calculate p-values with correlated data.
Default is \code{NULL}. See \link{tests} for methods that utilize the \code{group} argument.}

\item{include}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
Variables to include in output. Default is \code{everything()}.}

\item{test.args}{(\code{\link[=syntax]{formula-list-selector}})\cr
Containing additional arguments to pass to tests that accept arguments.
For example, add an argument for all t-tests, use
\code{test.args = all_tests("t.test") ~ list(var.equal = TRUE)}.}

\item{...}{Not used}
}
\value{
a gtsummary table of class \code{"tbl_summary"}
}
\description{
Adds p-values to tables created by \code{\link[=tbl_summary]{tbl_summary()}} by comparing values across groups.
}
\examples{
\dontshow{if (gtsummary:::is_pkg_installed("cardx", reference_pkg = "gtsummary") && gtsummary:::is_pkg_installed("broom", reference_pkg = "cardx")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Example 1 ----------------------------------
add_p_ex1 <-
  trial |>
  tbl_summary(by = trt, include = c(age, grade)) |>
  add_p()

# Example 2 ----------------------------------
add_p_ex2 <-
  trial \%>\%
  select(trt, age, marker) |>
  tbl_summary(by = trt, missing = "no") |>
  add_p(
    # perform t-test for all variables
    test = everything() ~ "t.test",
    # assume equal variance in the t-test
    test.args = all_tests("t.test") ~ list(var.equal = TRUE)
  )
\dontshow{\}) # examplesIf}
}
