% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_p.R
\name{add_p.tbl_survfit}
\alias{add_p.tbl_survfit}
\title{Adds p-value to survfit table}
\usage{
\method{add_p}{tbl_survfit}(x, include = everything(), footnote_text = NULL, quiet = FALSE, ...)
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}
Calculate and add a p-value
}
