% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_q.R
\name{add_q}
\alias{add_q}
\title{Add a column of q values to objects to account for
multiple comparisons}
\usage{
add_q(x, ...)
}
\arguments{
\item{x}{\code{fmt_table1} or \code{fmt_uni_regression} object}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Add a column of q values to objects to account for
multiple comparisons
}
