% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_q.R
\name{add_q}
\alias{add_q}
\title{Add a column of q values to account for
multiple comparisons}
\usage{
add_q(x, ...)
}
\arguments{
\item{x}{\code{tbl_summary} or \code{tbl_uvregression} object}

\item{...}{Additional arguments passed to other methods.}
}
\description{
Add a column of q values to account for
multiple comparisons
}
\seealso{
\code{\link{add_q.tbl_summary}}, \code{\link{add_q.tbl_uvregression}},
\code{\link{tbl_summary}}, \code{\link{tbl_uvregression}}
}
\author{
Esther Drill, Daniel D. Sjoberg
}
