% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_stat_label.R
\name{add_stat_label}
\alias{add_stat_label}
\title{Adds a column showing a label for the summary statistics shown in each row}
\usage{
add_stat_label(x)
}
\arguments{
\item{x}{object with class \code{tbl_summary} from the \code{\link{tbl_summary}} function}
}
\description{
Rather than simply printing the summary statistics, with the use of \code{add_stat_label()},
a column labeling the summary statistics is added.
}
\examples{
tbl_stat <- mtcars \%>\% tbl_summary() \%>\% add_stat_label()
}
