% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_gt.R
\name{as_gt}
\alias{as_gt}
\title{Convert gtsummary object to a gt object}
\usage{
as_gt(x, include = everything(), return_calls = FALSE, ...)
}
\arguments{
\item{x}{An object of class `"gtsummary"}

\item{include}{Commands to include in output. Input may be a vector of
quoted or unquoted names. tidyselect and gtsummary select helper
functions are also accepted.
Default is \code{everything()}.}

\item{return_calls}{Logical. Default is \code{FALSE}. If \code{TRUE}, the calls are returned
as a list of expressions.}

\item{...}{Arguments passed on to \code{gt::gt(...)}}
}
\value{
A \code{gt_tbl} object
}
\description{
Function converts a gtsummary object to a \code{"gt_tbl"} object,
that is, a table created with \code{gt::gt()}.
Function is used in the background when the results are printed or knit.
A user can use this function if they wish to add customized formatting
available via the \href{https://gt.rstudio.com/index.html}{gt package}.
}
\examples{
# # Example 1 ----------------------------------
# as_gt_ex1 <-
#   trial[c("trt", "age", "response", "grade")] \%>\%
#   tbl_summary(by = trt) \%>\%
#   as_gt()
}
\author{
Daniel D. Sjoberg
}
\concept{gtsummary output types}
