% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_kable.R
\name{as_kable}
\alias{as_kable}
\title{Convert to knitr_kable object}
\usage{
as_kable(x, omit = NULL)
}
\arguments{
\item{x}{object created by a function from the gtsummary package
(e.g. \code{\link{tbl_summary}} or \code{\link{tbl_regression}})}

\item{omit}{vector of named kable commands to omit. Default is \code{NULL}}
}
\description{
Function converts object to a knitr_kable object.  This function is used in the
background when the results are printed or knit.  A user can use this function
if they wish to add customized formatting available via \link[knitr:kable]{knitr::kable}
}
\examples{
trial \%>\%
tbl_summary(by = "trt") \%>\%
as_kable()
}
\seealso{
\link{tbl_summary} \link{tbl_regression} \link{tbl_uvregression}
}
\author{
Daniel D. Sjoberg
}
