% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_kable_extra.R
\name{as_kable_extra}
\alias{as_kable_extra}
\title{Convert gtsummary object to a kableExtra object}
\usage{
as_kable_extra(
  x,
  include = everything(),
  return_calls = FALSE,
  strip_md_bold = TRUE,
  linebreak = FALSE,
  ...
)
}
\arguments{
\item{x}{Object created by a function from the gtsummary package
(e.g. \link{tbl_summary} or \link{tbl_regression})}

\item{include}{Commands to include in output. Input may be a vector of
quoted or unquoted names. tidyselect and gtsummary select helper
functions are also accepted.
Default is \code{everything()}, which includes all commands in \code{x$kable_calls}.}

\item{return_calls}{Logical. Default is \code{FALSE}. If \code{TRUE}, the calls are returned
as a list of expressions.}

\item{strip_md_bold}{When TRUE, all double asterisk (markdown language for
bold weight) in column labels and spanning headers are removed.
Default is TRUE}

\item{linebreak}{When \code{TRUE} character columns and column labels will
be interpretted with \link[kableExtra:linebreak]{kableExtra::linebreak} allowing for
line breaks to be inserted with \code{"\\n"} with PDF output. Default is \code{FALSE}}

\item{...}{Additional arguments passed to \link[knitr:kable]{knitr::kable}}
}
\value{
A {kableExtra} object
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}
Function converts a gtsummary object to a knitr_kable + kableExtra object.
A user can use this function if they wish to add customized formatting
available via \link[knitr:kable]{knitr::kable} and {kableExtra}.
}
\examples{
# in PDF output, we can make the table with width of the output using
# kableExtra::kable_styling(). This is helpful for scaling wide tables to fit
trial \%>\%
  tbl_summary(by = trt) \%>\%
  modify_header(stat_by = "{level}\nN = {n}")
  as_kable_extra() \%>\%
  kableExtra::kable_styling(latex_options = "scale_down")
}
\seealso{
Other gtsummary output types: 
\code{\link{as_flextable}()},
\code{\link{as_gt}()},
\code{\link{as_kable}()},
\code{\link{as_tibble.gtsummary}()}
}
\author{
Daniel D. Sjoberg
}
\concept{gtsummary output types}
