% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_tests.R
\name{assign_tests}
\alias{assign_tests}
\alias{assign_tests.tbl_summary}
\title{Assign Test}
\usage{
assign_tests(x, ...)

\method{assign_tests}{tbl_summary}(
  x,
  test = NULL,
  group = NULL,
  include,
  calling_fun = c("add_p", "add_difference"),
  ...
)
}
\arguments{
\item{x}{(\code{gtsummary})\cr
a table of class \code{'gtsummary'}}

\item{...}{Passed to \code{\link[rlang:abort]{rlang::abort()}}, \code{\link[rlang:abort]{rlang::warn()}} or
\code{\link[rlang:abort]{rlang::inform()}}.}

\item{test}{(named \code{list})\cr
a named list of tests.}

\item{group}{(\code{string})\cr
a variable name indicating the grouping column for correlated data.
Default is \code{NULL}.}

\item{include}{(\code{character})\cr
Character vector of column names to assign a default tests.}

\item{calling_fun}{(\code{string})\cr
Must be one of \code{'add_p'} and \code{'add_difference'}. Depending on the context,
different defaults are set.}
}
\value{
A table of class \code{'gtsummary'}
}
\description{
This function is used to assign default tests for \code{add_p()}
and \code{add_difference()}.
}
\examples{
trial |>
  tbl_summary(
    by = trt,
    include  = c(age, stage)
  ) |>
  assign_tests(include = c("age", "stage"), calling_fun = "add_p")
}
