% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bold_italicise_labels_levels.R
\name{bold_italicize_labels_levels}
\alias{bold_italicize_labels_levels}
\alias{tab_style_bold_labels}
\alias{tab_style_bold_levels}
\alias{tab_style_italicize_labels}
\alias{tab_style_italicize_levels}
\title{Bold or Italicize labels or levels in gtsummary tables}
\usage{
tab_style_bold_labels(x)

tab_style_bold_levels(x)

tab_style_italicize_labels(x)

tab_style_italicize_levels(x)
}
\arguments{
\item{x}{an object created using gtsummary functions}
}
\description{
Bold or Italicize labels or levels in gtsummary tables
}
\section{Functions}{
\itemize{
\item \code{tab_style_bold_labels}: Bold labels in gtsummary tables

\item \code{tab_style_bold_levels}: Bold levels in gtsummary tables

\item \code{tab_style_italicize_labels}: Italicize labels in gtsummary tables

\item \code{tab_style_italicize_levels}: Italicize levels in gtsummary tables
}}

\author{
Daniel Sjoberg
}
