% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bold_italicise_labels_levels.R
\name{bold_italicize_labels_levels}
\alias{bold_italicize_labels_levels}
\alias{bold_labels}
\alias{italicize_labels}
\alias{bold_levels}
\alias{italicize_levels}
\alias{bold_labels.gtsummary}
\alias{bold_levels.gtsummary}
\alias{italicize_labels.gtsummary}
\alias{italicize_levels.gtsummary}
\alias{bold_labels.tbl_cross}
\alias{bold_levels.tbl_cross}
\alias{italicize_labels.tbl_cross}
\alias{italicize_levels.tbl_cross}
\title{Bold or Italicize labels or levels in gtsummary tables}
\usage{
bold_labels(x)

italicize_labels(x)

bold_levels(x)

italicize_levels(x)

\method{bold_labels}{gtsummary}(x)

\method{bold_levels}{gtsummary}(x)

\method{italicize_labels}{gtsummary}(x)

\method{italicize_levels}{gtsummary}(x)

\method{bold_labels}{tbl_cross}(x)

\method{bold_levels}{tbl_cross}(x)

\method{italicize_labels}{tbl_cross}(x)

\method{italicize_levels}{tbl_cross}(x)
}
\arguments{
\item{x}{Object created using gtsummary functions}
}
\value{
Functions return the same class of gtsummary object supplied
}
\description{
Bold or Italicize labels or levels in gtsummary tables
}
\section{Example Output}{


\if{html}{Example 1}

\if{html}{\out{
<img src="https://raw.githubusercontent.com/ddsjoberg/gtsummary/main/man-images/tbl_bold_ital_ex1.png" alt="image of rendered example table" style="width:50\%;">
}}
}

\examples{
# Example 1 ----------------------------------
tbl_bold_ital_ex1 <-
  trial[c("trt", "age", "grade")] \%>\%
  tbl_summary() \%>\%
  bold_labels() \%>\%
  bold_levels() \%>\%
  italicize_labels() \%>\%
  italicize_levels()
}
\seealso{
Other tbl_summary tools: 
\code{\link{add_n.tbl_summary}()},
\code{\link{add_overall}()},
\code{\link{add_p.tbl_summary}()},
\code{\link{add_q}()},
\code{\link{add_stat_label}()},
\code{\link{inline_text.tbl_summary}()},
\code{\link{inline_text.tbl_survfit}()},
\code{\link{modify}},
\code{\link{separate_p_footnotes}()},
\code{\link{tbl_custom_summary}()},
\code{\link{tbl_merge}()},
\code{\link{tbl_split}()},
\code{\link{tbl_stack}()},
\code{\link{tbl_strata}()},
\code{\link{tbl_summary}()}

Other tbl_regression tools: 
\code{\link{add_global_p}()},
\code{\link{add_q}()},
\code{\link{combine_terms}()},
\code{\link{inline_text.tbl_regression}()},
\code{\link{modify}},
\code{\link{tbl_merge}()},
\code{\link{tbl_regression}()},
\code{\link{tbl_split}()},
\code{\link{tbl_stack}()},
\code{\link{tbl_strata}()}

Other tbl_uvregression tools: 
\code{\link{add_global_p}()},
\code{\link{add_q}()},
\code{\link{inline_text.tbl_uvregression}()},
\code{\link{modify}},
\code{\link{tbl_merge}()},
\code{\link{tbl_split}()},
\code{\link{tbl_stack}()},
\code{\link{tbl_strata}()},
\code{\link{tbl_uvregression}()}
}
\author{
Daniel D. Sjoberg
}
\concept{tbl_regression tools}
\concept{tbl_summary tools}
\concept{tbl_uvregression tools}
