% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bold_p.R
\name{bold_p}
\alias{bold_p}
\title{Bold significant p-values in Rmarkdown}
\usage{
bold_p(x, ...)
}
\arguments{
\item{x}{\code{fmt_table1}, \code{fmt_regression}, or \code{fmt_uni_regression} object}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Bold p-values in \code{fmt_table1}, \code{fmt_regression}, and \code{fmt_uni_regression}
objects by adding '__' to each side of the string
}
\seealso{
\code{\link{bold_p.fmt_table1}}, \code{\link{bold_p.fmt_regression}}
}
