% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bold_p.R
\name{bold_p}
\alias{bold_p}
\title{Bold significant p-values or q-values}
\usage{
bold_p(x, ...)
}
\arguments{
\item{x}{Object created using gtsummary functions}

\item{...}{Additional arguments passed to other methods.}
}
\description{
Bold values below a chosen threshold (e.g. <0.05)
in gtsummary tables.
}
\seealso{
\code{\link{bold_p.tbl_summary}},
\code{\link{bold_p.tbl_regression}},
\code{\link{bold_p.tbl_uvregression}}
}
\author{
Daniel D. Sjoberg, Esther Drill
}
