% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brdg_hierarchical.R
\name{brdg_hierarchical}
\alias{brdg_hierarchical}
\title{Hierarchical Summary Table Bridges}
\usage{
brdg_hierarchical(
  cards,
  hierarchies,
  type,
  by,
  id,
  include,
  statistic,
  labels_hierarchy,
  missing = "no",
  missing_stat = "{N_miss}",
  missing_text = "Unknown"
)
}
\description{
Hierarchical Summary Table Bridges
}
