% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brdg_hierarchical.R
\name{brdg_hierarchical}
\alias{brdg_hierarchical}
\alias{pier_summary_hierarchical}
\title{Hierarchy table bridge}
\usage{
brdg_hierarchical(
  cards,
  hierarchies,
  by,
  include,
  statistic,
  type,
  overall_row,
  count,
  label
)

pier_summary_hierarchical(cards, variables, include, statistic)
}
\arguments{
\item{cards}{(\code{card})\cr
an ARD object of class \code{"card"} created with \code{cards::ard_hierarchical_stack()}.}

\item{hierarchies}{(\code{character})\cr
character list of hierarchy variables.}

\item{by}{(\code{string})\cr
string indicating the stratifying column.}

\item{include}{(\code{character})\cr
character list of hierarchy variables to include summary statistics for.}

\item{statistic}{(named \code{list})\cr
named list of summary statistic names.}

\item{type}{(named \code{list})\cr
named list of summary types.}

\item{overall_row}{(scalar \code{logical})\cr
whether an overall summmary row should be included at the top of the table.
The default is \code{FALSE}.}

\item{count}{(scalar \code{logical})\cr
whether \code{tbl_hierarchical_count()} (\code{TRUE}) or \code{tbl_hierarchical()} (\code{FALSE}) is being applied.}

\item{label}{(named \code{list})\cr
named list of hierarchy variable labels.}
}
\value{
a gtsummary object
}
\description{
Bridge function for converting \code{tbl_hierarchical()} (and similar) cards to basic gtsummary objects.
}
\seealso{
Review \link[=syntax]{list, formula, and selector syntax} used throughout gtsummary
}
