% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brdg_hierarchical.R
\name{brdg_hierarchical}
\alias{brdg_hierarchical}
\title{Hierarchical Summary Table Bridges}
\usage{
brdg_hierarchical(
  cards,
  hierarchies,
  type,
  by,
  include,
  statistic,
  cards_summary,
  labels_hierarchy
)
}
\description{
Hierarchical Summary Table Bridges
}
