% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge_summary.R
\name{bridge_summary}
\alias{bridge_summary}
\alias{pier_summary_dichotomous}
\alias{pier_summary_categorical}
\alias{pier_summary_continuous2}
\alias{pier_summary_continuous}
\alias{pier_summary_missing_row}
\title{Summary Table Bridges}
\usage{
pier_summary_dichotomous(x, variables, value = x$inputs$value)

pier_summary_categorical(x, variables, missing, missing_text, missing_stat)

pier_summary_continuous2(x, variables, missing, missing_text, missing_stat)

pier_summary_continuous(x, variables, missing, missing_text, missing_stat)

pier_summary_missing_row(x, variables = x$inputs$include)
}
\arguments{
\item{x}{gtsummary object of class \code{"tbl_summary"}}

\item{variables}{character list of variables}

\item{value}{named list of values to be summarized. the names are the
variable names.}

\item{missing}{TODO:}

\item{missing_text}{TODO:}

\item{missing_stat}{TODO:}
}
\value{
data frame
}
\description{
Bridge function for converting \code{tbl_summary()} (and similar) cards to table bodies.
All bridge functions begin with prefix \verb{brdg_*()}.

This file also contains helper functions for constructing the bridge,
referred to as the piers (supports for a bridge) and begin with \verb{pier_*()}.
}
\examples{
tbl <-
  tbl_summary(
    data = mtcars,
    type =
      list(
        cyl = "categorical",
        am = "dichotomous",
        mpg = "continuous",
        hp = "continuous2"
      ),
    value = list(am = 1),
    statistic =
      list(
        c(cyl, am) ~ "{n} ({p}\%)",
        mpg = "{mean} ({sd})",
        hp = c("{mean}", "{median}")
      )
  )

pier_summary_dichotomous(
 x = tbl,
 variables = "am",
 value = list(am = 1),
 missing = "ifany",
 missing_text = "Unknown",
 missing_stat = "{N_miss}"
)

pier_summary_categorical(
 x = tbl,
 variables = "cyl",
 missing = "ifany",
 missing_text = "Unknown",
 missing_stat = "{N_miss}"
)

pier_summary_continuous2(
 x = tbl,
 variables = "hp",
 missing = "ifany",
 missing_text = "Unknown",
 missing_stat = "{N_miss}"
)

pier_summary_continuous(
 x = tbl,
 variables = "mpg",
 missing = "ifany",
 missing_text = "Unknown",
 missing_stat = "{N_miss}"
)
}
