% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-calculate_pvalue.R
\name{calculate_pvalue}
\alias{calculate_pvalue}
\title{This function takes in the meta-data and calculates an appropriate p-value}
\usage{
calculate_pvalue(data, variable, by, test, type, id)
}
\arguments{
\item{data}{input data set}

\item{variable}{categorical or continuous variable for which a test with \code{by_var} is desired}

\item{by}{categorical variable}

\item{test}{list of statistical tests from meta data}

\item{type}{the type of variable, one of categorical or continuous, from the metadata}

\item{id}{the id variable for clustered data}
}
\value{
a table of p-values for each variable
}
\description{
This function takes in the meta-data and calculates an appropriate p-value
}
\keyword{internal}
