% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_stat_labels.R
\name{default_stat_labels}
\alias{default_stat_labels}
\title{Default Statistics Labels}
\usage{
default_stat_labels()
}
\value{
named list
}
\description{
Default Statistics Labels
}
\keyword{internal}
