% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{add_comparison}
\alias{add_global}
\title{Deprecated functions}
\usage{
add_comparison(...)

add_global(...)
}
\description{
Deprecated functions
}
\keyword{internal}
