% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{add_comparison}
\alias{add_global}
\title{Deprecated functions}
\usage{
add_comparison(...)

add_global(...)
}
\description{
Some functions have been deprecated and are no longer being actively
supported. Please use these functions instead:
\link{add_p}, \link{add_global_p}.
}
\keyword{internal}
