% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{add_comparison}
\alias{add_global}
\alias{tab_style_bold_p}
\alias{tab_style_bold_labels}
\alias{tab_style_italicize_levels}
\alias{tab_style_italicize_labels}
\alias{tab_style_bold_levels}
\alias{tbl_summary_}
\alias{add_p_}
\alias{as_flextable}
\alias{all_numeric}
\alias{all_character}
\alias{all_integer}
\alias{all_double}
\alias{all_logical}
\alias{all_factor}
\title{Deprecated functions}
\usage{
add_comparison(...)

add_global(...)

tab_style_bold_p(...)

tab_style_bold_labels(...)

tab_style_italicize_levels(...)

tab_style_italicize_labels(...)

tab_style_bold_levels(...)

tbl_summary_(...)

add_p_(...)

as_flextable(...)

all_numeric()

all_character()

all_integer()

all_double()

all_logical()

all_factor()
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("deprecated")}
Some functions have been deprecated and are no longer being actively
supported.
}
\keyword{internal}
