% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{modify_cols_merge}
\title{Deprecated functions}
\usage{
modify_cols_merge(...)
}
\description{
\lifecycle{deprecated}\cr
Some functions have been deprecated and are no longer being actively
supported.
}
\keyword{internal}
