% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_ci_column.R
\name{deprecated_ci_column}
\alias{deprecated_ci_column}
\title{\code{"ci"} Column Deprecated}
\description{
\lifecycle{deprecated}\cr
\subsection{Overview}{

When the gtsummary package was first written, the gt package was not on CRAN
and the version of the package that was available did not have the ability
to merge columns.
Due to these limitations, the \code{"ci"} column was added to show the combined
\code{"conf.low"} and \code{"conf.high"} columns.

Column merging in both gt and gtsummary packages has matured over the years,
and we are now adopting a more modern approach by using these features.
As a result, the \code{"ci"} column will eventually be dropped from \code{.$table_body}.

By using column merging, the \code{conf.low} and \code{conf.high} remain numeric
and we can to continue to update how these columns are formatted.
}

\subsection{How to update?}{

In most cases it is a simple change to adapt your code to the updated
structure: simply swap \code{ci} with \code{conf.low}.
For example, if you had updated the column header with \code{modify_header(ci = "Confidence Interval")},
you would update your code to

\if{html}{\out{<div class="sourceCode r">}}\preformatted{modify_header(conf.low = "Confidence Interval")
}\if{html}{\out{</div>}}

If you need to update a reference from the \code{pattern} argument in \code{modify_column_merge()} or
\code{inline_text()}, you can replace the \code{"{ci}"} reference with \code{"{conf.low}, {conf.high}"}.
}
}
\keyword{internal}
