% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glance_fun.R
\name{glance_fun_s3}
\alias{glance_fun_s3}
\alias{glance_fun_s3.default}
\alias{glance_fun_s3.mira}
\title{Default glance function}
\usage{
glance_fun_s3(x, ...)

\method{glance_fun_s3}{default}(x, ...)

\method{glance_fun_s3}{mira}(x, ...)
}
\arguments{
\item{x}{(regression model)\cr
a regression model object}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
a function
}
\description{
This is an S3 generic used as the default function in \code{add_glance*(glance_fun)}.
It's provided so various regression model classes can have their own default
functions for returning statistics.
}
\examples{
\dontshow{if ((identical(Sys.getenv("NOT_CRAN"), "true") || identical(Sys.getenv("IN_PKGDOWN"), "true")) && gtsummary:::is_pkg_installed("cardx")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
mod <- lm(age ~ trt, trial)

glance_fun_s3(mod)
\dontshow{\}) # examplesIf}
}
\keyword{internal}
