% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtsummary-package.R
\docType{package}
\name{gtsummary-package}
\alias{gtsummary}
\alias{gtsummary-package}
\title{gtsummary: Presentation-Ready Data Summary and Analytic Result
    Tables}
\description{
The {gtsummary} package creates presentation-ready
    tables summarizing data sets, regression models, and more. The code to
    create the tables is concise, and also highly customizable. Data
    frames can be summarized any function that summarizes data, e.g.
    mean(), sd(), median(), or any user-written function. Regression
    models are summarized including the reference row for categorical
    independent variables.  Common regression models, such as, logistic
    regression and Cox proportional hazards regression, are automatically
    identified and the tables are pre-filled with the appropriate column
    headers.
}
\author{
\strong{Maintainer}: Daniel D. Sjoberg \email{danieldsjoberg@gmail.com}

Authors:
\itemize{
  \item Emily C. Zabor
  \item Margie Hannum
  \item Karissa Whiting
}

}
\keyword{internal}
