% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtsummary-package.R
\docType{package}
\name{gtsummary-package}
\alias{gtsummary}
\alias{gtsummary-package}
\title{gtsummary: The goal of gtsummary is to make reporting of tabular
    analytic results easy, beautiful, and reproducible}
\description{
Written with the goal to make reporting of tabular
    analytic results simple, beautiful, and reproducible.
    Publication-ready tables are created with minimal code, and can be
    exported to Word, HTML, and PDF documents with the use of Rmarkdown.
    The clintable functions are flexible, allowing you to easily customize
    the aesthetics of tables to suit your needs.
}
\author{
\strong{Maintainer}: Daniel D. Sjoberg \email{danieldsjoberg@gmail.com}

Authors:
\itemize{
  \item Emily C. Zabor
  \item Margie Hannum
}

}
\keyword{internal}
