% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtsummary-package.R
\docType{package}
\name{gtsummary-package}
\alias{gtsummary}
\alias{gtsummary-package}
\title{gtsummary: Presentation-Ready Data Summary and Analytic Result
    Tables}
\description{
Creates presentation-ready tables summarizing data
    sets, regression models, and more. The code to create the tables is
    concise and highly customizable. Data frames can be summarized with
    any function, e.g. mean(), median(), even user-written functions.
    Regression models are summarized and include the reference rows for
    categorical variables. Common regression models, such as logistic
    regression and Cox proportional hazards regression, are automatically
    identified and the tables are pre-filled with appropriate column
    headers.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://www.danieldsjoberg.com/gtsummary/}
  \item Report bugs at \url{https://github.com/ddsjoberg/gtsummary/issues}
}

}
\author{
\strong{Maintainer}: Daniel D. Sjoberg \email{danield.sjoberg@gmail.com}

Authors:
\itemize{
  \item Emily C. Zabor
  \item Margie Hannum
  \item Karissa Whiting
}

}
\keyword{internal}
