% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head.R
\name{head.gtsummary}
\alias{head.gtsummary}
\title{Display the first rows of a gtsummary table}
\usage{
\method{head}{gtsummary}(x, n = 6L, ...)
}
\arguments{
\item{x}{A \code{gtsummary} object.}

\item{n}{Number of rows to return. Default is \code{6L} as default in \code{head()}.}

\item{...}{Additional arguments passed to \code{head()}.}
}
\value{
A \code{gtsummary} object with only the first \code{n} rows in \code{table_body}.
}
\description{
Display the first rows of a gtsummary table
}
\examples{
trial |>
  tbl_summary(by = trt) |>
  tail(n = 2L)

}
