% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_text.R
\name{inline_text.gtsummary}
\alias{inline_text.gtsummary}
\title{Report statistics from summary tables inline}
\usage{
\method{inline_text}{gtsummary}(x, variable, level = NULL, column = NULL, pattern = NULL)
}
\arguments{
\item{x}{gtsummary object}

\item{variable}{Variable name of statistic to present}

\item{level}{Level of the variable to display for categorical variables.
Default is \code{NULL}}

\item{column}{Column name to return from \code{x$table_body}.}

\item{pattern}{String indicating the statistics to return.
Uses \link[glue:glue]{glue::glue} formatting. Default is \code{NULL}}
}
\description{
Report statistics from summary tables inline
}
