% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_text.R
\name{inline_text}
\alias{inline_text}
\title{Report statistics from \code{fmt_table1} and \code{fmt_regression} inline in an Rmarkdown document}
\usage{
inline_text(x, ...)
}
\arguments{
\item{x}{\code{fmt_table1} or \code{fmt_regression} object}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Report statistics from \code{fmt_table1} and \code{fmt_regression} inline in an Rmarkdown document
}
\seealso{
\code{\link{inline_text.fmt_table1}}, \code{\link{inline_text.fmt_regression}}
}
