% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_text.R
\name{inline_text}
\alias{inline_text}
\title{Report statistics from gtsummary tables inline}
\usage{
inline_text(x, ...)
}
\arguments{
\item{x}{object created from a gtsummary function}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Report statistics from gtsummary tables inline
}
\seealso{
\code{\link{inline_text.tbl_summary}},
\code{\link{inline_text.tbl_regression}},
<<<<<<< HEAD
\code{\link{inline_text.tbl_uvregression}},
\code{\link{inline_text.tbl_survival}}
=======
\link{inline_text.tbl_uvregression}
>>>>>>> origin/master
}
\author{
Daniel D. Sjoberg
}
