% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_text.R
\name{inline_text}
\alias{inline_text}
\title{Report statistics from gtsummary tables inline}
\usage{
inline_text(x, ...)
}
\arguments{
\item{x}{object created from a gtsummary funciton}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Report statistics from gtsummary tables inline
}
\seealso{
\link{inline_text.tbl_summary}, \link{tbl_summary}, \link{tbl_regression}, \link{tbl_uvregression}
}
\author{
Daniel Sjoberg
}
