% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_text.R
\name{inline_text}
\alias{inline_text}
\title{Report statistics from gtsummary tables inline}
\usage{
inline_text(x, ...)
}
\arguments{
\item{x}{Object created from a gtsummary function}

\item{...}{Additional arguments passed to other methods.}
}
\value{
A string reporting results from a gtsummary table
}
\description{
Report statistics from gtsummary tables inline
}
\author{
Daniel D. Sjoberg
}
\keyword{internal}
