% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_text.R
\name{inline_text}
\alias{inline_text}
\title{Report statistics from gtsummary tables inline}
\usage{
inline_text(x, ...)
}
\arguments{
\item{x}{(\code{gtsummary})\cr
Object created from a gtsummary function}

\item{...}{Additional arguments passed to other methods.}
}
\value{
A string reporting results from a gtsummary table
}
\description{
Report statistics from gtsummary tables inline
}
\seealso{
\code{\link[=inline_text.tbl_summary]{inline_text.tbl_summary()}}, \code{\link[=inline_text.tbl_svysummary]{inline_text.tbl_svysummary()}},
\code{\link[=inline_text.tbl_regression]{inline_text.tbl_regression()}}, \code{\link[=inline_text.tbl_uvregression]{inline_text.tbl_uvregression()}},
\code{\link[=inline_text.tbl_survfit]{inline_text.tbl_survfit()}}, \code{\link[=inline_text.tbl_cross]{inline_text.tbl_cross()}}, \code{\link[=inline_text.gtsummary]{inline_text.gtsummary()}}
}
\author{
Daniel D. Sjoberg
}
\keyword{internal}
