% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_text.R
\name{inline_text.tbl_regression}
\alias{inline_text.tbl_regression}
\title{Report statistics from regression summary tables inline}
\usage{
\method{inline_text}{tbl_regression}(x, variable, level = NULL,
  pattern = "{coef} ({conf.level*100}\% CI {ll}, {ul}; {p.value})",
  coef_fun = x$inputs$coef_fun, pvalue_fun = function(x)
  style_pvalue(x, prepend_p = TRUE), ...)
}
\arguments{
\item{x}{object created from  \link{tbl_regression}}

\item{variable}{variable name of statistic to present}

\item{level}{level of the variable to display for categorical variables.
Default is \code{NULL}, returning the top row in the table for the variable.}

\item{pattern}{statistics to return.  Uses \link[glue:glue]{glue::glue} formatting.
Default is \code{"{coef} ({conf.level }\% CI  {ll}, {ul}; {p.value})"}.  All columns from
\code{x$table_body} are available to print as well as the confidence level (conf.level).
Uses \link[glue:glue]{glue::glue} formatting. See below for details.}

\item{coef_fun}{function to style model coefficients.
Columns 'coef', 'll', and 'ul' are formatted. Default is \code{x$inputs$coef_fun}}

\item{pvalue_fun}{function to style p-values and/or q-values.
Default is \code{function(x) style_pvalue(x, prepend_p = TRUE)}}

\item{...}{not used}
}
\description{
Functions takes an object with class \code{tbl_regression}, and the
location of the statistic to report and returns the statistic for reporting
inline in an R markdown document.  Detailed examples in the
\href{https://github.com/pages/ddsjoberg/gtsummary/articles/tbl_regression.html}{tbl_regression vignette}
}
\section{pattern argument}{

The following items are available to print.  Use \code{print(x$table_body)} to
print the table the estimates are extracted from.
\itemize{
\item \code{{coef}} coeficient estiamte formatted with 'coef_fun'
\item \code{{ll}} lower limit of confidence interval formmated with 'coef_fun'
\item \code{{ul}} upper limit of confidence interval formmated with 'coef_fun'
\item \code{{ci}} confidence interval formmated with x$estimate_fun
\item \code{{p.value}} p-value formatted with 'pvalue_fun'
\item \code{{N}} number of observations in model
\item \code{{label}} variable/variable level label
}
}

\examples{
inline_text_ex1 <-
  glm(response ~ age + grade, trial, family = binomial(link = "logit")) \%>\%
  tbl_regression(exponentiate = TRUE)

inline_text(inline_text_ex1, variable = "age")
inline_text(inline_text_ex1, variable = "grade", level = "III")
}
\seealso{
Other tbl_regression tools: \code{\link{add_global.tbl_regression}},
  \code{\link{add_nevent.tbl_regression}},
  \code{\link{bold_italicize_labels_levels}},
  \code{\link{tbl_merge}}, \code{\link{tbl_regression}}
}
\author{
Daniel D. Sjoberg
}
\concept{tbl_regression tools}
