% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_text.R
\name{inline_text.tbl_regression}
\alias{inline_text.tbl_regression}
\title{Report statistics from regression summary tables inline}
\usage{
\method{inline_text}{tbl_regression}(x, variable, level = NULL,
  pattern = "{coef} ({conf.level*100}\% CI {ll}, {ul}; {pvalue})",
  coef_fun = x$inputs$coef_fun, pvalue_fun = function(x)
  style_pvalue(x, prepend_p = TRUE), ...)
}
\arguments{
\item{x}{object created from  \link{tbl_regression}}

\item{variable}{variable name of statistic to present}

\item{level}{level of the variable to display for categorical variables.
Default is \code{NULL}, returning the top row in the table for the variable.}

\item{pattern}{statistics to return.  Uses \link[glue]{glue} formatting.
Default is \code{"{coef} ({conf.level }\% CI  {ll}, {ul}; {pvalue})"}.  All columns from
\code{.$table_body} are available to print as well as the confidence level (conf.level)}

\item{coef_fun}{function to style model coefficients.
Columns 'coef', 'll', and 'ul' are formatted. Default is \code{x$inputs$coef_fun}}

\item{pvalue_fun}{function to style p-values and/or q-values.
Default is \code{function(x) style_pvalue(x, prepend_p = TRUE)}}

\item{...}{not used}
}
\description{
Functions takes an object with class \code{tbl_regression}, and the
location of the statistic to report and returns the statistic for reporting
inline in an R markdown document.  Detailed examples in the
\href{https://github.com/pages/ddsjoberg/gtsummary/articles/tbl_regression.html}{\code{tbl_regression} vignette}
}
\seealso{
\link{tbl_regression}
}
\author{
Daniel Sjoberg
}
