% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_text.R
\name{inline_text.tbl_summary}
\alias{inline_text.tbl_summary}
\title{Report statistics from summary tables inline}
\usage{
\method{inline_text}{tbl_summary}(x, variable, level = NULL,
  column = ifelse(is.null(x$by), "stat_0", stop("Must specify column")),
  pvalue_fun = purrr::partial(style_pvalue, prepend_p = TRUE), ...)
}
\arguments{
\item{x}{object created from  \link{tbl_summary}}

\item{variable}{variable name of statistic to present}

\item{level}{level of the variable to display for categorical variables.
Can also specify the 'Unknown' row.  Default is \code{NULL}}

\item{column}{name column to return.  Default is \code{stat_0}, when
no by variable is passed to \code{tbl_summary()}.  See below for details.}

\item{pvalue_fun}{function for rounding/formatting p-values.
Default is \code{\link{style_pvalue}}.
The function must have a single input (the numeric, exact p-value),
and return a string that is the rounded/formatted p-value (e.g.
\code{pvalue_fun = function(x) style_pvalue(x, digits = 2)} or equivalently,
\code{purrr::partial(style_pvalue, digits = 2)}).}

\item{...}{not used}
}
\description{
Functions takes an object with class \code{tbl_summary}, and the
location of the statistic to report and returns the statistic for reporting
inline in an R markdown document
}
\section{Selecting Columns}{

Using the \code{column} argument, select the column to return.  The argument
accepts any column name in the \code{.$table_body} data frame.  Overall statistics
are saved in \code{stat_0}.  When a \code{by} argument is passed to \code{tbl_summary()}, the
first by level (when ordered) is \code{stat_1}, the second is \code{stat_2}, and so on.
A helper function \code{by_level} is available to aid in selecting the correct
column.  If a \code{by} variable has levels 'Drug' and 'Placebo',
then \code{by_level('Drug')} would return \code{stat_1}
}

\author{
Daniel Sjoberg
}
