% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_text.R
\name{inline_text.tbl_summary}
\alias{inline_text.tbl_summary}
\title{Report statistics from summary tables inline}
\usage{
\method{inline_text}{tbl_summary}(x, variable, level = NULL,
  column = ifelse(is.null(x$by), "stat_0", stop("Must specify column")),
  pvalue_fun = function(x) style_pvalue(x, prepend_p = TRUE), ...)
}
\arguments{
\item{x}{object created from  \link{tbl_summary}}

\item{variable}{variable name of statistic to present}

\item{level}{level of the variable to display for categorical variables.
Can also specify the 'Unknown' row.  Default is \code{NULL}}

\item{column}{name column name to return from \code{x$table_body}.
Can also pass the level of a by variable.}

\item{pvalue_fun}{function for rounding/formatting p-values.
Default is \code{\link{style_pvalue}}.
The function must have a single input (the numeric, exact p-value),
and return a string that is the rounded/formatted p-value (e.g.
\code{pvalue_fun = function(x) style_pvalue(x, digits = 2)} or equivalently,
\code{partial(style_pvalue, digits = 2)}).}

\item{...}{not used}
}
\description{
Functions takes an object with class \code{tbl_summary}, and the
location of the statistic to report and returns the statistic for reporting
inline in an R markdown document.  Detailed examples in the
\href{https://github.com/pages/ddsjoberg/gtsummary/articles/tbl_summary.html}{\code{tbl_summary} vignette}
}
\seealso{
Other tbl_summary: \code{\link{add_comparison}},
  \code{\link{add_n}}, \code{\link{add_overall}},
  \code{\link{add_q.tbl_summary}},
  \code{\link{add_stat_label}},
  \code{\link{bold_italicize_labels_levels}},
  \code{\link{cols_label_summary}},
  \code{\link{tab_style_bold_p.tbl_summary}},
  \code{\link{tbl_summary}}
}
\author{
Daniel Sjoberg
}
\concept{tbl_summary}
