% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_text.R
\name{inline_text.tbl_survival}
\alias{inline_text.tbl_survival}
\title{Report statistics from survival summary tables inline}
\usage{
\method{inline_text}{tbl_survival}(x, strata = NULL, time = NULL,
  prob = NULL,
  pattern = "{estimate} ({conf.level*100}\% CI {lower}, {upper})", ...)
}
\arguments{
\item{x}{object created from  \link{tbl_survival}}

\item{strata}{if \code{tbl_survival} estimates are stratified, level of the stratum
report. Default is \code{NULL} when \code{tbl_survival} have no specified strata.}

\item{time}{time for which to return survival probability.}

\item{prob}{probability for which to return survival time.}

\item{pattern}{statistics to return.  Uses \link[glue]{glue} formatting.
Default is \code{'{surv} ({conf.level.100}\% {lower}, {upper})'}.  All columns from
\code{.$table_body} are available to print as well as the confidence level (conf.level)}

\item{...}{not used}
}
\description{
Functions takes an object with class \code{tbl_survival}, and the
location of the statistic to report and return the statistic for reporting
inline in an R markdown document.
}
\examples{
library(survival)
survfit(Surv(ttdeath, death) ~ trt, trial) \%>\%
  tbl_survival(
    times = c(12, 24)
  ) \%>\%
  inline_text(
    strata = "Drug",
    time = 12
  )
}
\author{
Karissa Whiting
}
\concept{tbl_survival}
