% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_text.R
\name{inline_text.tbl_survival}
\alias{inline_text.tbl_survival}
\title{Report statistics from survival summary tables inline}
\usage{
\method{inline_text}{tbl_survival}(x, ...)
}
\arguments{
\item{x}{object created from  \link{tbl_survival}}

\item{...}{not used}
}
\description{
Functions takes an object with class \code{tbl_survival}, and the
location of the statistic to report and returns the statistic for reporting
inline in an R markdown document.
}
\seealso{
Other tbl_survival: \code{\link{tbl_survival.survfit}},
  \code{\link{tbl_survival}}
}
\concept{tbl_survival}
