% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{is_date_time}
\alias{is_date_time}
\title{Is a date/time}
\usage{
is_date_time(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a scalar logical
}
\description{
\code{is_date_time()}: Predicate for date, time, or date-time vector identification.
}
\examples{
iris |>
  dplyr::mutate(date = as.Date("2000-01-01") + dplyr::row_number()) |>
  lapply(gtsummary:::is_date_time)
}
\keyword{internal}
