% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_abbreviation.R
\name{modify_abbreviation}
\alias{modify_abbreviation}
\title{Modify Abbreviations}
\usage{
modify_abbreviation(x, abbreviation)
}
\arguments{
\item{x}{(\code{gtsummary})\cr
A gtsummary object}

\item{abbreviation}{(\code{string})\cr
a string}
}
\value{
Updated gtsummary object
}
\description{
All abbreviations will be coalesced when printing the final table into
a single source note.
}
\examples{
# TODO: Add examples
}
