% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_column_hide.R
\name{modify_column_hide}
\alias{modify_column_hide}
\alias{modify_column_unhide}
\title{Modify Hidden Columns}
\usage{
modify_column_hide(x, column)

modify_column_unhide(x, column)
}
\arguments{
\item{x}{gtsummary object}

\item{column}{vector or selector of columns in \code{x$table_body}}
}
\description{
Use these functions to hide or unhide columns in a gtsummary tables.
}
\examples{
modify_column_hide_ex1 <-
  lm(age ~ marker + grade, trial) \%>\%
  tbl_regression() \%>\%
  modify_column_hide(column = ci) \%>\%
  modify_column_unhide(column = std.error)
}
