% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_footnote.R
\name{modify_footnote2}
\alias{modify_footnote2}
\alias{modify_footnote_header}
\alias{modify_footnote_body}
\alias{remove_footnote_header}
\alias{remove_footnote_body}
\title{Modify Footnotes}
\usage{
modify_footnote_header(x, footnote, columns, text_interpret = c("md", "html"))

modify_footnote_body(
  x,
  footnote,
  columns,
  rows,
  text_interpret = c("md", "html")
)

remove_footnote_header(x, columns)

remove_footnote_body(x, columns, rows)
}
\arguments{
\item{x}{(\code{gtsummary})\cr
A gtsummary object}

\item{footnote}{(\code{string})\cr
a string}

\item{columns}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
columns to add footnote}

\item{text_interpret}{(\code{string})\cr
String indicates whether text will be interpreted with
\code{\link[gt:md]{gt::md()}} or \code{\link[gt:html]{gt::html()}}. Must be \code{"md"} (default) or \code{"html"}.
Applies to tables printed with \code{{gt}}.}

\item{rows}{(predicate \code{expression})\cr
Predicate expression to select rows in \code{x$table_body}.
Review \link[=rows_argument]{rows argument details}.}
}
\value{
Updated gtsummary object
}
\description{
Modify Footnotes
}
\examples{
# TODO: Add examples
}
