% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_table_header.R
\name{modify_table_header}
\alias{modify_table_header}
\title{Modify table_header}
\usage{
modify_table_header(
  x,
  column,
  label = NULL,
  hide = NULL,
  align = NULL,
  missing_emdash = NULL,
  indent = NULL,
  text_interpret = NULL,
  bold = NULL,
  italic = NULL,
  fmt_fun = NULL,
  footnote_abbrev = NULL,
  footnote = NULL,
  spanning_header = NULL
)
}
\arguments{
\item{x}{gtsummary object}

\item{column}{columns to update}

\item{label}{string of column label}

\item{hide}{logical indicating whether to hide column from output}

\item{align}{string indicating alignment of column, must be one of
\code{c("left", "right", "center")}}

\item{missing_emdash}{string that evaluates to logical identifying rows to
include em-dash for missing values, e.g. \code{"reference_row == TRUE"}}

\item{indent}{string that evaluates to logical identifying rows to indent}

\item{text_interpret}{string, must be one of \code{"gt::md"} or \code{"gt::html"}}

\item{bold}{string that evaluates to logical identifying rows to bold}

\item{italic}{string that evaluates to logical identifying rows to italicize}

\item{fmt_fun}{function that formats the statistics in the column}

\item{footnote_abbrev}{string with abbreviation definition, e.g.
\code{"CI = Confidence Interval"}}

\item{footnote}{string with text for column footnote}

\item{spanning_header}{string with text for spanning header}
}
\value{
gtsummary object
}
\description{
This is a function meant for advanced users to gain
more control over the characteristics of the resulting
gtsummary table.
}
\details{
Review the
\href{http://www.danieldsjoberg.com/gtsummary/articles/gtsummary_definition.html}{gtsummary definition}
vignette for information on \code{.$table_header} objects.
}
\examples{
lm(mpg ~ factor(cyl), mtcars) \%>\%
  tbl_regression() \%>\%
  modify_table_header(column = estimate,
                      label = "estimate test",
                      fmt_fun = function(x) style_sigfig(x, digits = 5),
                      footnote_abbrev = "test = unit test",
                      footnote = "this is a footnote") \%>\%
  modify_table_header(column = "p.value",
                      hide = TRUE)
}
