% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_table_body_styling.R
\name{modify_table_styling}
\alias{modify_table_styling}
\title{Modify table_body_styling}
\usage{
modify_table_styling(
  x,
  columns,
  rows = NA_character_,
  label = NULL,
  spanning_header = NULL,
  hide = NULL,
  align = NULL,
  footnote = NULL,
  footnote_abbrev = NULL,
  missing_symbol = NULL,
  fmt_fun = NULL,
  indent = FALSE,
  bold = FALSE,
  italic = FALSE,
  undo_text_format = FALSE,
  text_interpret = c("gt::md", "gt::html")
)
}
\arguments{
\item{x}{gtsummary object}

\item{columns}{vector or selector of columns in \code{x$table_body}}

\item{rows}{predicate string expression to select rows. \code{rows = NA} often indicates all rows.}

\item{label}{string of column label(s)}

\item{spanning_header}{string with text for spanning header}

\item{hide}{logical indicating whether to hide column from output}

\item{align}{string indicating alignment of column, must be one of
\code{c("left", "right", "center")}}

\item{footnote}{string with text for footnote}

\item{footnote_abbrev}{string with abbreviation definition, e.g.
\code{"CI = Confidence Interval"}}

\item{missing_symbol}{string indicating how missing values are formatted.}

\item{fmt_fun}{function that formats the statistics in the
columns/rows in \verb{columns=} and \verb{rows=}}

\item{indent}{logical indicating if the rows selected in \verb{rows=} should be indented}

\item{bold}{logical indicating if the rows selected in \verb{rows=} should be bold}

\item{italic}{logical indicating if the rows selected in \verb{rows=} should be italic}

\item{undo_text_format}{rarely used. Logical that undoes the indent, bold,
and italic styling when \code{TRUE}}

\item{text_interpret}{string, must be one of \code{"gt::md"} or \code{"gt::html"}}
}
\description{
This is a function meant for advanced users to gain
more control over the characteristics of the resulting
gtsummary table.
}
\details{
Review the
\href{http://www.danieldsjoberg.com/gtsummary/articles/gtsummary_definition.html}{gtsummary definition}
vignette for information on \code{.$table_header} objects.
}
\seealso{
\code{modify_table_body()}

See \href{http://www.danieldsjoberg.com/gtsummary/articles/gtsummary_definition.html}{gtsummary internals vignette}
}
