% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\alias{plot.tbl_regression}
\alias{plot.tbl_uvregression}
\title{Plot Regression Coefficients}
\usage{
\method{plot}{tbl_regression}(x, remove_header_rows = TRUE, remove_reference_rows = FALSE, ...)

\method{plot}{tbl_uvregression}(x, remove_header_rows = TRUE, remove_reference_rows = FALSE, ...)
}
\arguments{
\item{x}{(\code{tbl_regression}, \code{tbl_uvregression})\cr
A 'tbl_regression' or 'tbl_uvregression' object}

\item{remove_header_rows}{(scalar \code{logical})\cr
logical indicating whether to remove header rows
for categorical variables. Default is \code{TRUE}}

\item{remove_reference_rows}{(scalar \code{logical})\cr
logical indicating whether to remove reference rows
for categorical variables. Default is \code{FALSE}.}

\item{...}{arguments passed to \code{ggstats::ggcoef_plot(...)}}
}
\value{
a ggplot
}
\description{
The \code{plot()} function extracts \code{x$table_body} and passes the it to
\code{ggstats::ggcoef_plot()} along with formatting options.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}\cr
}
\examples{
\dontshow{if (gtsummary:::is_pkg_installed("ggstats")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
glm(response ~ marker + grade, trial, family = binomial) |>
  tbl_regression(
    add_estimate_to_reference_rows = TRUE,
    exponentiate = TRUE
  ) |>
  plot()
\dontshow{\}) # examplesIf}
}
