% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print_gtsummary}
\alias{print_gtsummary}
\alias{print.gtsummary}
\title{print and knit_print methods for gtsummary objects}
\usage{
\method{print}{gtsummary}(x, print_engine = "gt", ...)
}
\arguments{
\item{x}{An object created using gtsummary functions}

\item{print_engine}{String indicating the print method. Must be one of
\code{"gt"}, \code{"kable"}, \code{"kable_extra"}, \code{"flextable"}, \code{"tibble"}}

\item{...}{Not used}
}
\description{
print and knit_print methods for gtsummary objects
}
\author{
Daniel D. Sjoberg
}
\keyword{internal}
