% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print_gtsummary}
\alias{print_gtsummary}
\alias{print.gtsummary}
\alias{knit_print.gtsummary}
\title{print and knit_print methods for gtsummary objects}
\usage{
\method{print}{gtsummary}(x, ...)

\method{knit_print}{gtsummary}(x, ...)
}
\arguments{
\item{x}{An object created using gtsummary functions}

\item{...}{Not used}
}
\description{
print and knit_print methods for gtsummary objects
}
\seealso{
\link{tbl_summary} \link{tbl_regression} \link{tbl_uvregression} \link{tbl_merge} \link{tbl_stack}
}
\author{
Daniel D. Sjoberg
}
