% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as_tibble}
\alias{knit_print}
\alias{vars}
\alias{select}
\alias{as_huxtable}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{select}}, \code{\link[dplyr]{vars}}}

  \item{huxtable}{\code{\link[huxtable]{as_huxtable}}}

  \item{knitr}{\code{\link[knitr]{knit_print}}}

  \item{tibble}{\code{\link[tibble]{as_tibble}}}
}}

