% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{as_tibble}
\alias{knit_print}
\alias{vars}
\alias{starts_with}
\alias{ends_with}
\alias{contains}
\alias{matches}
\alias{one_of}
\alias{everything}
\alias{num_range}
\alias{last_col}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{vars}}}

  \item{knitr}{\code{\link[knitr]{knit_print}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{tibble}{\code{\link[tibble]{as_tibble}}}

  \item{tidyselect}{\code{\link[tidyselect]{starts_with}}, \code{\link[tidyselect]{ends_with}}, \code{\link[tidyselect]{contains}}, \code{\link[tidyselect]{matches}}, \code{\link[tidyselect]{one_of}}, \code{\link[tidyselect]{everything}}, \code{\link[tidyselect]{num_range}}, \code{\link[tidyselect]{last_col}}}
}}

