% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_helpers.R
\docType{import}
\name{select_helpers}
\alias{select_helpers}
\alias{all_continuous}
\alias{all_continuous2}
\alias{reexports}
\alias{all_dichotomous}
\alias{all_categorical}
\title{Select helper functions}
\usage{
all_continuous(continuous2 = TRUE)

all_continuous2()
}
\arguments{
\item{continuous2}{Logical indicating whether to include continuous2 variables.
Default is \code{TRUE}}

\item{dichotomous}{Logical indicating whether to include dichotomous variables.
Default is \code{TRUE}}
}
\value{
A character vector of column names selected
}
\description{
Set of functions to supplement the {tidyselect} set of functions for selecting
columns of data frames. \code{all_continuous()}, \code{all_continuous2()}, \code{all_categorical()}, and
\code{all_dichotomous()} may only be used with \code{tbl_summary()}, where each variable
has been classified into one of these three groups. All other helpers
are available throughout the package.
}
\examples{
select_ex1 <-
  trial \%>\%
  select(age, response, grade) \%>\%
  tbl_summary(
    statistic = all_continuous() ~ "{mean} ({sd})",
    type = all_dichotomous() ~ "categorical"
  )
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{broom.helpers}{\code{\link[broom.helpers:select_helpers]{all_categorical}}, \code{\link[broom.helpers:select_helpers]{all_dichotomous}}}
}}

