% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_helpers.R
\name{select_helpers}
\alias{select_helpers}
\alias{all_continuous}
\alias{all_continuous2}
\alias{all_dichotomous}
\alias{all_categorical}
\alias{all_interaction}
\alias{all_intercepts}
\alias{all_contrasts}
\title{Select helper functions}
\usage{
all_continuous(continuous2 = TRUE)

all_continuous2()

all_dichotomous()

all_categorical(dichotomous = TRUE)

all_interaction()

all_intercepts()

all_contrasts(type = c("treatment", "sum", "poly", "helmert"))
}
\arguments{
\item{continuous2}{Logical indicating whether to include continuous2 variables.
Default is \code{TRUE}}

\item{dichotomous}{Logical indicating whether to include dichotomous variables.
Default is \code{TRUE}}
}
\value{
A character vector of column names selected
}
\description{
Set of functions to supplement the {tidyselect} set of
functions for selecting columns of data frames (and other items as well).
\itemize{
\item \code{all_continuous()} selects continuous variables
\item \code{all_continuous2()} selects only type \code{"continuous2"}
\item \code{all_categorical()} selects categorical (including \code{"dichotomous"}) variables
\item \code{all_dichotomous()} selects only type \code{"dichotomous"}
\item \code{all_interaction()} selects interaction terms from a regression model
\item \code{all_intercepts()} selects intercept terms from a regression model
\item \code{all_contrasts()} selects variables in regression model based on their type of contrast
}
}
\examples{
select_ex1 <-
  trial \%>\%
  select(age, response, grade) \%>\%
  tbl_summary(
    statistic = all_continuous() ~ "{mean} ({sd})",
    type = all_dichotomous() ~ "categorical"
  )
}
