% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_helpers.R
\name{select_helpers}
\alias{select_helpers}
\alias{all_continuous}
\alias{all_categorical}
\alias{all_dichotomous}
\alias{all_numeric}
\alias{all_character}
\alias{all_integer}
\alias{all_double}
\alias{all_logical}
\alias{all_factor}
\title{Select helper functions}
\usage{
all_continuous()

all_categorical(dichotomous = TRUE)

all_dichotomous()

all_numeric()

all_character()

all_integer()

all_double()

all_logical()

all_factor()
}
\arguments{
\item{dichotomous}{Logical indicating whether to include dichotomous variables.
Default is \code{TRUE}}
}
\value{
A character vector of column names selected
}
\description{
Set of functions to supplement the {tidyselect} set of functions for selecting
columns of data frames. \code{all_continuous()}, \code{all_categorical()}, and
\code{all_dichotomous()} may only be used with \code{tbl_summary()}, where each variable
has been classified into one of these three groups. All other helpers
are available throughout the package.
}
\examples{
select_ex1 <-
  trial \%>\%
  dplyr::select(age, response, grade) \%>\%
  tbl_summary(
    statistic = all_continuous() ~ "{mean} ({sd})",
    type = all_dichotomous() ~ "categorical"
  )
}
