% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_gtsummary_theme.R
\name{set_gtsummary_theme}
\alias{set_gtsummary_theme}
\alias{gtsummary_theme_reset}
\alias{gtsummary_theme_jama}
\title{Set a gtsummary theme}
\usage{
set_gtsummary_theme(x)

gtsummary_theme_reset()

gtsummary_theme_jama()
}
\arguments{
\item{x}{A named list defining a gtsummary theme. See details below.}
}
\description{
Use this function to set preferences for the display of gtsummary tables.
}
\section{Details}{


The following fields are available to set:
\subsection{Package-wide Settings}{
\itemize{
\item \code{pkg:print_engine} default print engine
\item \code{pkg:pvalue_fun} default p-value formatting function
}
}

\subsection{Function Settings}{
\subsection{\code{tbl_summary()}}{
\subsection{Set Default Function Arguments}{
\itemize{
\item \code{fn:tbl_summary-arg:label}
\item \code{fn:tbl_summary-arg:statistic}
\item \code{fn:tbl_summary-arg:digits}
\item \code{fn:tbl_summary-arg:type}
\item \code{fn:tbl_summary-arg:value}
\item \code{fn:tbl_summary-arg:missing}
\item \code{fn:tbl_summary-arg:missing_text}
\item \code{fn:tbl_summary-arg:percent}
\item \code{fn:tbl_summary-arg:sort}
}
}

\subsection{Other Settings}{
\itemize{
\item \code{fn:tbl_summary-attr:percent_fun} default function for styling percentages
}
}

}

\subsection{\code{add_p.tbl_summary()}}{
\subsection{Set Default Function Arguments}{
\itemize{
\item \code{fn:add_p.tbl_summary-arg:test}
\item \code{fn:add_p.tbl_summary-arg:pvalue_fun}
\item \code{fn:add_p.tbl_summary-arg:group}
}
}

\subsection{Other Settings}{
\itemize{
\item \code{fn:add_p-attr:test.continuous_by2} default test for continuous variables with 2-level by variable
\item \code{fn:add_p-attr:test.continuous}	default test for continuous variables with 3- or more level by variable
\item \code{fn:add_p-attr:test.categorical}	default test for categorical/dichotomous variables
\item \code{fn:add_p-attr:test.categorical.low_count}	default test for categorical/dichotomous variables with minimum expected count <5 in one cell
\item \code{fn:add_p-attr:test.categorical.group_by2}	default test for categorical/dichotomous grouped/correlated variables with 2-level by variable
\item \code{fn:add_p-attr:test.continuous.group_by2}	default test for continuous grouped/correlated variables with 2-level by variable
}
}

}

\subsection{\code{add_q()}}{
\subsection{Set Default Function Arguments}{
\itemize{
\item \code{fn:add_q-arg:method}
\item \code{fn:add_q-arg:pvalue_fun}
}
}

}

}
}

\examples{
# Setting JAMA theme for gtsummary
set_gtsummary_theme(gtsummary_theme_jama())

set_gtsummary_theme_ex1 <-
  trial \%>\%
  dplyr::select(age, grade, trt) \%>\%
  tbl_summary(by = trt)

# reset gtsummary theme
gtsummary_theme_reset()
}
