% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_tbl_hierarchical.R
\name{sort_tbl_hierarchical}
\alias{sort_tbl_hierarchical}
\alias{tbl_sort}
\alias{tbl_sort.tbl_hierarchical}
\title{Sort Hierarchical Tables}
\usage{
tbl_sort(x, ...)

\method{tbl_sort}{tbl_hierarchical}(x, sort = "frequency", desc = FALSE, .stat = "n")
}
\arguments{
\item{x}{(\code{tbl_hierarchical})\cr
A hierarchical gtsummary table of class \code{'tbl_hierarchical'}.}

\item{sort}{(\code{string})\cr
Specifies sorting to perform. Values must be one of \code{c("alphanumeric", "frequency")}. Default is \code{"frequency"}.}

\item{desc}{(scalar \code{logical})\cr
Whether to sort rows in ascending or descending order. Default is descending (\code{desc = TRUE}).}

\item{.stat}{(\code{string})\cr
Statistic to use for sorting when \code{sort = "frequency"}. This statistic must be present in the table for all
hierarchy levels.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}\cr

This function is used to sort hierarchical tables. Options for sorting criteria are:
\enumerate{
\item Alphanumeric - rows are ordered alphanumerically by label text (default).
\item Frequency - within each section of the hierarchy table, frequency sums are calculated for each row and rows are
ordered accordingly.
}
}
\examples{
ADAE_subset <- cards::ADAE |>
  dplyr::filter(
    AESOC \%in\% unique(cards::ADAE$AESOC)[1:5],
    AETERM \%in\% unique(cards::ADAE$AETERM)[1:5]
  )

tbl <- tbl_hierarchical(
  data = ADAE_subset,
  variables = c(SEX, RACE, AETERM),
  by = TRTA,
  denominator = cards::ADSL |> mutate(TRTA = ARM),
  id = USUBJID,
  overall_row = TRUE
)

tbl_sort(tbl)
}
