% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_hierarchical.R
\name{tbl_hierarchical}
\alias{tbl_hierarchical}
\title{Create hierarchical table}
\usage{
tbl_hierarchical(
  data,
  hierarchies,
  by = NULL,
  id = NULL,
  denominator = NULL,
  include = everything(),
  statistic = ifelse(!missing(id), "{n} ({p})", "{n}"),
  digits = NULL,
  overall_row = FALSE
)
}
\description{
Create hierarchical table
}
\examples{
tbl_hierarchical(
  data = cards::ADAE,
  hierarchies = c(AESOC, AETERM),
  by = ARM,
  statistic = "{n} ({p})",
  denominator = cards::ADSL,
  id = USUBJID
)

}
