% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_merge.R
\name{tbl_merge}
\alias{tbl_merge}
\title{Merge two or more gtsummary regression objects}
\usage{
tbl_merge(tbls, header_span = NULL)
}
\arguments{
\item{tbls}{list of gtsummary regression objects}

\item{header_span}{Spanning headers.  Vector with same length as \code{tbls}}
}
\description{
The \code{tbl_merge} function merges two or more \code{tbl_regression} or
\code{tbl_uvregression} objects and adds appropriate spanning headers.
}
