% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_regression.R
\name{tbl_regression}
\alias{tbl_regression}
\title{Display regression model results in table}
\usage{
tbl_regression(x, exponentiate = FALSE, label = NULL, include = NULL,
  exclude = NULL, show_yesno = NULL, conf.level = 0.95,
  intercept = FALSE, coef_fun = ifelse(exponentiate == TRUE,
  style_ratio, style_sigfig), pvalue_fun = style_pvalue)
}
\arguments{
\item{x}{regression model object}

\item{exponentiate}{logical argument passed directly to
\code{tidy} function. Default is \code{FALSE}}

\item{label}{list of labels to write in the output. \code{list(age60 = "Age > 60")}}

\item{include}{names of variables to include in output.}

\item{exclude}{names of variables to exclude from output.}

\item{show_yesno}{By default yes/no categorical variables are printed on a single row,
when the 'No' category is the reference group.  To print both levels in the output table, include
the variable name in the show_yesno vector, e.g. `show_yesno = c("var1", "var2")``}

\item{conf.level}{confidence level passed directly to \code{tidy} function. Default is 0.95.}

\item{intercept}{logical argument indicates whether to include the intercept
in the output.  Default is \code{FALSE}}

\item{coef_fun}{function to round and format beta coefficients.  Default
is \code{\link{style_sigfig}} when the coefficients are printed, and
\code{\link{style_ratio}} when the coefficients have been exponentiated.}

\item{pvalue_fun}{function to round and format p-values.
Default is \code{\link{style_pvalue}}.
The function must have a numeric vector input (the numeric, exact p-value),
and return a string that is the rounded/formatted p-value (e.g.
\code{pvalue_fun = function(x) style_pvalue(x, digits = 2)} or equivalently,
\code{purrr::partial(style_pvalue, digits = 2)}).}
}
\description{
This function uses \code{broom::tidy} from the \code{broom} or \code{broom.mixed} packages
to perform the initial model formatting. Review the \code{tbl_regression}
\href{http://www.danieldsjoberg.com/gtsummary/articles/tbl_regression.html}{vignette}
for detailed examples.
}
\section{Note}{

The N reported in the \code{tbl_uvregression()} output is the number of observations
in the data frame \code{model.frame(x)}. Depending on the model input, this N
may represent different quantities. In most cases, it is the number of people or
units in your model.  Here are some common exceptions.
\enumerate{
\item Survival regression models including time dependent covariates.
\item Random- or mixed-effects regression models with clustered data.
\item GEE regression models with clustered data.
}

This list is not exhaustive, and care should be taken for each number reported.
}

\examples{
mod1 <-
  lm(hp ~ mpg + factor(cyl), mtcars) \%>\%
  tbl_regression()

mod2 <-
  glm(response ~ age + grade + stage, trial, family = binomial(link = "logit")) \%>\%
  tbl_regression(exponentiate = TRUE)

library(lme4)
mod_glmer <-
  glmer(am ~ hp + (1 | gear), mtcars, family = binomial) \%>\%
  tbl_regression(exponentiate = TRUE)
}
\seealso{
Other tbl_regression: \code{\link{add_global.tbl_regression}},
  \code{\link{add_nevent.tbl_regression}},
  \code{\link{bold_italicize_labels_levels}},
  \code{\link{inline_text.tbl_regression}},
  \code{\link{tab_style_bold_p.tbl_regression}}
}
\author{
Daniel D. Sjoberg
}
\concept{tbl_regression}
