% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_regression_methods.R
\name{tbl_regression_methods}
\alias{tbl_regression_methods}
\alias{tbl_regression.lmerMod}
\alias{tbl_regression.glmerMod}
\alias{tbl_regression.survreg}
\alias{tbl_regression.multinom}
\title{Methods for tbl_regression}
\usage{
\method{tbl_regression}{lmerMod}(
  x,
  tidy_fun = function(x, ...) broom.mixed::tidy(x, ..., effects = "fixed"),
  ...
)

\method{tbl_regression}{glmerMod}(
  x,
  tidy_fun = function(x, ...) broom.mixed::tidy(x, ..., effects = "fixed"),
  ...
)

\method{tbl_regression}{survreg}(
  x,
  tidy_fun = function(x, ...) broom::tidy(x, ...) \%>\% dplyr::filter(.data$term !=
    "Log(scale)"),
  ...
)

\method{tbl_regression}{multinom}(x, group_header = NULL, ...)
}
\arguments{
\item{x}{Regression model object}

\item{tidy_fun}{Option to specify a particular tidier function if the
model is not a \link[=vetted_models]{vetted model} or you need to implement a
custom method. Default is \code{NULL}}

\item{...}{arguments passed to \code{tbl_regression.default()}}

\item{group_header}{Character vector with table headers where length matches
the length of \verb{tbls=}}
}
\description{
Most regression models are handled by \link{tbl_regression.default},
which uses \link[broom:reexports]{broom::tidy} to perform initial tidying of results. There are,
however, some model types that have modified default printing behavior.
Those methods are listed below.
}
\section{Methods}{


The default method for \code{tbl_regression()} model summary uses \code{broom::tidy(x)}
to perform the initial tidying of the model object. There are, however,
a few \link[=vetted_models]{vetted model} that use \link[=tbl_regression_methods]{modifications}.
\itemize{
\item \code{"lmerMod"} or \code{"glmerMod"}: These mixed effects models use \code{broom.mixed::tidy(x, effects = "fixed")}
\item \code{"survreg"}: The scale parameter is removed, \code{broom::tidy(x) \%>\% dplyr::filter(term != "Log(scale)")}
\item \code{"multinom"}: This multinomial outcome is complex, and the returned object is a \code{tbl_stack()} object with the parameters for each outcome stacked into a final object
}
}

