% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_split.R
\name{tbl_split}
\alias{tbl_split}
\title{Split gtsummary Table}
\usage{
tbl_split(x, variables)
}
\arguments{
\item{x}{gtsummary table}

\item{variables}{variables at which to split the gtsummary table rows (tables
will be separated after each of these variables)}

\item{...}{not used}
}
\value{
tbl_split object
}
\description{
The \code{tbl_split} function splits a single gtsummary table into multiple tables
}
\examples{
library(gtsummary)

tbl <-
  tbl_summary(trial) \%>\%
  tbl_split(variables = c(marker, grade))

}
\concept{gtsummary-related functions}
