% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_split.R
\name{tbl_split_by}
\alias{tbl_split_by}
\alias{tbl_split_by_rows}
\alias{tbl_split_by_columns}
\alias{first_column}
\alias{print.tbl_split}
\title{Split gtsummary table by rows and/or columns}
\usage{
tbl_split_by_rows(
  x,
  variables,
  footnotes = c("all", "first", "last"),
  caption = c("all", "first")
)

tbl_split_by_columns(
  x,
  keys,
  groups,
  footnotes = c("all", "first", "last"),
  caption = c("all", "first")
)

first_column(x)

\method{print}{tbl_split}(x, ...)
}
\arguments{
\item{x}{(\code{gtsummary} or \code{list})\cr
gtsummary table.}

\item{variables}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
variables at which to split the gtsummary table rows (tables
will be separated after each of these variables).}

\item{footnotes}{(\code{string}) \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}\cr
can be either \code{"first"}, \code{"all"}, or \code{"last"}, to locate global footnotes
only on the first, in each, or in the last table, respectively. It defaults
to \code{"all"}. Reference footnotes are always present wherever they appear.}

\item{caption}{(\code{string}) \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}\cr
similarly to footnotes, makes the titles, and subtitles appear only on the
\code{"first"} or \code{"all"} table. It defaults to \code{"all"}.}

\item{keys}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
columns to be repeated in each table split. It defaults to \code{first_column(x)}
if missing (usually label column).}

\item{groups}{(list of \code{character} vectors)\cr
list of column names that appear in \code{x$table_body}.
Each group of column names represent a different table in the output list.}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
\code{tbl_split} object. If multiple splits are performed (e.g., both by
row and columns), the output is returned a single level list.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}\cr
The \code{tbl_split_by_rows()} and \code{tbl_split_by_columns()} functions split a single
gtsummary table into multiple tables.
Both column-wise splitting (that is, splits by columns in
\code{x$table_body}) and row-wise splitting is possible.
}
\details{
Run \code{\link[=show_header_names]{show_header_names()}} to print all column names to split by.

Footnotes and caption handling are experimental and may change in the future.
}
\examples{
\dontshow{if (identical(Sys.getenv("NOT_CRAN"), "true") || identical(Sys.getenv("IN_PKGDOWN"), "true")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Example 1 ----------------------------------
# Split by rows
trial |>
  tbl_summary(by = trt) |>
  tbl_split_by_rows(variables = c(marker, grade)) |>
  tail(n = 1) # Print only last table for simplicity

# Example 2 ----------------------------------
# Split by columns
trial |>
  tbl_summary(by = trt, include = c(death, ttdeath)) |>
  tbl_split_by_columns(groups = list("stat_1", "stat_2")) |>
  tail(n = 1) # Print only last table for simplicity

# Example 3 ----------------------------------
# Both row and column splitting
trial |>
  tbl_summary(by = trt) |>
  tbl_split_by_rows(variables = c(marker, grade)) |>
  tbl_split_by_columns(groups = list("stat_1", "stat_2")) |>
  tail(n = 1) # Print only last table for simplicity

# Example 4 ------------------------------
trial |>
  tbl_summary(by = trt, missing = "no") |>
  modify_footnote_header(
    footnote = "All but four subjects received both treatments in a crossover design",
    columns = all_stat_cols(),
    replace = FALSE
  ) |>
  modify_footnote_body(
    footnote = "Tumor grade was assessed _before_ treatment began",
    columns = "label",
    rows = variable == "grade" & row_type == "label"
  ) |>
  modify_spanning_header(
    c(stat_1, stat_2) ~ "**TRT**"
  ) |>
  modify_abbreviation("I = 1, II = 2, III = 3") |>
  modify_caption("_Some caption_") |>
  modify_footnote_spanning_header(
    footnote = "Treatment",
    columns = c(stat_1)
  ) |>
  modify_source_note("Some source note!") |>
  tbl_split_by_rows(variables = c(marker, stage, grade), footnotes = "last", caption = "first") |>
  tail(n = 2) |>
  head(n = 1) # Print only one but not last table for simplicity
\dontshow{\}) # examplesIf}
}
