% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_split.R
\name{tbl_split_by}
\alias{tbl_split_by}
\alias{tbl_split_by_rows}
\alias{tbl_split_by_columns}
\alias{first_column}
\alias{print.tbl_split}
\title{Split gtsummary table bz rows and/or columns}
\usage{
tbl_split_by_rows(x, variables, footnotes = "each", caption = "each")

tbl_split_by_columns(x, keys, groups, footnotes = "each", caption = "each")

first_column(x)

\method{print}{tbl_split}(x, ...)
}
\arguments{
\item{x}{(\code{gtsummary} or \code{list})\cr
gtsummary table.}

\item{variables}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
variables at which to split the gtsummary table rows (tables
will be separated after each of these variables).}

\item{footnotes}{(\code{string})\cr
can be either \code{"first"}, \code{"each"}, or \code{"last"}, to locate global footnotes
only on the first, in each, or in the last table, respectively. It defaults
to \code{"each"}. Reference footnotes are always present wherever they appear.}

\item{caption}{(\code{string})\cr
similarly to footnotes, makes the titles, and subtitles appear only on the
\code{"first"} or \code{"each"} table. It defaults to \code{"each"}.}

\item{keys}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
columns to be repeated in each table split. It defaults to \code{first_column(x)}
if missing (usually label column).}

\item{groups}{(list of \code{character} vectors)\cr
list of column names that appear in \code{x$table_body}.
Each group of column names represent a different table in the output list.}
}
\value{
\code{tbl_split} object. They can also be identified further by subclass
definitions \code{tbl_split_by_rows} and \code{tbl_split_by_columns}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}\cr
The \code{tbl_split_by_rows} and \verb{tbl_split_by_columns`` functions split a single gtsummary table into multiple tables. Both horizontal (column-wise, (that is, splits by columns in }x$table_body`) and vertical (row-wise) splits are possible.
Updates to the print method are expected.
}
\details{
Run \code{gtsummary::show_header_names()} to print all column names to split by.
}
\examples{
# create standard table
tbl <- tbl_summary(trial, by = trt)

# split by rows
tbl_list_rows <- tbl |>
  tbl_split_by_rows(variables = c(marker, grade))

# print column names
gtsummary::show_header_names(tbl)

# split by columns
tbl_list_cols <- tbl |>
  tbl_split_by_columns(groups = list("stat_1", "stat_2"))

# take the list of splitted rows and split again by columns
tbl_list_rows_cols <- tbl_list_rows |>
  tbl_split_by_columns(groups = list("stat_1", "stat_2"))

# Handle footnotes

}
