% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_stack.R
\name{tbl_stack}
\alias{tbl_stack}
\title{Stacks two or more gtsummary regression objects}
\usage{
tbl_stack(tbls)
}
\arguments{
\item{tbls}{list of gtsummary regression objects}
}
\description{
Stacks two or more \code{tbl_regression} or \code{tbl_uvregression} objects.
{gt} attributes from the first regression object are utilized for output table.
If combining \code{tbl_stack()} and \code{tbl_merge}, merge first then stack.
}
\examples{
t1 <-
  glm(response ~ age, trial, family = binomial) \%>\%
  tbl_regression(exponentiate = TRUE,
                 label = list(age = "Age, Unadjusted"))

t2 <-
  glm(response ~ age + grade + stage + marker, trial, family = binomial) \%>\%
  tbl_regression(include = "age",
                 exponentiate = TRUE,
                 label = list(age = "Age, Adjusted"))

tbl_stack_ex1 <- tbl_stack(list(t1, t2))
}
