% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_strata.R
\name{tbl_strata}
\alias{tbl_strata}
\title{Stratified gtsummary table}
\usage{
tbl_strata(
  data,
  strata,
  .f,
  ...,
  .sep = ", ",
  combine_with = c("tbl_merge", "tbl_stack")
)
}
\arguments{
\item{data}{a date frame}

\item{strata}{character vector or tidyselector of columns in data to
stratify tables by}

\item{...}{Additional arguments passed on to the \code{.tbl_fun} function.}

\item{.sep}{when more than one stratifying variable is passed, this string is
used to separate the levels. Default is \code{", "}}

\item{combine_with}{One of \code{c("tbl_merge", "tbl_stack")}. Names the function
used to combine the stratified tables.}

\item{.tbl_fun}{A function or formula. If a \emph{function}, it is used as is.
If a formula, e.g. \code{~ .x \%>\% tbl_summary() \%>\% add_p()}, it is converted to a function.
The stratified data frame is passed to this function.}
}
\description{
Build a stratified gtsummary table. Any gtsummary table that accepts
a data frame as its first argument is accepted
}
\examples{
tbl_strata_ex1 <-
  trial \%>\%
  select(age, grade, stage, trt) \%>\%
  mutate(grade = paste("Grade", grade)) \%>\%
  tbl_strata(
    strata = grade,
    ~ .x \%>\%
      tbl_summary(by = trt) \%>\%
      add_p()
  )
}
\author{
Daniel D. Sjoberg
}
