% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_summary.R
\name{tbl_summary}
\alias{tbl_summary}
\title{Calculates and formats descriptive statistics for a data frame}
\usage{
tbl_summary(data, by = NULL, label = NULL, type = NULL,
  statistic = NULL, digits = NULL, group = NULL,
  missing = c("ifany", "always", "no"))
}
\arguments{
\item{data}{a data frame}

\item{by}{a character name of a categorical variable in data, \code{by = "group"}.
Summary statistics will be calculated separately for each level of the by variable.
If \code{NULL}, summary statistics
are calculated using all observations.}

\item{label}{A list of variable labels,
e.g. \code{list(age = "Age, yrs", ptstage = "Path T Stage")}.  If \code{NULL}, the function
will take the label attribute (\code{attr(data$age, "label")}).  If
the label doesn't exist, then the label is assigned as the variable name.}

\item{type}{A list that includes specified summary types.  Accepted values
are \code{c("continuous", "categorical", "dichotomous")},
e.g. \code{type = list(age = "continuous", female = "dichotomous")}.
If type not specified for a variable, the function
will default to an appropriate summary type.}

\item{statistic}{A list of the type of statistics to return.  The list can contain
two names lists (\code{continuous} and \code{categorical}).  The value within the list is the types of
summary statistics to be returned.  For continuous variables the choices are:
\code{median}, \code{p25} (first quartile), \code{p75} (third quartile), \code{mean}, \code{sd} (standard deviation),
\code{min} (minimum), \code{max} (maximum).  For categorical variables the choices are \code{n} (frequency),
\code{N} (denominator, or cohort size), \code{p} (percent).  The defaults are
\code{continuous = "{median} ({p25}, {p75})"} and \code{categorical = "{n} ({p}\\\%)"}.
The syntax follows from the \code{\link[glue]{glue}} function.  Dichotomous variables
follow the same format as categorical.}

\item{digits}{integer indicating the number of decimal places to round continuous
summary statistics. \code{sprintf(glue::glue("\%.{digits}f"), x)}}

\item{group}{Character vector of an ID or grouping variable.  Summary statistics
will not be printed for this column. The column may be used in \code{\link{add_comparison}} to
calculate p-values with correlated data. Default is \code{NULL}}

\item{missing}{whether to include \code{NA} values in the table. \code{missing} controls
if the table includes counts of \code{NA} values: the allowed values correspond to
never (\code{"no"}), only if the count is positive (\code{"ifany"}) and even for
zero counts (\code{"always"}). Default is \code{"ifany"}.}
}
\value{
List of summary statistics formattable to a \code{gt} object
}
\description{
The \code{tbl_summary} function calculates descriptive statistics by groups for
continuous, categorical, and dichotomous variables.  Review the
\href{https://github.com/pages/ddsjoberg/gtsummary/articles/tbl_summary.html}{\code{tbl_summary} vignette}
for detailed examples.
}
\author{
Daniel Sjoberg
}
