% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_summary.R
\name{tbl_summary}
\alias{tbl_summary}
\title{Calculates and formats descriptive statistics for a data frame}
\usage{
tbl_summary(data, by = NULL, label = NULL, type = NULL,
  value = NULL, statistic = NULL, digits = NULL, group = NULL,
  missing = c("ifany", "always", "no"))
}
\arguments{
\item{data}{a data frame}

\item{by}{a character name of a categorical variable in data, \code{by = "group"}.
Summary statistics will be calculated separately for each level of the by variable.
If \code{NULL}, summary statistics
are calculated using all observations.}

\item{label}{A list of variable labels,
e.g. \code{list(age = "Age, yrs", ptstage = "Path T Stage")}.  If \code{NULL}, the function
will take the label attribute (\code{attr(data$age, "label")}).  If
the label doesn't exist, the variable name will be used.}

\item{type}{A list that includes specified summary types.  Accepted values
are \code{c("continuous", "categorical", "dichotomous")},
e.g. \code{type = list(age = "continuous", female = "dichotomous")}.
If type not specified for a variable, the function
will default to an appropriate summary type.  See below for details.}

\item{value}{A list that specifies the value to display for dichotomous
values only.  See below for details.}

\item{statistic}{A list of the type of statistics to return.  The list can contain
two named elements (\code{continuous} and \code{categorical}). The default is
\code{list(continuous = "{median} ({p25}, {p75})", categorical = "{n} ({p}\%)")}.
The syntax follows from the \code{\link[glue]{glue}} function.
For categorical variables the choices the following statistics are available to
report: \code{{n}} (frequency), \code{{N}} (denominator, or cohort size), \code{{p}} (percent).
For continuous variables, any quantile may be returned with \code{{p##}}, where ##
is any integer from 0 to 100. For example, \code{{p25}} and \code{{p75}} returns
the 25th and 75th quantiles.  The median (\code{{median}}), mean (\code{{mean}}),
standard deviation (\code{{sd}}), variance (\code{{var}}), minimum (\code{{min}}), and
maximum (\code{{max}}) are available.  In fact, any function that takes the form
\code{foo(x, na.rm = TRUE)} should work.}

\item{digits}{integer indicating the number of decimal places to round continuous
summary statistics. \code{sprintf(glue::glue("\%.{digits}f"), x)}}

\item{group}{Character vector of an ID or grouping variable.  Summary statistics
will not be printed for this column. The column may be used in \code{\link{add_comparison}} to
calculate p-values with correlated data. Default is \code{NULL}}

\item{missing}{whether to include \code{NA} values in the table. \code{missing} controls
if the table includes counts of \code{NA} values: the allowed values correspond to
never (\code{"no"}), only if the count is positive (\code{"ifany"}) and even for
zero counts (\code{"always"}). Default is \code{"ifany"}.}
}
\value{
List of summary statistics to be converted to a \code{gt} object
}
\description{
The \code{tbl_summary} function calculates descriptive statistics by groups for
continuous, categorical, and dichotomous variables.  Review the
\href{http://www.danieldsjoberg.com/gtsummary/articles/tbl_summary.html}{\code{tbl_summary} vignette}
for detailed examples.
}
\section{Specifying Variable Types}{

tbl_summary displays summary statistics for three types of data:
continuous, categorical, and dichotomous. If the type is not specified,
tbl_summary will do its best to guess the type.  Dichotomous variables
are categorical variables that are displayed on a single row in the
output table, rather than one row per level of the variable.  For dichotomous
variables coded as TRUE/FALSE, 0/1, or yes/no, the TRUE, 1, and yes rows
will be displayed.  Otherwise, the value to display must be specified in
the \code{value} argument, e.g. \code{value = list(varname = "level to show")}
}

\examples{
tbl_overall <- tbl_summary(trial)
tbl_trt <- tbl_summary(trial, by = "trt")
tbl_lbls <- mtcars \%>\% tbl_summary(label = list(cyl = "No. Cylinders"))
}
\author{
Daniel Sjoberg
}
