% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_summary.R
\name{tbl_summary}
\alias{tbl_summary}
\title{Title}
\usage{
tbl_summary(
  data,
  by = NULL,
  label = NULL,
  statistic = NULL,
  digits = NULL,
  type = NULL,
  value = NULL,
  missing = c("ifany", "no", "always"),
  missing_text = "Unknown",
  missing_stat = "{N_miss}",
  sort = NULL,
  percent = c("column", "row", "cell"),
  include = c(everything(), -by)
)
}
\arguments{
\item{data}{TODO:}

\item{by}{TODO:}

\item{label}{TODO:}

\item{statistic}{TODO:}

\item{digits}{TODO:}

\item{type}{TODO:}

\item{value}{TODO:}

\item{missing}{TODO:}

\item{missing_text}{TODO:}

\item{missing_stat}{TODO:}

\item{sort}{TODO:}

\item{percent}{TODO:}

\item{include}{TODO:}
}
\value{
a gtsummary table of class \code{"tbl_summary"}
}
\description{
Title
}
\examples{
tbl <-
  tbl_summary(
    data = mtcars,
    type =
      list(
        cyl = "categorical",
        am = "dichotomous",
        mpg = "continuous",
        hp = "continuous2"
      ),
    value = list(am = 1),
    statistic =
      list(
        c(cyl, am) ~ "{n} ({p}\%)",
        mpg = "{mean} ({sd})",
        hp = c("{mean}", "{median}")
      )
  )
}
