% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_survfit.R
\name{tbl_survfit}
\alias{tbl_survfit}
\alias{tbl_survfit.list}
\title{Survival table}
\usage{
tbl_survfit(x, ...)

\method{tbl_survfit}{list}(
  x,
  times = NULL,
  probs = NULL,
  statistic = "{estimate} ({conf.low}, {conf.high})",
  label = NULL,
  label_header = ifelse(!is.null(times), "Time {time}", "{prob} Percentile"),
  estimate_fun = ifelse(!is.null(times), styfn_percent(), styfn_sigfig()),
  missing = "--",
  conf.level = 0.95,
  type = NULL,
  reverse = FALSE,
  quiet = TRUE,
  ...
)
}
\arguments{
\item{x}{(\code{survfit}, \code{list}, \code{data.frame})\cr
a survfit object, list of survfit objects, or a data frame.
If a data frame is passed, a list of survfit objects is constructed using
each variable as a stratifying variable.}

\item{...}{For \code{\link[=tbl_survfit.data.frame]{tbl_survfit.data.frame()}}  and \code{\link[=tbl_survfit.survfit]{tbl_survfit.survfit()}} the arguments
are passed to \code{\link[=tbl_survfit.list]{tbl_survfit.list()}}. They are not used when \code{\link[=tbl_survfit.list]{tbl_survfit.list()}}
is called directly.}

\item{times}{(\code{numeric})\cr
a vector of times for which to return survival probabilities.}

\item{probs}{(\code{numeric})\cr
a vector of probabilities with values in (0,1) specifying the survival quantiles to return.}

\item{statistic}{(\code{string})\cr
string defining the statistics to present in the table.
Default is \code{"{estimate} ({conf.low}, {conf.high})"}}

\item{label}{(\code{\link[=syntax]{formula-list-selector}})\cr
List of formulas specifying variables labels,
e.g. \code{list(age = "Age, yrs", stage = "Path T Stage")}, or a string for a
single variable table.}

\item{label_header}{(\code{string})\cr
string specifying column labels above statistics. Default
is \code{"{prob} Percentile"} for survival percentiles, and \code{"Time {time}"} for n-year
survival estimates}

\item{estimate_fun}{(\code{function})\cr
function to format the Kaplan-Meier estimates. Default
is \code{\link[=styfn_percent]{styfn_percent()}} for survival probabilities and \code{\link[=styfn_sigfig]{styfn_sigfig()}} for
survival times}

\item{missing}{(\code{string})\cr
text to fill when estimate is not estimable. Default is \code{"--"}}

\item{conf.level}{(scalar \code{numeric})\cr ]
Confidence level for confidence intervals. Default is \code{0.95}}

\item{type}{(\code{string} or \code{NULL})\cr
type of statistic to report. Available for Kaplan-Meier time estimates only, otherwise \code{type}
is ignored. Default is \code{NULL}.
Must be one of the following:\tabular{ll}{
   type \tab transformation \cr
   \code{"survival"} \tab \code{x} \cr
   \code{"risk"} \tab \code{1 - x} \cr
   \code{"cumhaz"} \tab \code{-log(x)} \cr
}}

\item{reverse}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}

\item{quiet}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}

\item{y}{outcome call, e.g. \code{y = Surv(ttdeath, death)}}

\item{include}{Variable to include as stratifying variables.}
}
\description{
Function takes a \code{survfit} object as an argument, and provides a
formatted summary table of the results
}
\examples{
\dontshow{if (gtsummary:::is_pkg_installed("survival", reference_pkg = "gtsummary")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(survival)
\dontshow{\}) # examplesIf}
}
\author{
Daniel D. Sjoberg
}
