% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_survival.R
\name{tbl_survival}
\alias{tbl_survival}
\title{Creates table of univariate summary statistics for time-to-event endpoints}
\usage{
tbl_survival(x, ...)
}
\arguments{
\item{x}{A survfit object}

\item{...}{Additional arguments passed to other methods}
}
\description{
\lifecycle{experimental}
}
\seealso{
\link{tbl_survival.survfit}
}
