% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_survival.R
\name{tbl_survival}
\alias{tbl_survival}
\title{Tabulate survival probabilites for time-to-event endpoints}
\usage{
tbl_survival(x, ...)
}
\arguments{
\item{x}{survival object}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Tabulate survival probabilites for time-to-event endpoints
}
\seealso{
\code{\link{tbl_survival.survfit}},
\code{\link{inline_text.tbl_regression}},
\link{inline_text.tbl_uvregression}
}
\author{
Daniel D. Sjoberg
}
