% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_survival.R
\name{tbl_survival}
\alias{tbl_survival}
\title{Tabulate Kaplan-Meier survival probabilities for time-to-event endpoints}
\usage{
tbl_survival(x, times, time_label = "{time}",
  level_label = "{level}, N = {n}", header_time = md("**Time**"),
  header_prob = md("**Probability**"))
}
\arguments{
\item{x}{a survfit object with a single stratifying variable}

\item{times}{numeric vector of survival times}

\item{time_label}{string defining the label shown for the time column.
Default is \code{"{time}"}.  The input uses \code{\link[glue]{glue}} notation to
convert the string into a label.  A common label may be \code{"{time} Months"}, which
would resolve to "6 Months" or "12 Months" depending on the times requested.}

\item{level_label}{used when survival results are stratified.
It is a string defining the label shown.  The input uses
\code{\link[glue]{glue}} notation to convert the string into a label.
The default is \code{"{level}, N = {n}"}.  Other information available to
call are \code{'{n}'}, \code{'{level}'}, \code{'{variable}'}, \code{'{n.event.tot}'}, and \code{'{strata}'}.}

\item{header_time}{string to be displayed as column header of the time column.
Default is \code{md('**Time**')}}

\item{header_prob}{string to be displayed as column header of the Kaplan-Meier
estimate.  Default is \code{md('**Probability**')}}

\item{...}{Arguments passed on to \code{survival::survfit.formula}
\describe{
  \item{weights}{
    The weights must be nonnegative and it is strongly recommended that  
    they be strictly positive, since zero weights are ambiguous, compared 
    to use of the \code{subset} argument.
  }
  \item{subset}{
    expression saying that only a subset of the rows of the data 
    should be used in the fit. 
  }
  \item{na.action}{
    a missing-data filter function, applied to the model frame, after any 
    \code{subset} argument has been used. 
    Default is \code{options()$na.action}. 
  }
  \item{etype}{
    a variable giving the type of event.  This has been superseded by
    multi-state Surv objects; see example below.
  }
  \item{id}{
    identifies individual subjects, when a given person can have multiple
    lines of data.
  }
  \item{istate}{for multi-state models, identifies the initial state of
    each subject}
  \item{timefix}{process times through the \code{aeqSurv} function to
  eliminate potential roundoff issues.}
}}
}
\description{
Function takes a \code{survfit} object as an argument, and provides a tabulated and
formatted summary of the results
}
\examples{
library(survival)
fit1 <- survfit(Surv(ttdeath, death) ~ trt, trial)
tbl_strata <-
  tbl_survival(
    fit1,
    times = c(12, 24),
    time_label = "{time} Months"
  )

fit2 <- survfit(Surv(ttdeath, death) ~ 1, trial)
tbl_nostrata <-
  tbl_survival(
    fit2,
    times = c(12, 24),
    time_label = "{time} Months"
  )
}
\seealso{
\link[gt]{md}, \link[gt]{html}
}
\author{
Daniel D. Sjoberg
}
