% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_survival.R
\name{tbl_survival}
\alias{tbl_survival}
\title{Creates table of univariate summary statistics for time-to-event endpoints}
\usage{
tbl_survival(...)
}
\arguments{
\item{...}{Not used}
}
\description{
\lifecycle{deprecated}
Please use \link{tbl_survfit}.
}
\keyword{internal}
