% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_survival.R
\name{tbl_survival}
\alias{tbl_survival}
\title{Creates table of univariate summary statistics for time-to-event endpoints}
\usage{
tbl_survival(x, ...)
}
\arguments{
\item{x}{A survfit object}

\item{...}{Additional arguments passed to other methods}
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("questioning")}
Please use \link{tbl_survfit}.
Function takes a \code{survfit} object as an argument, and provides a
formatted summary of the results
}
\seealso{
\link{tbl_survival.survfit}
}
\keyword{internal}
