% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_svysummary.R
\name{tbl_svysummary}
\alias{tbl_svysummary}
\title{Create a table of summary statistics from a survey object}
\usage{
tbl_svysummary(
  data,
  by = NULL,
  label = NULL,
  statistic = NULL,
  digits = NULL,
  type = NULL,
  value = NULL,
  missing = NULL,
  missing_text = NULL,
  sort = NULL,
  percent = NULL,
  include = everything()
)
}
\arguments{
\item{data}{A survey object created with created with \code{\link[survey:svydesign]{survey::svydesign()}}}

\item{by}{A column name (quoted or unquoted) in \code{data}.
Summary statistics will be calculated separately for each level of the \code{by}
variable (e.g. \code{by = trt}). If \code{NULL}, summary statistics
are calculated using all observations.}

\item{label}{List of formulas specifying variables labels,
e.g. \code{list(age ~ "Age, yrs", stage ~ "Path T Stage")}.  If a
variable's label is not specified here, the label attribute
(\code{attr(data$age, "label")}) is used.  If
attribute label is \code{NULL}, the variable name will be used.}

\item{statistic}{List of formulas specifying types of summary statistics to
display for each variable.  The default is
\code{list(all_continuous() ~ "{median} ({p25}, {p75})", all_categorical() ~ "{n} ({p}\%)")}.
See below for details.}

\item{digits}{List of formulas specifying the number of decimal
places to round continuous summary statistics. If not specified,
\code{tbl_summary} guesses an appropriate number of decimals to round statistics.
When multiple statistics are displayed for a single variable, supply a vector
rather than an integer.  For example, if the
statistic being calculated is \code{"{mean} ({sd})"} and you want the mean rounded
to 1 decimal place, and the SD to 2 use \code{digits = list(age ~ c(1, 2))}.}

\item{type}{List of formulas specifying variable types. Accepted values
are \code{c("continuous", "categorical", "dichotomous")},
e.g. \code{type = list(age ~ "continuous", female ~ "dichotomous")}.
If type not specified for a variable, the function
will default to an appropriate summary type. See below for details.}

\item{value}{List of formulas specifying the value to display for dichotomous
variables.  See below for details.}

\item{missing}{Indicates whether to include counts of \code{NA} values in the table.
Allowed values are \code{"no"} (never display NA values),
\code{"ifany"} (only display if any NA values), and \code{"always"}
(includes NA count row for all variables). Default is \code{"ifany"}.}

\item{missing_text}{String to display for count of missing observations.
Default is \code{"Unknown"}.}

\item{sort}{List of formulas specifying the type of sorting to perform for
categorical data. Options are \code{frequency} where results are sorted in
descending order of frequency and \code{alphanumeric},
e.g. \code{sort = list(everything() ~ "frequency")}}

\item{percent}{Indicates the type of percentage to return. Must be one of
\code{"column"}, \code{"row"}, or \code{"cell"}. Default is \code{"column"}.}

\item{include}{variables to include in the summary table. Default is \code{everything()}}
}
\value{
A \code{tbl_svysummary} object
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}
}
\details{
The \code{tbl_svysummary} function calculates descriptive statistics for
continuous, categorical, and dichotomous variables taking into account survey weights and design.
It is similar to \code{\link[=tbl_summary]{tbl_summary()}}.

See \code{\link[=tbl_summary]{tbl_summary()}} for more details on the different arguments.
}
\section{statistic argument}{

The statistic argument specifies the statistics presented in the table. The
input is a list of formulas that specify the statistics to report. For example,
\code{statistic = list(age ~ "{mean} ({sd})")} would report the mean and
standard deviation for age; \code{statistic = list(all_continuous() ~ "{mean} ({sd})")}
would report the mean and standard deviation for all continuous variables.
A statistic name that appears between curly brackets
will be replaced with the numeric statistic (see \link[glue:glue]{glue::glue}).

For categorical variables the following statistics are available to display.
\itemize{
\item \code{{n}} frequency
\item \code{{N}} denominator, or cohort size
\item \code{{p}} formatted percentage
\item \code{{n_unweighted}} unweighted frequency
\item \code{{N_unweighted}} unweighted denominator
\item \code{{p_unweighted}} unweighted formatted percentage
}
For continuous variables the following statistics are available to display.
\itemize{
\item \code{{median}} median
\item \code{{mean}} mean
\item \code{{sd}} standard deviation
\item \code{{var}} variance
\item \code{{min}} minimum
\item \code{{max}} maximum
\item \verb{\{p##\}} any integer percentile, where \verb{##} is an integer from 0 to 100
}

Unlike \code{\link[=tbl_summary]{tbl_summary()}}, It's not possible to pass a custom function.

For both categorical and continuous variables, statistics on the number of
missing and non-missing observations and their proportions are available to
display.
\itemize{
\item \code{{N_obs}} total number of observations
\item \code{{N_miss}} number of missing observations
\item \code{{N_nonmiss}} number of non-missing observations
\item \code{{p_miss}} percentage of observations missing
\item \code{{p_nonmiss}} percentage of observations not missing
\item \code{{N_obs_unweighted}} unweighted total number of observations
\item \code{{N_miss_unweighted}} unweighted number of missing observations
\item \code{{N_nonmiss_unweighted}} unweighted number of non-missing observations
\item \code{{p_miss_unweighted}} unweighted percentage of observations missing
\item \code{{p_nonmiss_unweighted}} unweighted percentage of observations not missing
}

Note that for categorical variables, \code{{N_obs}}, \code{{N_miss}} and \code{{N_nonmiss}} refer
to the total number, number missing and number non missing observations
in the denominator, not at each level of the categorical variable.
}

\examples{
# A simple weighted dataset
d <- survey::svydesign(~1, data = as.data.frame(Titanic), weights = ~Freq)
tbl_svysummary(d, by = "Survived", include = -Freq, percent = "row")

# A dataset with a complex design
data(api, package = "survey")
d_clust <- survey::svydesign(id = ~dnum, weights = ~pw, data = apiclus1, fpc = ~fpc)
tbl_svysummary(d_clust, include = c("cname", "api00", "api99", "both"))
tbl_svysummary(d_clust, include = c("cname", "api00", "api99", "both"), by = "both")
}
\seealso{
Other tbl_svysummary tools: 
\code{\link{add_overall}()}
}
\author{
Joseph Larmarange
}
\concept{tbl_svysummary tools}
