% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble_as_cli.R
\name{tibble_as_cli}
\alias{tibble_as_cli}
\title{Print tibble with cli}
\usage{
tibble_as_cli(x, na_value = "", label = list(), padding = 3L)
}
\arguments{
\item{x}{(\code{data.frame})\cr
a data frame with all character columns.}

\item{na_value}{(\code{string})\cr
a string indicating how an \code{NA} value will appear in printed table.}

\item{label}{(named \code{list})\cr
named list of column labels to use. Default is to print the column names.}

\item{padding}{(\code{integer})\cr
an integer indicating the amount of padding between columns.}
}
\description{
Print a tibble or data frame using cli styling and formatting.
}
\examples{
trial[1:3, ] |>
  dplyr::mutate_all(as.character) |>
  gtsummary:::tibble_as_cli()
}
\keyword{internal}
