% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_examples.R
\name{utils_examples}
\alias{utils_examples}
\alias{write_example_output}
\alias{read_example_output}
\alias{write_read_example_output}
\title{Utilities to write and read tables in html for function documentation}
\usage{
write_example_output(
  tbl,
  example_name = "example",
  out_var_name = NULL,
  decoration_function = NULL
)

read_example_output(out_var_name, example_name = "example")

write_read_example_output(
  tbl,
  example_name = "example",
  out_var_name = deparse(substitute(tbl))
)
}
\arguments{
\item{tbl}{(\code{gtsummary} or \code{html})\cr
Tables are converted to html. Any html code (table or not) can be saved and
retrieved for documentation purposes.}

\item{example_name}{(\code{string})\cr
Usually the name of the function or file where the table is used.}

\item{out_var_name}{(\code{string})\cr
The name of the variable that will be saved. It defaults to the name of the
variable passed to the function (\code{tbl}).}

\item{decoration_function}{(\code{NULL} or \code{function})\cr
Function of the form \code{function(x) tab_options(x) |> tab_style()} where additional
specifications can be added to the table before saving it. If \code{NULL} only a white
\code{gt::cell_fill} is used.}
}
\description{
This is still experimental. We are planning to potentially expand this to other
html output. Principally based on \code{\link[=as_gt]{as_gt()}} and \code{\link[gt:as_raw_html]{gt::as_raw_html()}}.
}
\keyword{internal}
