% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_examples.R
\name{utils_examples}
\alias{utils_examples}
\alias{write_example_output}
\alias{read_example_output}
\alias{write_read_example_output}
\title{Utilities to write and read tables in html for function documentation}
\usage{
write_example_output(
  tbl,
  example_name = "example",
  out_var_name = deparse(substitute(tbl))
)

read_example_output(out_var_name, example_name = "example")

write_read_example_output(
  tbl,
  example_name = "example",
  out_var_name = deparse(substitute(tbl))
)
}
\arguments{
\item{tbl}{(\code{gtsummary} or \code{html})\cr tables are converted to html. Any html code (table or not)
can be saved and retrieved for documentation purposes.}

\item{example_name}{(\code{string})\cr usually the name of the function or file where the
table is used.}

\item{out_var_name}{(\code{string})\cr the name of the variable that will be saved. It defaults
to the name of the variable passed to the function (\code{tbl}).}

\item{method}{(\code{string})\cr saving method. Default is \code{"rds"} as it produces a smaller file,
but we are planning to extend this to other, non-html outputs like \code{"png"}.}
}
\description{
This is still experimental. We are planning to potentially expand this to other
html output. Principally based on \code{\link[=as_gt]{as_gt()}} and \code{\link[gt:as_raw_html]{gt::as_raw_html()}}.
}
\keyword{internal}
